### 2019-04-14 #1382 @whedon
[PRE REVIEW]: PyVista: 3D plotting and mesh analysis through a streamlined interface for the Visualization Toolkit (VTK)
**Submitting author:** @banesullivan (<a href="http://orcid.org/0000-0001-8628-4566">C. Bane Sullivan</a>)
**Repository:** <a href="https://github.com/pyvista/pyvista/" target ="_blank">https://github.com/pyvista/pyvista/</a>
**Version:** 0.18.1
