### 2019-04-14 #1382 @banesullivan
Thanks, @labarba! It looks like whedon successfully rendered the PDF!

From the preview, it seems I may need to change a few things in the Markdown:

- One code snippet runs off the page - I'll try to fix this.
- I'm wondering if I added the figure captions in the correct way in Markdown?
- I may need to shift the figures around. Is there a way to specify the width in markdown that the latex compiler will recognize?

If I edit on the master branch, will whedon update after using a command?
