### 2019-04-14 4e22058
Simplify paper's filtering chain example
