### 2019-04-14 d733c5d
Attempt resize of paper figures
