### 2019-04-15 83fa649
multiblock weakref bug fix
