### 2019-04-15 75af32c
Further fix for #158
