### 2019-04-15 bcce52a
Remove volume form meta description of objects
