+++ vtki/common.py
@@ -766,8 +766,8 @@ class Common(DataSetFilters, object):
         attrs.append(("X Bounds", (bds[0], bds[1]), "{:.3e}, {:.3e}"))
         attrs.append(("Y Bounds", (bds[2], bds[3]), "{:.3e}, {:.3e}"))
         attrs.append(("Z Bounds", (bds[4], bds[5]), "{:.3e}, {:.3e}"))
-        if self.n_cells <= vtki.REPR_VOLUME_MAX_CELLS and self.n_cells > 0:
-            attrs.append(("Volume", (self.volume), "{:.3e}"))
+        # if self.n_cells <= vtki.REPR_VOLUME_MAX_CELLS and self.n_cells > 0:
+        #     attrs.append(("Volume", (self.volume), "{:.3e}"))
         return attrs
 
 
