### 2019-04-16 4c24a87
removed polydata smoothing
