### 2019-04-16 2cc54f9
Feature: Calculate geodesics on PolyData (#167) (#170)
* Feature: Calculate geodesics on PolyData (#167)

* Added functionality for calculating geodesics on polydata

* removed polydata smoothing

* Added tests for geodesics on polydata

* Add Geodesic path example
