+++ examples/01-filter/geodesic.py
@@ -8,7 +8,7 @@ Calculates the geodesic path betweeen two vertices using Dijkstra's algorithm
 import vtki
 from vtki import examples
 
-sphere = vtki.Sphere()
+sphere = examples.load_globe()
 
 ###############################################################################
 # Get teh geodesic path as a new :class:`vtki.PolyData` object:
@@ -19,8 +19,8 @@ geodesic = sphere.geodesic(0, sphere.n_points - 1)
 # Render the path along the sphere
 
 p = vtki.Plotter()
-p.add_mesh(geodesic, line_width=10, color='red', label='Sphere')
-p.add_mesh(sphere, show_edges=True, color=True, label='Geodesic Path')
+p.add_mesh(geodesic, line_width=10, color='red', label='Geodesic Path')
+p.add_mesh(sphere, show_edges=True, )
 p.camera_position = [-1,-1,1]
 p.add_legend()
 p.show()
