### 2019-04-16 dac2d58
Update resampling dataset filter
