### 2019-04-16 040519b
Update resample filter docs, add example, add test
