### 2019-04-16 2f8a8c1
Rename resample filter to `sample`
