### 2019-04-17 919f096
cleaned up .travis.yml
