+++ .travis.yml
@@ -78,8 +78,6 @@ script:
       make html;
     fi
   - cd ..
-  - echo $TRAVIS_REPO_SLUG
-  - echo $TRAVIS_BRANCH
   - |  # deploy only on the master branch for our repo
     if [[ ($TRAVIS_PYTHON_VERSION == 3.7) || ($TRAVIS_REPO_SLUG == 'vtkiorg/vtki') || ($TRAVIS_BRANCH == 'master') ]]; then
         doctr deploy --built-docs ./docs/_build/html .;
