+++ .travis.yml
@@ -79,7 +79,7 @@ script:
     fi
   - cd ..
   - |  # deploy only on the master branch for our repo
-    if [[ ($TRAVIS_PYTHON_VERSION == 3.7) || ($TRAVIS_REPO_SLUG == 'vtkiorg/vtki') || ($TRAVIS_BRANCH == 'master') ]]; then
+    if [[ ($TRAVIS_PYTHON_VERSION == 3.7) && ($TRAVIS_REPO_SLUG == 'vtkiorg/vtki') && ($TRAVIS_BRANCH == 'master') ]]; then
         doctr deploy --built-docs ./docs/_build/html .;
         cookiecutter -f --no-input --config-file vtki-binder-config.yml -o .. https://github.com/vtkiorg/cookiecutter-vtki-binder.git;
         rm -rf ../vtki-examples/notebooks/;
