### 2019-04-17 157cd5a
Attempt test plotting on AppVeyor
