### 2019-04-17 c938903
Add __str__ methods to all objects
