### 2019-04-18 5180973
Fix `set_position` bug
