### 2019-04-18 c4e0daa
Try install open GL on AppVeyor
