+++ appveyor.yml
@@ -25,8 +25,7 @@ init:
   - "ECHO %PYTHON% %PYTHON_VERSION% %PYTHON_ARCH%"
 
 install:
-  # https://help.appveyor.com/discussions/problems/3586-black-screenshots#comment_38618584
-  - ps: iex ((new-object net.webclient).DownloadString('https://raw.githubusercontent.com/appveyor/ci/master/scripts/enable-desktop.ps1'))
+  - "powershell ./.appveyor/install_opengl.ps1" # Install OpenGL
   - "%PYTHON% -m ensurepip"
   - "%PYTHON% -m pip install -r requirements.txt"
   - "%PYTHON% -m pip install pytest"
