### 2019-04-18 1802686
Edit temporary files so tests pass on Windows
