### 2019-04-18 7d1c03c
Install ffmpeg on AppVeyor
