### 2019-04-18 ef8149b
Fix AppVeyor Syntax
