### 2019-04-18 b307fc7
BUG: Escape char on re
