+++ vtki/errors.py
@@ -30,7 +30,7 @@ class Observer:
     @staticmethod
     def parse_message(message):
         # Message format
-        regex = re.compile('([A-Z]+):\sIn\s(.+),\sline\s.+\n\w+\s(.+):\s(.+)')
+        regex = re.compile(r'([A-Z]+):\sIn\s(.+),\sline\s.+\n\w+\s(.+):\s(.+)')
         kind, path, address, alert = regex.findall(message)[0]
         return kind, path, address, alert
 
