### 2019-04-18 0912986
Use a deploy script to fix #169
