+++ .travis.yml
@@ -78,24 +78,6 @@ script:
       make html;
     fi
   - cd ..
-  - |  # deploy only on the master branch for our repo
-    if [[ ($TRAVIS_PYTHON_VERSION == 3.7) && ($TRAVIS_REPO_SLUG == 'vtkiorg/vtki') && ($TRAVIS_BRANCH == 'master') ]]; then
-        doctr deploy --built-docs ./docs/_build/html .;
-        cookiecutter -f --no-input --config-file vtki-binder-config.yml -o .. https://github.com/vtkiorg/cookiecutter-vtki-binder.git;
-        rm -rf ../vtki-examples/notebooks/;
-        cd ./docs/;
-        find ./examples -type f -name '*.ipynb' | cpio -p -d -v ../../vtki-examples/;
-        cd ../../vtki-examples/;
-        git init;
-        git add .;
-        git commit -m "Autogenerated notebooks from Travis";
-        REMOTE="https://${GH_TOKEN}@github.com/vtkiorg/vtki-examples";
-        git config --global user.name "${GH_NAME}";
-        git config --global user.email "${GH_EMAIL}";
-        git remote add origin ${REMOTE};
-        git push -uf origin master;
-        cd ../vtki/
-    fi
 
 cache:
   directories:
@@ -109,6 +91,10 @@ deploy:
         condition: "$TRAVIS_PYTHON_VERSION == 3.6"
         tags: true
     password: $PYPI_PASSWORD
+  - provider: script
+    script: docs/deploy.sh
+    on:
+        branch: master
 
 notifications:
   email: false
