+++ appveyor.yml
@@ -41,7 +41,7 @@ build_script:
 # it's not apparent that the root directory changes
 # test_qt_plotting.py fails to collect in windows
 test_script:
-  - "del /Q tests\\test_qt_plotting.py"
+  - "del /Q tests\\test_qt_plotting.py" # TODO: run these tests
   - "cd tests"
   - "%PYTHON% -m pytest --cov -v"
   - "cd .."
