### 2019-04-18 964801b
Better control over plotting tests
