+++ tests/test_renderer.py
@@ -4,7 +4,7 @@ import pytest
 
 
 import vtki
-from vtki.plotting import running_xserver
+from vtki.plotting import system_supports_plotting
 
 if __name__ != '__main__':
     OFF_SCREEN = 'pytest' in sys.modules
@@ -12,7 +12,7 @@ else:
     OFF_SCREEN = False
 
 
-@pytest.mark.skipif(not running_xserver(), reason="Requires X11")
+@pytest.mark.skipif(not system_supports_plotting(), reason="Requires X11")
 def test_camera_position():
     plotter = vtki.Plotter(off_screen=OFF_SCREEN)
     plotter.add_mesh(vtki.Sphere())
