+++ tests/test_examples.py
@@ -7,7 +7,7 @@ import vtk
 
 import vtki
 from vtki import examples
-from vtki.plotting import running_xserver
+from vtki.plotting import system_supports_plotting
 
 TEST_DOWNLOADS = False
 try:
@@ -17,7 +17,7 @@ except KeyError:
     pass
 
 
-@pytest.mark.skipif(not running_xserver(), reason="Requires X11")
+@pytest.mark.skipif(not system_supports_plotting(), reason="Requires X11")
 def test_docexample_advancedplottingwithnumpy():
     import vtki
     import numpy as np
@@ -44,7 +44,7 @@ def test_docexample_advancedplottingwithnumpy():
     plotter.close()
 
 
-@pytest.mark.skipif(not running_xserver(), reason="Requires X11")
+@pytest.mark.skipif(not system_supports_plotting(), reason="Requires X11")
 def test_creatingagifmovie(tmpdir, off_screen=True):
     if tmpdir:
         filename = str(tmpdir.mkdir("tmpdir").join('wave.gif'))
@@ -84,18 +84,18 @@ def test_creatingagifmovie(tmpdir, off_screen=True):
     plotter.close()
 
 
-@pytest.mark.skipif(not running_xserver(), reason="Requires X11")
+@pytest.mark.skipif(not system_supports_plotting(), reason="Requires X11")
 def test_plot_wave():
     points = examples.plot_wave(wavetime=0.1, off_screen=True)
     assert isinstance(points, np.ndarray)
 
 
-@pytest.mark.skipif(not running_xserver(), reason="Requires X11")
+@pytest.mark.skipif(not system_supports_plotting(), reason="Requires X11")
 def test_beam_example():
     examples.beam_example(off_screen=True)
 
 
-@pytest.mark.skipif(not running_xserver(), reason="Requires X11")
+@pytest.mark.skipif(not system_supports_plotting(), reason="Requires X11")
 def test_plot_ants_plane():
     examples.plot_ants_plane(off_screen=True)
 
