### 2019-04-18 36ab3a4
Add conditional to docs deploy script
