### 2019-04-18 42e6a72
Fix travis typo
