### 2019-04-18 eb7fde5
Fix Travis script
