+++ .travis.yml
@@ -91,11 +91,12 @@ deploy:
         condition: "$TRAVIS_PYTHON_VERSION == 3.6"
         tags: true
     password: $PYPI_PASSWORD
+
   - provider: script
     script: docs/deploy.sh
     on:
         branch: master
-        condition: "($TRAVIS_PYTHON_VERSION == 3.7) && ($TRAVIS_REPO_SLUG == 'vtkiorg/vtki') && ($TRAVIS_BRANCH == 'master')"
+        condition: ($TRAVIS_PYTHON_VERSION == 3.7) && ($TRAVIS_REPO_SLUG == 'vtkiorg/vtki') && ($TRAVIS_BRANCH == 'master')
 
 notifications:
   email: false
