### 2019-04-18 dc26ff1
Fix vtkjs multi access issue during testing
