### 2019-04-18 3549759
Change error message for no X11
