+++ tests/test_plotting.py
@@ -26,7 +26,7 @@ sphere = vtki.Sphere()
 sphere_b = vtki.Sphere(1.0)
 sphere_c = vtki.Sphere(2.0)
 
-@pytest.mark.skipif(NO_PLOTTING, reason="Requires X11")
+@pytest.mark.skipif(NO_PLOTTING, reason="Requires system to support plotting")
 def test_plot(tmpdir):
     filename = os.path.join(vtki.USER_DATA_PATH, 'tmp.png')
     scalars = np.arange(sphere.n_points)
@@ -50,20 +50,20 @@ def test_plot(tmpdir):
     os.remove(filename)
 
 
-@pytest.mark.skipif(NO_PLOTTING, reason="Requires X11")
+@pytest.mark.skipif(NO_PLOTTING, reason="Requires system to support plotting")
 def test_plot_invalid_style():
     with pytest.raises(Exception):
         vtki.plot(sphere, style='not a style')
 
 
-@pytest.mark.skipif(NO_PLOTTING, reason="Requires X11")
+@pytest.mark.skipif(NO_PLOTTING, reason="Requires system to support plotting")
 def test_plot_bounds_axes_with_no_data():
     plotter = vtki.Plotter()
     plotter.show_bounds()
     plotter.close()
 
 
-@pytest.mark.skipif(NO_PLOTTING, reason="Requires X11")
+@pytest.mark.skipif(NO_PLOTTING, reason="Requires system to support plotting")
 def test_plot_show_grid():
     plotter = vtki.Plotter()
     plotter.show_grid()
@@ -71,7 +71,7 @@ def test_plot_show_grid():
     plotter.close()
 
 
-@pytest.mark.skipif(NO_PLOTTING, reason="Requires X11")
+@pytest.mark.skipif(NO_PLOTTING, reason="Requires system to support plotting")
 def test_set_camera_position():
     # with pytest.raises(Exception):
     cpos = [(2.085387555594636, 5.259683527170288, 13.092943022481887),
@@ -91,7 +91,7 @@ def test_set_camera_position():
     assert cpos_out == cpos
 
 
-@pytest.mark.skipif(NO_PLOTTING, reason="Requires X11")
+@pytest.mark.skipif(NO_PLOTTING, reason="Requires system to support plotting")
 def test_plot_no_active_scalars():
     plotter = vtki.Plotter(off_screen=OFF_SCREEN)
     plotter.add_mesh(sphere)
@@ -101,7 +101,7 @@ def test_plot_no_active_scalars():
         plotter.update_scalars(np.arange(sphere.n_faces))
 
 
-@pytest.mark.skipif(NO_PLOTTING, reason="Requires X11")
+@pytest.mark.skipif(NO_PLOTTING, reason="Requires system to support plotting")
 def test_plot_show_bounds():
     plotter = vtki.Plotter(off_screen=OFF_SCREEN)
     plotter.add_mesh(sphere)
@@ -122,7 +122,7 @@ def test_plot_show_bounds():
                             use_2d=True)
     plotter.plot()
 
-@pytest.mark.skipif(NO_PLOTTING, reason="Requires X11")
+@pytest.mark.skipif(NO_PLOTTING, reason="Requires system to support plotting")
 def test_plot_label_fmt():
     plotter = vtki.Plotter(off_screen=OFF_SCREEN)
     plotter.add_mesh(sphere)
@@ -130,7 +130,7 @@ def test_plot_label_fmt():
     plotter.plot()
 
 
-@pytest.mark.skipif(NO_PLOTTING, reason="Requires X11")
+@pytest.mark.skipif(NO_PLOTTING, reason="Requires system to support plotting")
 @pytest.mark.parametrize('grid', [True, 'both', 'front', 'back'])
[…]
