+++ tests/test_ipy_tools.py
@@ -15,7 +15,7 @@ except:
 
 
 
-@pytest.mark.skipif(not system_supports_plotting(), reason="Requires X11")
+@pytest.mark.skipif(not system_supports_plotting(), reason="Requires system to support plotting")
 @pytest.mark.skipif(not has_pyqt5, reason="requires pyqt5")
 def test_ipy_scaled_plotter(qtbot):
     data = examples.load_uniform()
@@ -23,7 +23,7 @@ def test_ipy_scaled_plotter(qtbot):
     p.add_mesh(data)
     p.close()
 
-@pytest.mark.skipif(not system_supports_plotting(), reason="Requires X11")
+@pytest.mark.skipif(not system_supports_plotting(), reason="Requires system to support plotting")
 @pytest.mark.skipif(not has_pyqt5, reason="requires pyqt5")
 def test_ipy_orthoganl_slicer(qtbot):
     data = examples.load_uniform()
@@ -33,7 +33,7 @@ def test_ipy_orthoganl_slicer(qtbot):
     tool.plotter.close()
 
 
-@pytest.mark.skipif(not system_supports_plotting(), reason="Requires X11")
+@pytest.mark.skipif(not system_supports_plotting(), reason="Requires system to support plotting")
 @pytest.mark.skipif(not has_pyqt5, reason="requires pyqt5")
 def test_ipy_many_slices_along_axis(qtbot):
     data = examples.load_uniform()
@@ -43,7 +43,7 @@ def test_ipy_many_slices_along_axis(qtbot):
     tool.plotter.close()
 
 
-@pytest.mark.skipif(not system_supports_plotting(), reason="Requires X11")
+@pytest.mark.skipif(not system_supports_plotting(), reason="Requires system to support plotting")
 @pytest.mark.skipif(not has_pyqt5, reason="requires pyqt5")
 def test_ipy_threshold(qtbot):
     data = examples.load_uniform()
@@ -53,7 +53,7 @@ def test_ipy_threshold(qtbot):
     tool.plotter.close()
 
 
-@pytest.mark.skipif(not system_supports_plotting(), reason="Requires X11")
+@pytest.mark.skipif(not system_supports_plotting(), reason="Requires system to support plotting")
 @pytest.mark.skipif(not has_pyqt5, reason="requires pyqt5")
 def test_ipy_clip(qtbot):
     data = examples.load_uniform()
@@ -63,7 +63,7 @@ def test_ipy_clip(qtbot):
     tool.plotter.close()
 
 
-@pytest.mark.skipif(not system_supports_plotting(), reason="Requires X11")
+@pytest.mark.skipif(not system_supports_plotting(), reason="Requires system to support plotting")
 @pytest.mark.skipif(not has_pyqt5, reason="requires pyqt5")
 def test_ipy_integrated(qtbot):
     data = examples.load_uniform()
@@ -75,7 +75,7 @@ def test_ipy_integrated(qtbot):
     clipper = vtki.Clip(data, plotter=p)
     p.close()
 
-@pytest.mark.skipif(not system_supports_plotting(), reason="Requires X11")
+@pytest.mark.skipif(not system_supports_plotting(), reason="Requires system to support plotting")
 @pytest.mark.skipif(not has_pyqt5, reason="requires pyqt5")
 def test_ipy_isocontour(qtbot):
     data = examples.load_uniform()

[…]
