### 2019-04-18 04aa629
Attempt fix travis docs deploy
