+++ docs/deploy.sh
@@ -1,4 +1,4 @@
-cd "$(dirname "$0")"
+cd "$(dirname "$0")";
 doctr deploy --built-docs ./_build/html .;
 cookiecutter -f --no-input --config-file vtki-binder-config.yml -o .. https://github.com/vtkiorg/cookiecutter-vtki-binder.git;
 rm -rf ../../vtki-examples/notebooks/;
@@ -6,10 +6,10 @@ find ./examples -type f -name '*.ipynb' | cpio -p -d -v ../../vtki-examples/;
 cd ../../vtki-examples/;
 git init;
 git add .;
-git commit -m "Autogenerated notebooks from Travis";
+git commit -m "Autogenerated notebooks from Travis build $TRAVIS_BUILD_NUMBER";
 REMOTE="https://${GH_TOKEN}@github.com/vtkiorg/vtki-examples";
 git config --global user.name "${GH_NAME}";
 git config --global user.email "${GH_EMAIL}";
 git remote add origin ${REMOTE};
 git push -uf origin master;
-cd ../vtki/
+cd ../vtki/;
