### 2019-04-18 6c5a0b0
attempt fix travis docs deploy
