+++ .travis.yml
@@ -99,7 +99,7 @@ deploy:
     password: $PYPI_PASSWORD
 
   - provider: script
-    script: docs/deploy.sh
+    script: docs/deploy.sh ${GH_TOKEN} ${GH_NAME} ${GH_EMAIL} ${TRAVIS_BUILD_NUMBER}
     on:
         branch: master
         condition: ($TRAVIS_PYTHON_VERSION == 3.7) && ($TRAVIS_REPO_SLUG == 'vtkiorg/vtki') && ($TRAVIS_BRANCH == 'master')
