### 2019-04-18 20b7c5e
Streamline requirements
