+++ requirements.txt
@@ -3,3 +3,10 @@ imageio<=2.4.1
 vtk
 matplotlib
 appdirs
+pytest
+pytest-cov
+codecov
+PyQt5==5.11.3
+pytest-qt
+nbval
+ipywidgets
