+++ .travis.yml
@@ -40,11 +40,8 @@ before_script:
 install:
   - pip install -r requirements.txt
   - pip install -r requirements_docs.txt
-  - pip install pytest-cov
-  - pip install codecov
   - |
     if [[ $TRAVIS_PYTHON_VERSION != 2.7 ]]; then
-      pip install -r requirements_test.txt
       pip install -U doctr
       pip install cookiecutter
     fi
