+++ docs/deploy.sh
@@ -1,17 +1,0 @@
-#!/bin/sh
-# Useage: deploy.sh ${GH_TOKEN} ${GH_NAME} ${GH_EMAIL} ${TRAVIS_BUILD_NUMBER}
-cd "$(dirname "$0")";
-doctr deploy --built-docs ./_build/html .;
-cookiecutter -f --no-input --config-file vtki-binder-config.yml -o ../.. https://github.com/vtkiorg/cookiecutter-vtki-binder.git;
-rm -rf ../../vtki-examples/notebooks/;
-find ./examples -type f -name '*.ipynb' | cpio -p -d -v ../../vtki-examples/;
-cd ../../vtki-examples/;
-git init;
-git add .;
-git commit -m "Autogenerated notebooks from Travis build $4";
-REMOTE="https://$1@github.com/vtkiorg/vtki-examples";
-git config --global user.name "$2";
-git config --global user.email "$3";
-git remote add origin ${REMOTE};
-git push -uf origin master;
-cd ../vtki/;
