+++ .travis.yml
@@ -84,18 +84,29 @@ script:
       make html;
     fi
   - cd ..
-
+  - |
+    if [[ ($TRAVIS_PYTHON_VERSION == 3.7) && ($TRAVIS_REPO_SLUG == 'vtkiorg/vtki') && ($TRAVIS_BRANCH == 'master') ]]; then
+        doctr deploy --built-docs ./docs/_build/html .;
+        cookiecutter -f --no-input --config-file vtki-binder-config.yml -o .. https://github.com/vtkiorg/cookiecutter-vtki-binder.git;
+        rm -rf ../vtki-examples/notebooks/;
+        cd ./docs/;
+        find ./examples -type f -name '*.ipynb' | cpio -p -d -v ../../vtki-examples/;
+        cd ../../vtki-examples/;
+        git init;
+        git add .;
+        git commit -m "Autogenerated notebooks from Travis";
+        REMOTE="https://${GH_TOKEN}@github.com/vtkiorg/vtki-examples";
+        git config --global user.name "${GH_NAME}";
+        git config --global user.email "${GH_EMAIL}";
+        git remote add origin ${REMOTE};
+        git push -uf origin master;
+        cd ../vtki/
+    fi
 cache:
   directories:
   - "$HOME/.cache/pip"
 
 deploy:
-  - provider: script
-    script: docs/deploy.sh ${GH_TOKEN} ${GH_NAME} ${GH_EMAIL} ${TRAVIS_BUILD_NUMBER}
-    skip_cleanup: true
-    on:
-        branch: master
-        condition: ($TRAVIS_PYTHON_VERSION == 3.7) && ($TRAVIS_REPO_SLUG == 'vtkiorg/vtki') && ($TRAVIS_BRANCH == 'master')
   - provider: pypi
     user: akaszynski
     distributions: sdist
@@ -103,7 +114,6 @@ deploy:
       condition: "$TRAVIS_PYTHON_VERSION == 3.6"
       tags: true
     password: $PYPI_PASSWORD
-
 
 notifications:
   email: false
