### 2019-04-18 90f0d7a
Stop testing on Python 2.7 per #164
