+++ .travis.yml
@@ -94,7 +94,7 @@ script:
         cd ../../vtki-examples/;
         git init;
         git add .;
-        git commit -m "Autogenerated notebooks from Travis";
+        git commit -m "$TRAVIS_BUILD_NUMBER : Autogenerated notebooks from Travis";
         REMOTE="https://${GH_TOKEN}@github.com/vtkiorg/vtki-examples";
         git config --global user.name "${GH_NAME}";
         git config --global user.email "${GH_EMAIL}";
