### 2019-04-19 63b5c37
Fix figure formatting for the JOSS paper
Captions should be in the `[]`, ie `![caption here](figure.png)`.
