### 2019-04-19 #1382 @leouieda
Hi @banesullivan, I've been going through the paper and I think it's a bit out of our usual format. Quoting from our [guidelines](https://joss.readthedocs.io/en/latest/submitting.html#what-should-my-paper-contain):

> ... JOSS papers are only permitted to contain a limited set of metadata (see header below), **Statement of Need, Summary, Acknowledgements, and References sections**. ... **software documentation** such as API (Application Programming Interface) functionality **should not be in the paper** and instead should be outlined in the software documentation.

Could you please reformat the paper before we proceed with the review? The information is already in the paper so it's only a matter of cutting and reorganizing. You don't have to remove all the figures. In particular, make sure there are:

* A summary describing the high-level functionality and purpose of the software for a **diverse, non-specialist audience**
* A clear statement of need that illustrates the purpose of the software

It would also help if you clarified wether the citations in Mentions are actually using vtki or if they are examples of possible applications of vtki.
