### 2019-04-19 6e2270d
Edit JOSS paper per pre-review
