+++ joss/paper.md
@@ -23,39 +23,30 @@ bibliography: paper.bib
 
 # Summary
 
-The Visualization Toolkit (VTK) is an excellent visualization software
-library, and with Python bindings it combines the speed of C++ with
-the rapid prototyping of Python [@vtkbook].  Despite this, VTK code
-programmed in Python using the base package provided by Kitware is
-unnecessarily complex as the Python package merely wraps existing C++
-calls. The `vtki` Python package seeks to simplify common mesh creation
-and plotting routines without compromising on the speed of the C++ VTK
-backend, enabling researchers to rapidly explore large datasets,
-communicate their spatial findings, and facilitate reproducibility.
-Figure 1 demonstrates an integrated scene of geospatial data generated
-by `vtki`; to learn more about how this visualization was created,
+The availability of software for 3D visualization in Python is limited to
+libraries that are either not capable of handling large, spatially referenced
+datasets or have inherently complex application programming interfaces (APIs).
+One of these software libraries is the Visualization Toolkit (VTK) [@vtkbook];
+a powerful scientific visualization software library, and with Python bindings,
+it combines the speed of C++ with the rapid prototyping of Python.
+Despite this, VTK code programmed in Python using the base VTK python package
+is unnecessarily complicated as its API merely binds existing C++ calls.
+The `vtki` Python package provides a concise, well-documented interface
+exposing VTK’s powerful visualization backend; enabling researchers to
+rapidly explore large datasets, communicate their spatial findings, and
+facilitate reproducibility. `vtki` further seeks to simplify standard mesh
+creation and plotting routines without compromising on the speed of the C++
+VTK backend. Figure 1 demonstrates an integrated scene of geospatial data
+generated by `vtki`; to learn more about how this visualization was created,
 please visit [PVGeo's FORGE project website](http://forge.pvgeo.org).
 
 
-![A visually integrated scene of geospatial data (FORGE Geothermal Site). 
-This rendering includes a digital land surface with overlain satellite 
-imagery and geologic map, a subsurface temperature model, scattered points 
-of the sampled temperature values, geophysical well logging data, GIS site 
+![A visually integrated scene of geospatial data (FORGE Geothermal Site).
+This rendering includes a digital land surface with overlain satellite
+imagery and geologic map, a subsurface temperature model, scattered points
+of the sampled temperature values, geophysical well logging data, GIS site
 boundary, and interpreted faulting surfaces.](./images/forge-iso.png)
 
-
-
-At its core, `vtki` is a pure Python helper module for VTK
-that interfaces back to VTK data objects through NumPy [@numpy]
-and direct array access adhering to VTK's object oriented approach to
-3D visualization [@vtkbook]. This package expands upon VTK's data types
-by creating classes that extend their VTK counterpart.
-The `vtki` wrapping layer provides an accessible and intuitive interface back
-to the VTK library to facilitate rapid prototyping, analysis, and visual
-integration of spatially referenced datasets.
-
-
-## Simplified Plotting Routines
 
 Plotting VTK datasets using only the VTK Python package or a similar
 visualization library is often an ambitious programming endeavor.
@@ -63,68 +54,28 @@ Reading a VTK supported file and plotting it requires a user to write a
 complicated sequence of routines to render the data object while
 having to remember which VTK classes to use for file reading and dataset mapping.
 An example can be found in [this creative commons VTK example](https://vtk.org/Wiki/VTK/Examples/Python/STLReader).
-
 `vtki` includes plotting routines that are intended to be intuitive and
 highly controllable with `matplotlib` [@matplotlib] similar syntax and keyword
 arguments. These plotting routines are defined to make the process of
[…]
