### 2019-04-19 4c7d9a1
Update JOSS code example file
