### 2019-04-19 47fc023
Updates to JOSS paper
