### 2019-04-19 cdcd7cc
Update filters figure for JOSS paper
