### 2019-04-19 4bb4248
Update AFRL mentions
