### 2019-04-19 #1382 @banesullivan
Thank you for the feedback @leouieda. I have made updates to the paper addressing your concerns.

1. On the note of the paper's format:

    - I have significantly trimmed the content and made sure the paper only contains: *Statement of Need, Summary, Mentions, and References sections*
    - I have removed all code examples to adhere to JOSS' criteria that *"software documentation such as API (Application Programming Interface) functionality should not be in the paper and instead should be outlined in the software documentation"*

2. On the note of ensuring there is a *"summary describing the high-level functionality and purpose of the software for a diverse, non-specialist audience"*:

    - I have rewritten the summary to describe the high-level purpose of why we built `vtki` and what need it fills. Particularly, the paper now focuses on what `vtki` achieves and what that looks like for a novice programmer/new user rather than the nuances of how it is implemented.

3. On the note of ensuring there is a *"clear statement of need that illustrates the purpose of the software"*

    - The paper now conveys a more clear narrative of the current state of 3D visualization in Python and what role `vtki` fills in that space.
    - Specifically, I focused on including this in the first paragraph of the summary:
        > The availability of software for 3D visualization in Python is limited to libraries that are either not capable of handling large, spatially referenced datasets or have inherently complex application programming interfaces (APIs). One of these software libraries is the Visualization Toolkit (VTK) [@vtkbook]; a powerful scientific visualization software library, and with Python bindings, it combines the speed of C++ with the rapid prototyping of Python. Despite this, VTK code programmed in Python using the base VTK python package is unnecessarily complicated as its API merely binds existing C++ calls. The vtki Python package provides a concise, well-documented interface exposing VTK’s powerful visualization backend; enabling researchers to rapidly explore large datasets, communicate their spatial findings, and facilitate reproducibility.

4. I have clarified the mentions to state that the listed publications do indeed leverage `vtki` for 3D visualization.
