### 2019-04-19 bf6b5fd
Specify supported Python versions per #164
