+++ README.rst
@@ -44,6 +44,7 @@ papers as well as a supporting module for other mesh dependent Python modules.
 
 Documentation
 =============
+
 Refer to the `documentation <http://docs.vtki.org/>`_ for detailed
 installation and usage details.
 
@@ -58,7 +59,9 @@ of the developers an email. The project support team can be reached at
 
 Installation
 ============
-Installation is simply::
+
+``vtki`` can be installed from `PyPI <http://pypi.python.org/pypi/vtki>`_
+using ``pip`` on Python >= 3.5::
 
     pip install vtki
 
