### 2019-04-19 3db9b33
Update geometric objects example
