### 2019-04-19 79a3f50
Remove unused code from parallel example
