### 2019-04-19 f48eaee
Fix `set_position` bug (#173)
