### 2019-04-19 02753f9
Handle plotting if matplotlib is not installed
