### 2019-04-19 57423c3
Fix geometric object example
