### 2019-04-19 4376a6d
Use source on travis gl-ci-helper script
