+++ .travis.yml
@@ -20,7 +20,7 @@ before_script:
   - echo $TRAVIS_REPO_SLUG
   # configure a headless display
   - git clone --depth 1 git://github.com/vtkiorg/gl-ci-helpers.git
-  - sh gl-ci-helpers/travis/setup_headless_display.sh
+  - source ./gl-ci-helpers/travis/setup_headless_display.sh
 
 install:
   - pip install -r requirements.txt
