### 2019-04-19 cafe5c9
Reorganize Travis script
