### 2019-04-19 6186588
Update CI installation details in docs
