### 2019-04-19 aafeff3
Use gl-ci-helpers on Travis (#181)
* Use gl-ci-helpers on Travis

* Use source on travis gl-ci-helper script

* Reorganize Travis script

* Update CI installation details in docs
