+++ .travis.yml
@@ -60,29 +60,17 @@ script:
   - make doctest;
   - make html;
   - cd ..
-  - |
-    if [[ ($TRAVIS_PYTHON_VERSION == 3.7) && ($TRAVIS_REPO_SLUG == 'vtkiorg/vtki') && ($TRAVIS_BRANCH == 'master') && ($DOCS_TRIGGER == 'BANE_RULES')]]; then
-        doctr deploy --built-docs ./docs/_build/html .;
-        cookiecutter -f --no-input --config-file vtki-binder-config.yml -o .. https://github.com/vtkiorg/cookiecutter-vtki-binder.git;
-        rm -rf ../vtki-examples/notebooks/;
-        cd ./docs/;
-        find ./examples -type f -name '*.ipynb' | cpio -p -d -v ../../vtki-examples/;
-        cd ../../vtki-examples/;
-        git init;
-        git add .;
-        git commit -m "$TRAVIS_BUILD_NUMBER : Autogenerated notebooks from Travis";
-        REMOTE="https://${GH_TOKEN}@github.com/vtkiorg/vtki-examples";
-        git config --global user.name "${GH_NAME}";
-        git config --global user.email "${GH_EMAIL}";
-        git remote add origin ${REMOTE};
-        git push -uf origin master;
-        cd ../vtki/
-    fi
+
 cache:
   directories:
   - "$HOME/.cache/pip"
 
 deploy:
+  - provider: script
+    script: sh ./docs/deploy_docs_from_travis.sh
+    skip_cleanup: true
+    on:
+      condition: ($TRAVIS_PYTHON_VERSION == 3.7) && ($TRAVIS_REPO_SLUG == 'vtkiorg/vtki') && ($TRAVIS_BRANCH == 'master') && ($DOCS_TRIGGER == 'BANE_RULES')
   - provider: pypi
     user: akaszynski
     distributions: sdist
