### 2019-04-20 fa66ef4
Fix rare error parsing bug
