### 2019-04-20 47cb402
Repeat build matrix to test across Python versions
