### 2019-04-20 f09a1d4
Only deploy to PyPI on Python 3.7
