+++ .travis.yml
@@ -126,7 +126,7 @@ deploy:
     user: akaszynski
     distributions: sdist
     on:
-      condition: "$TRAVIS_PYTHON_VERSION == 3.6"
+      condition: "$TRAVIS_PYTHON_VERSION == 3.7"
       tags: true
     password: $PYPI_PASSWORD
 
