### 2019-04-20 c6f3443
Remove Python 2.7 tests
