### 2019-04-20 c545197
Only run documentation tests on Python 3.7
