### 2019-04-20 0b5dc16
Report coverage for all 3.7 builds with coverage reports
