### 2019-04-20 9b9c9a0
Debug coverage reporting
