### 2019-04-20 f40762f
Attempt fix codecov conditional
