### 2019-04-20 7594881
report coverage on module doctests
