### 2019-04-20 069a4ae
[Travis]: Update build matrix (#182)
* Fix interactor style bugs

* Allow downloads failures

* Remove fast finish

* Update travis build matrix

* Attempt use script matrix

* Attempt use nested python jobs in build matrix

* Repeat build matrix to test across Python versions

* Only deploy to PyPI on Python 3.7

* Remove Python 2.7 tests

* Only run documentation tests on Python 3.7

* Report coverage for all 3.7 builds with coverage reports

* Debug coverage reporting

* Attempt fix codecov conditional

* Attempt fix codecov syntax

* report coverage on module doctests
