+++ .travis.yml
@@ -1,12 +1,12 @@
 sudo: required
+
 dist: xenial
+
 os:
   - linux
+
 language: python
-python:
-  - '3.5'
-  - '3.6'
-  - '3.7'
+
 env:
   global:
     # Doctr deploy key for vtkiorg/vtki
@@ -35,31 +35,77 @@ before_script:
   - echo $TRAVIS_BUILD_NUMBER
   - echo $TRAVIS_REPO_SLUG
 
-script:
-  # unset VTKI_OFFSCREEN which is set by travis CI helper
-  - unset VTKI_OFFSCREEN
-  # Run the test suite and generate coverage report
-  - pytest -v --cov vtki
-  # Run all code examples in the docstrings
-  - pytest -v --doctest-modules vtki
-  - |  # report code coverage
-    if [[ $TRAVIS_PYTHON_VERSION == 3.7 ]]; then
-      codecov
+matrix:
+  include:
+    # Test the core API on Python 3.5, 3.6, and 3.7
+    - name: Core API 3.5
+      python: 3.5
+      # Run the test suite and generate coverage report
+      script: pytest -v --cov vtki
+    - name: Core API 3.6
+      python: 3.6
+      # Run the test suite and generate coverage report
+      script: pytest -v --cov vtki
+    - name: Core API 3.7
+      python: 3.7
+      env:
+        - REPORT_COVERAGE: true
+      # Run the test suite and generate coverage report
+      script: pytest -v --cov vtki
+
+    # Test the package docstrings on Python 3.5, 3.6, and 3.7
+    - name: Package Docstrings 3.5
+      python: 3.5
+      # Run all code examples in the docstrings
+      script: pytest -v --doctest-modules vtki
+    - name: Package Docstrings 3.6
+      python: 3.6
+      # Run all code examples in the docstrings
+      script: pytest -v --cov --doctest-modules vtki
+    - name: Package Docstrings 3.7
+      python: 3.7
+      env:
+        - REPORT_COVERAGE: true
+      # Run all code examples in the docstrings
+      script: pytest -v --cov --doctest-modules vtki
+
+    # Test notebooks on Python 3.6 and 3.7
+    - name: Notebooks 3.6
+      python: 3.6
+      # Now make sure notebooks are running
+      # TODO: add below: pytest -v --nbval-lax --current-env --disable-warnings tests/*.ipynb;
+      script: pytest -v --nbval-lax --current-env --disable-warnings notebooks/*.ipynb;
+    - name: Notebooks 3.7
+      python: 3.7
+      env:
+        - REPORT_COVERAGE: false
+      # Now make sure notebooks are running
+      # TODO: add below: pytest -v --nbval-lax --current-env --disable-warnings tests/*.ipynb;
+      script: pytest -v --nbval-lax --current-env --disable-warnings notebooks/*.ipynb;
+
+    # Build/test documentation on 3.7 only
+    - name: Documentation and Deploy 3.7
+      python: 3.7
+      env:
+        - REPORT_COVERAGE: false
+      # Run `make html` before `make doctest` to avoid segfault.
+      # Then rerun `make html` to use updated figures from `make doctest`
+      script: |
+        cd ./docs/;
+        export VTKI_OFFSCREEN=True;
+        set -e;
+        make html;
+        make doctest;
+        make html;
+        cd ..;
+
+
+
+after_success:
+  - if [[($TRAVIS_PYTHON_VERSION == 3.7) && ($REPORT_COVERAGE == 'true')]]; then
+      echo "Uploading coverage to Codecov";
+      codecov;
     fi
-  # Now make sure notebooks are running
[…]
