### 2019-04-21 34be73f
Fix environmental variable checking
