### 2019-04-22 f9f4242
Add transparent_background to rcParams
