### 2019-04-22 5bf86c2
updated argument options of polydata smoothing function
