### 2019-04-23 bc4824a
Address #186 by not calling to renderer after pressing close button
