### 2019-04-23 84f8808
Update notebook testing and test with panel installed
