+++ .travis.yml
@@ -74,14 +74,31 @@ matrix:
       python: 3.6
       # Now make sure notebooks are running
       # TODO: add below: pytest -v --nbval-lax --current-env --disable-warnings tests/*.ipynb;
-      script: pytest -v --nbval-lax --current-env --disable-warnings notebooks/*.ipynb;
+      script: |
+        pytest -v --nbval-lax --current-env --disable-warnings tests/*.ipynb;
+        pytest -v --nbval-lax --current-env --disable-warnings notebooks/*.ipynb;
+    - name: Notebooks 3.6 with panel
+      python: 3.6
+      # Now make sure notebooks are running
+      script: |
+        pip install panel;
+        pytest -v --nbval-lax --current-env --disable-warnings tests/*.ipynb;
+        pytest -v --nbval-lax --current-env --disable-warnings notebooks/*.ipynb;
+    - name: Notebooks 3.7 with panel
+      python: 3.7
+      # Now make sure notebooks are running
+      script: |
+        pip install panel;
+        pytest -v --nbval-lax --current-env --disable-warnings tests/*.ipynb;
+        pytest -v --nbval-lax --current-env --disable-warnings notebooks/*.ipynb;
     - name: Notebooks 3.7
       python: 3.7
       env:
         - REPORT_COVERAGE: false
       # Now make sure notebooks are running
-      # TODO: add below: pytest -v --nbval-lax --current-env --disable-warnings tests/*.ipynb;
-      script: pytest -v --nbval-lax --current-env --disable-warnings notebooks/*.ipynb;
+      script: |
+        pytest -v --nbval-lax --current-env --disable-warnings tests/*.ipynb;
+        pytest -v --nbval-lax --current-env --disable-warnings notebooks/*.ipynb;
 
     # Build/test documentation on 3.7 only
     - name: Documentation and Deploy 3.7
