### 2019-04-23 ea9e4b5
Allow panel tests to fail
