### 2019-04-23 458704b
added decimate_pro method
