+++ vtki/pointset.py
@@ -573,14 +573,15 @@ class PolyData(vtkPolyData, vtki.Common):
         trifilter.PassVertsOff()
         trifilter.PassLinesOff()
         trifilter.Update()
-        if inplace:
-            self.overwrite(trifilter.GetOutput())
-        else:
-            return PolyData(trifilter.GetOutput())
 
+        mesh = _get_output(trifilter)
+        if inplace:
+            self.overwrite(mesh)
+        else:
+            return mesh
 
     def smooth(self, n_iter=20, convergence=0.0, edge_angle=15, feature_angle=45,
-               boundary_smoothing=True, feature_smoothing=False):
+               boundary_smoothing=True, feature_smoothing=False, inplace=False):
         """Adjust point coordinates using Laplacian smoothing.
         The effect is to "relax" the mesh, making the cells better shaped and
         the vertices more evenly distributed.
@@ -606,6 +607,14 @@ class PolyData(vtkPolyData, vtki.Common):
         feature_smoothing : bool, optional
             Boolean flag to control smoothing of feature edges.
 
+        inplace : bool, optional
+            Updates mesh in-place while returning nothing.
+
+        Returns
+        -------
+        mesh : vtki.PolyData
+            Decimated mesh. None when inplace=True.
+
         """
         alg = vtk.vtkSmoothPolyDataFilter()
         alg.SetInputData(self)
@@ -617,7 +626,11 @@ class PolyData(vtkPolyData, vtki.Common):
         alg.SetBoundarySmoothing(boundary_smoothing)
         alg.Update()
 
-        return _get_output(alg)
+        mesh = _get_output(alg)
+        if inplace:
+            self.overwrite(mesh)
+        else:
+            return mesh
 
     def decimate_pro(self, reduction, feature_angle=45.0, split_angle=75.0, splitting=True,
                      pre_split_mesh=False, preserve_topology=False, inplace=False):
@@ -682,7 +695,7 @@ class PolyData(vtkPolyData, vtki.Common):
             return mesh
 
     def tube(self, radius=None, scalars=None, capping=True, n_sides=20,
-             radius_factor=10, preference='point'):
+             radius_factor=10, preference='point', inplace=False):
         """Generate a tube around each input line. The radius of the tube can be
         set to linearly vary with a scalar value.
 
@@ -705,6 +718,15 @@ class PolyData(vtkPolyData, vtki.Common):
 
         preference : str
             The field preference when searching for the scalar array by name
+
+        inplace : bool, optional
+            Updates mesh in-place while returning nothing.
+
+        Returns
+        -------
+        mesh : vtki.PolyData
+            Tube-filtered mesh. None when inplace=True.
+
         """
         if n_sides < 3:
             n_sides = 3
@@ -726,7 +748,12 @@ class PolyData(vtkPolyData, vtki.Common):
             tube.SetVaryRadiusToVaryRadiusByScalar()
         # Apply the filter
         tube.Update()
-        return _get_output(tube)
+
+        mesh = _get_output(tube)
+        if inplace:
+            self.overwrite(mesh)
+        else:
+            return mesh
 
     def subdivide(self, nsub, subfilter='linear', inplace=False):
         """
@@ -792,7 +819,8 @@ class PolyData(vtkPolyData, vtki.Common):
         sfilter.SetNumberOfSubdivisions(nsub)
         sfilter.SetInputData(self)
         sfilter.Update()
-        submesh = PolyData(sfilter.GetOutput())
+
+        submesh = _get_output(sfilter)
         if inplace:
             self.overwrite(submesh)
         else:
@@ -800,7 +828,7 @@ class PolyData(vtkPolyData, vtki.Common):
 
[…]
