### 2019-04-23 8d11ec3
updated inplace arguments in test
