### 2019-04-23 60d49f8
Fix issues introduced
