### 2019-04-23 3f7f715
Refactor in_place -> inplace
