### 2019-04-23 b503f76
do not call to renderer after pressing close button (#190)
* Address #186 by not calling to renderer after pressing close button

* Fix IPython import

* Update notebook testing and test with panel installed

* Allow panel tests to fail
