### 2019-04-23 c3ce9ca
Add lots more example data download functions
