+++ tests/test_examples.py
@@ -182,4 +182,256 @@ if TEST_DOWNLOADS:
         data = examples.download_blood_vessels()
         assert isinstance(data, vtki.UnstructuredGrid)
 
+    def test_download_bunny_coarse():
+        data = examples.download_bunny_coarse()
+        assert data.n_cells
+
+    def test_download_cow_head():
+        data = examples.download_cow_head()
+        assert data.n_cells
+
+    def test_download_knee_full():
+        data = examples.download_knee_full()
+        assert data.n_cells
+
+    def test_download_iron_pot():
+        data = examples.download_iron_pot()
+        assert data.n_cells
+
+    def test_download_tetrahedron():
+        data = examples.download_tetrahedron()
+        assert data.n_cells
+
+    def test_download_saddle_surface():
+        data = examples.download_saddle_surface()
+        assert data.n_cells
+
+    def test_download_foot_bones():
+        data = examples.download_foot_bones()
+        assert data.n_cells
+
+    def test_download_guitar():
+        data = examples.download_guitar()
+        assert data.n_cells
+
+    def test_download_quadratic_pyramid():
+        data = examples.download_quadratic_pyramid()
+        assert data.n_cells
+
+    def test_download_quadratic_pyramid():
+        data = examples.download_quadratic_pyramid()
+        assert data.n_cells
+
+    def test_download_bird():
+        data = examples.download_bird()
+        assert data.n_cells
+
+    def test_download_bird_texture():
+        data = examples.download_bird_texture()
+        assert isinstance(data, vtk.vtkTexture)
+
+    def test_download_office():
+        data = examples.download_office()
+        assert data.n_cells
+
+
+    def test_download_horse_points():
+        data = examples.download_horse_points()
+        assert data.n_points
+
+    def test_download_horse():
+        data = examples.download_horse()
+        assert data.n_cells
+
+    def test_download_cake_easy():
+        data = examples.download_cake_easy()
+        assert data.n_cells
+
+    def test_download_cake_easy_texture():
+        data = examples.download_cake_easy_texture()
+        assert isinstance(data, vtk.vtkTexture)
+
+    def test_download_rectilinear_grid():
+        data = examples.download_rectilinear_grid()
+        assert data.n_cells
+
+    def test_download_gourds():
+        data = examples.download_gourds()
+        assert data.n_cells
+        data = examples.download_gourds(zoom=True)
+        assert data.n_cells
+
+    def test_download_gourds_texture():
+        data = examples.download_gourds_texture()
+        assert isinstance(data, vtk.vtkTexture)
+        data = examples.download_gourds_texture(zoom=True)
+        assert isinstance(data, vtk.vtkTexture)
+
+    def test_download_unstructured_grid():
+        data = examples.download_unstructured_grid()
+        assert data.n_cells
+
+
+    def test_download_letter_k():
+        data = examples.download_letter_k()
+        assert data.n_cells
+
+    def test_download_letter_a():
+        data = examples.download_letter_a()
+        assert data.n_cells
+
+    def test_download_poly_line():
+        data = examples.download_poly_line()
+        assert data.n_cells
+
+    def test_download_cad_model():
+        data = examples.download_cad_model()
+        assert data.n_cells
+
+    def test_download_frog():
+        data = examples.download_frog()
+        assert data.n_cells
+
+    def test_download_prostate():
+        data = examples.download_prostate()
+        assert data.n_cells
+
+    def test_download_filled_contours():
+        data = examples.download_filled_contours()
+        assert data.n_cells
+
+    def test_download_doorman():
+        data = examples.download_doorman()
+        assert data.n_cells
+
+    def test_download_mug():
+        data = examples.download_mug()
+        assert data.n_blocks
+
[…]
