### 2019-04-24 e5ffcb9
Add lots more example data download functions (#195)
