+++ vtki/readers.py
@@ -38,7 +38,8 @@ READERS = {
     '.tiff': vtk.vtkTIFFReader,
     '.tif': vtk.vtkTIFFReader,
     # Other formats:
-    '.byu': vtk.vtkBYUReader, # TODO: not tested
+    '.byu': vtk.vtkBYUReader, # TODO: not tested with this extension
+    '.g': vtk.vtkBYUReader,
     # '.chemml': vtk.vtkCMLMoleculeReader, # TODO: not tested
     # '.cml': vtk.vtkCMLMoleculeReader, # vtkMolecule is not supported by vtki
     # TODO: '.csv': vtk.vtkCSVReader, # vtkTables are currently not supported
@@ -54,6 +55,8 @@ READERS = {
     # '.particles': vtk.vtkParticleReader, # TODO: not tested
     #TODO: '.pht': vtk.vtkPhasta??????,
     #TODO: '.vpc': vtk.vtkVPIC?????,
+    # '.bin': vtk.vtkMultiBlockPLOT3DReader,# TODO: non-default routine
+    # '.tri': vtk.vtkMCubesReader, # TODO: Breaks vtki wrapping on point arrays
 }
 
 
