+++ vtki/examples/downloads.py
@@ -56,7 +56,7 @@ def _download_and_read(filename, texture=False):
     return vtki.read(saved_file)
 
 
-# Textures:
+###############################################################################
 
 def download_masonry_texture():
     return _download_and_read('masonry.bmp', texture=True)
@@ -66,9 +66,6 @@ def download_usa_texture():
 
 def download_puppy_texture():
     return _download_and_read('puppy.jpg', texture=True)
-
-
-# Examples:
 
 def download_puppy():
     return _download_and_read('puppy.jpg')
@@ -82,8 +79,14 @@ def download_st_helens():
 def download_bunny():
     return _download_and_read('bunny.ply')
 
+def download_bunny_coarse():
+    return _download_and_read('Bunny.vtp')
+
 def download_cow():
     return _download_and_read('cow.vtp')
+
+def download_cow_head():
+    return _download_and_read('cowHead.vtp')
 
 def download_faults():
     return _download_and_read('faults.vtk')
@@ -116,6 +119,9 @@ def download_coastlines():
 def download_knee():
     return _download_and_read('DICOM_KNEE.dcm')
 
+def download_knee_full():
+    return _download_and_read('vw_knee.slc')
+
 def download_lidar():
     return _download_and_read('kafadar-lidar-interp.vtp')
 
@@ -132,3 +138,166 @@ def download_blood_vessels():
     local_path, _ = _download_file('pvtu_blood_vessels/blood_vessels.zip')
     filename = os.path.join(local_path, 'T0000000500.pvtu')
     return vtki.read(filename)
+
+def download_iron_pot():
+    return _download_and_read('ironProt.vtk')
+
+def download_tetrahedron():
+    return _download_and_read('Tetrahedron.vtu')
+
+def download_saddle_surface():
+    return _download_and_read('InterpolatingOnSTL_final.stl')
+
+def download_foot_bones():
+    return _download_and_read('fsu/footbones.ply')
+
+def download_guitar():
+    return _download_and_read('fsu/stratocaster.ply')
+
+def download_quadratic_pyramid():
+    return _download_and_read('QuadraticPyramid.vtu')
+
+def download_bird():
+    return _download_and_read('Pileated.jpg')
+
+def download_bird_texture():
+    return _download_and_read('Pileated.jpg', texture=True)
+
+def download_office():
+    return _download_and_read('office.binary.vtk')
+
+def download_horse_points():
+    return _download_and_read('horsePoints.vtp')
+
+def download_horse():
+    return _download_and_read('horse.vtp')
+
+def download_cake_easy():
+    return _download_and_read('cake_easy.jpg')
+
+def download_cake_easy_texture():
+    return _download_and_read('cake_easy.jpg', texture=True)
+
+def download_rectilinear_grid():
+    return _download_and_read('RectilinearGrid.vtr')
+
+def download_gourds(zoom=False):
+    if zoom:
+        return _download_and_read('Gourds.png')
+    return _download_and_read('Gourds2.jpg')
+
+def download_gourds_texture(zoom=False):
+    if zoom:
+        return _download_and_read('Gourds.png', texture=True)
+    return _download_and_read('Gourds2.jpg', texture=True)
+
+def download_unstructured_grid():
+    return _download_and_read('uGridEx.vtk')
+
+def download_letter_k():
+    return _download_and_read('k.vtk')
+
+def download_letter_a():
+    return _download_and_read('a.vtp')
+
+def download_poly_line():
+    return _download_and_read('polyline.vtk')
+
+def download_cad_model():
+    return _download_and_read('42400-IDGH.stl')
+
+def download_frog():
+    # TODO: there are other files with this
+    return _download_and_read('Frog/frog.mhd')
+
+def download_prostate():
[…]
