### 2019-04-25 2aebbf4
PolyData filters updates (#194)
@supersubscript's work

* Added functionality for calculating geodesics on polydata

* removed polydata smoothing

* Added tests for geodesics on polydata

* updated argument options of polydata smoothing function

* added decimate_pro method

* updated inplace arguments for consistency. Now defaulting to False

* removed old tests

* updated inplace arguments in test

* missing _pro

* Fix issues introduced

* Refactor in_place -> inplace
