+++ tests/test_common.py
@@ -416,7 +416,7 @@ def test_axis_rotation_invalid():
         vtki.common.axis_rotation(np.empty((3, 3)), 0, False, axis='not')
 
 
-def test_axis_rotation_not_in_place():
+def test_axis_rotation_not_inplace():
     p = np.eye(3)
     p_out = vtki.common.axis_rotation(p, 1, False, axis='x')
     assert not np.allclose(p, p_out)
