+++ tests/test_polydata.py
@@ -9,12 +9,12 @@ from vtki import examples
 from vtki.plotting import system_supports_plotting
 
 radius = 0.5
-sphere = vtki.Sphere(radius, theta_resolution=10, phi_resolution=10)
+SPHERE = vtki.Sphere(radius, theta_resolution=10, phi_resolution=10)
 
-sphere_shifted = vtki.Sphere(center=[0.5, 0.5, 0.5],
+SPHERE_SHIFTED = vtki.Sphere(center=[0.5, 0.5, 0.5],
                              theta_resolution=10, phi_resolution=10)
 
-dense_sphere = vtki.Sphere(radius, theta_resolution=100, phi_resolution=100)
+SPHERE_DENSE = vtki.Sphere(radius, theta_resolution=100, phi_resolution=100)
 
 try:
     test_path = os.path.dirname(os.path.abspath(__file__))
@@ -37,6 +37,7 @@ def test_init():
 
 
 def test_init_from_pdata():
+    sphere = SPHERE.copy()
     mesh = vtki.PolyData(sphere, deep=True)
     assert mesh.n_points
     assert mesh.n_cells
@@ -132,14 +133,19 @@ def test_invalid_file():
         # vtki.PolyData(examples.hexbeamfile)
 
 def test_geodesic():
+    sphere = SPHERE.copy()
     geodesic = sphere.geodesic(0, sphere.n_points - 1)
     assert isinstance(geodesic, vtki.PolyData)
 
+
 def test_geodesic_distance():
+    sphere = SPHERE.copy()
     distance = sphere.geodesic_distance(0, sphere.n_points - 1)
     assert isinstance(distance, float)
 
+
 def test_ray_trace():
+    sphere = SPHERE.copy()
     points, ind = sphere.ray_trace([0, 0, 0], [1, 1, 1])
     assert np.any(points)
     assert np.any(ind)
@@ -147,6 +153,7 @@ def test_ray_trace():
 
 @pytest.mark.skipif(not system_supports_plotting(), reason="Requires system to support plotting")
 def test_ray_trace_plot():
+    sphere = SPHERE.copy()
     points, ind = sphere.ray_trace([0, 0, 0], [1, 1, 1], plot=True, first_point=True,
                                    off_screen=True)
     assert np.any(points)
@@ -155,28 +162,35 @@ def test_ray_trace_plot():
 
 @pytest.mark.skipif(not system_supports_plotting(), reason="Requires system to support plotting")
 def test_plot_curvature():
+    sphere = SPHERE.copy()
     cpos = sphere.plot_curvature(off_screen=True)
     assert isinstance(cpos, list)
 
 
 def test_edge_mask():
+    sphere = SPHERE.copy()
     mask = sphere.edge_mask(10)
 
 
 def test_boolean_cut_inplace():
-    sub_mesh = sphere.copy()
+    sub_mesh = SPHERE.copy()
+    sphere_shifted = SPHERE_SHIFTED.copy()
     sub_mesh.boolean_cut(sphere_shifted, inplace=True)
     assert sub_mesh.n_points
     assert sub_mesh.n_cells
 
 
 def test_subtract():
+    sphere = SPHERE.copy()
+    sphere_shifted = SPHERE_SHIFTED.copy()
     sub_mesh = sphere - sphere_shifted
     assert sub_mesh.n_points
     assert sub_mesh.n_cells
 
 
 def test_add():
+    sphere = SPHERE.copy()
+    sphere_shifted = SPHERE_SHIFTED.copy()
     add_mesh = sphere + sphere_shifted
 
     npoints = sphere.n_points + sphere_shifted.n_points
@@ -187,6 +201,8 @@ def test_add():
 
 
[…]
