### 2019-04-25 5da4c1b
Fix Travis PyPI deployment issue
