+++ .travis.yml
@@ -50,6 +50,8 @@ matrix:
       python: 3.7
       env:
         - REPORT_COVERAGE: true
+        # Deploy to PyPI when core testing passes on 3.7 on tags
+        - VTKI_DEPLOY: true
       # Run the test suite and generate coverage report
       script: pytest -v --cov vtki
 
@@ -100,8 +102,8 @@ matrix:
         pytest -v --nbval-lax --current-env --disable-warnings tests/*.ipynb;
         pytest -v --nbval-lax --current-env --disable-warnings notebooks/*.ipynb;
 
-    # Build/test documentation on 3.7 only
-    - name: Documentation and Deploy 3.7
+    # Build/test/deploy documentation on 3.7 only
+    - name: Documentation Build/Deploy 3.7
       python: 3.7
       env:
         - REPORT_COVERAGE: false
@@ -152,7 +154,7 @@ deploy:
     user: akaszynski
     distributions: sdist
     on:
-      condition: "$TRAVIS_PYTHON_VERSION == 3.7"
+      condition: ($TRAVIS_PYTHON_VERSION == 3.7) && ($VTKI_DEPLOY == 'true')
       tags: true
     password: $PYPI_PASSWORD
 
