### 2019-04-26 a329f7d
Use link check in docs
