+++ docs/conf.py
@@ -33,7 +33,12 @@ extensions = ['sphinx.ext.autodoc',
               'notfound.extension',
               'sphinx_copybutton',
               'sphinx_gallery.gen_gallery',
+              'sphinx.ext.extlinks',
              ]
+
+
+linkcheck_retries = 3
+linkcheck_timeout = 500
 
 # Add any paths that contain templates here, relative to this directory.
 templates_path = ['_templates']
