### 2019-04-26 e41f2e2
Implement streamlines filter
