### 2019-04-28 821aaa3
Update documentation for streamlines filter
