+++ vtki/filters.py
@@ -978,22 +978,113 @@ class DataSetFilters(object):
 
 
     def streamlines(dataset, vectors=None, source_center=None,
-                    source_radius=None, integration_direction='both',
-                    integrator_type=45, n_points=100,
+                    source_radius=None, n_points=100,
+                    integrator_type=45, integration_direction='both',
                     surface_streamlines=False, initial_step_length=0.5,
                     step_unit='cl', min_step_length=0.01, max_step_length=1.0,
-                    max_steps=2000, max_streamline_length=None,
-                    terminal_speed=1e-12, max_error=1e-6, max_time=None,
-                    compute_vorticity=True, interpolator_type='point',
-                    rotation_scale=1.0, start_position=(0.0, 0.0, 0.0),
+                    max_steps=2000, terminal_speed=1e-12, max_error=1e-6,
+                    max_time=None, compute_vorticity=True, rotation_scale=1.0,
+                    interpolator_type='point', start_position=(0.0, 0.0, 0.0),
                     return_source=False):
         """Integrate a vector field to generate streamlines. The integration is
         performed using a specified integrator, by default Runge-Kutta2.
+        This supports integration through any type of dataset.
+        Thus if the dataset contains 2D cells like polygons or triangles, the
+        integration is constrained to lie on the surface defined by 2D cells.
+
+        This produces polylines as the output, with each cell
+        (i.e., polyline) representing a streamline. The attribute values
+        associated with each streamline are stored in the cell data, whereas
+        those associated with streamline-points are stored in the point data.
 
         This uses a Sphere as the source - set it's location and radius via
         the ``source_center`` and ``source_radius`` keyword arguments.
         You can retrieve the source as :class:`vtki.PolyData` by specifying
         ``return_source=True``.
+
+        Parameters
+        ----------
+        vectors : str
+            The string name of the active vector field to integrate across
+
+        source_center : tuple(float)
+            Length 3 tuple of floats defining the center of the source
+            particles. Defaults to the center of the dataset
+
+        source_radius : float
+            Float radius of the source particle cloud. Defaults to one-tenth of
+            the diagonal of the dataset's spatial extent
+
+        n_points : int
+            Number of particles present in source sphere
+
+        integrator_type : int
+            The integrator type to be used for streamline generation.
+            The default is Runge-Kutta45. The recognized solvers are:
+            RUNGE_KUTTA2 (``2``),  RUNGE_KUTTA4 (``4``), and RUNGE_KUTTA45
+            (``45``). Options are ``2``, ``4``, or ``45``. Default is ``45``.
+
+        integration_direction : str
+            Specify whether the streamline is integrated in the upstream or
+            downstream directions (or both). Options are ``'both'``,
+            ``'backward'``, or ``'forward'``.
+
+        surface_streamlines : bool
+            Compute streamlines on a surface. Default ``False``
+
+        initial_step_length : float
+            Initial step size used for line integration, expressed ib length
+            unitsL or cell length units (see ``step_unit`` parameter).
+            either the starting size for an adaptive integrator, e.g., RK45, or
+            the constant / fixed size for non-adaptive ones, i.e., RK2 and RK4)
+
+        step_unit : str
+            Uniform integration step unit. The valid unit is now limited to
+            only LENGTH_UNIT (``'l'``) and CELL_LENGTH_UNIT (``'cl'``).
+            Default is CELL_LENGTH_UNIT: ``'cl'``.
+
+        min_step_length : float
+            Minimum step size used for line integration, expressed in length or
+            cell length units. Only valid for an adaptive integrator, e.g., RK45
+
+        max_step_length : float
[…]
