### 2019-04-28 d5efac6
Fix streamline example
