### 2019-04-28 47b3b26
Add decimate example
