### 2019-04-28 3627ad2
Update decimate example
