+++ examples/01-filter/decimate.py
@@ -5,7 +5,7 @@ Decimation
 Decaimate a mesh
 
 """
-# sphinx_gallery_thumbnail_number = 1
+# sphinx_gallery_thumbnail_number = 4
 import vtki
 from vtki import examples
 
@@ -13,9 +13,10 @@ mesh = examples.download_face()
 
 # Define a camera potion the shows this mesh properly
 cpos = [(0.4,-0.07,-0.31), (0.05,-0.13,-0.06), (-.1,1,0.08)]
+dargs = dict(cpos=cpos, show_edges=True, color=True)
 
 # Preview the mesh
-mesh.plot(cpos=cpos, show_edges=True, color=True)
+mesh.plot(**dargs)
 
 ###############################################################################
 #  Now let's define a target reduction and compare the
@@ -26,10 +27,31 @@ print('Reducing {} percent out of the original mesh'.format(target_reduction * 1
 ###############################################################################
 decimated = mesh.decimate(target_reduction)
 
-decimated.plot(cpos=cpos, show_edges=True)
+decimated.plot(**dargs)
 
 
 ###############################################################################
 pro_decimated = mesh.decimate_pro(target_reduction, preserve_topology=True)
 
-pro_decimated.plot(cpos=cpos, show_edges=True, color=True)
+pro_decimated.plot(**dargs)
+
+
+###############################################################################
+# Side by side comparison:
+
+p = vtki.Plotter(shape=(1, 3))
+p.add_mesh(mesh, **dargs)
+p.add_text('Input mesh', font_size=24)
+p.camera_position = cpos
+p.reset_camera()
+p.subplot(0,1)
+p.add_mesh(decimated, **dargs)
+p.add_text('Decimated mesh', font_size=24)
+p.camera_position = cpos
+p.reset_camera()
+p.subplot(0,2)
+p.add_mesh(pro_decimated, **dargs)
+p.add_text('Pro Decimated mesh', font_size=24)
+p.camera_position = cpos
+p.reset_camera()
+p.show()
