### 2019-04-28 28bf055
Add select enclosed points filter
