### 2019-04-28 45ac5e8
streamlines filter (#199)
* Implement streamlines filter

* Update documentation for streamlines filter

* Fix streamline example
