+++ tests/test_filters.py
@@ -310,3 +310,13 @@ def test_resample():
     name = 'Spatial Point Data'
     assert name in result.scalar_names
     assert isinstance(result, type(mesh))
+
+
+def test_streamlines():
+    mesh = examples.download_carotid()
+    stream, src = mesh.streamlines(return_source=True, max_time=100.0,
+                            initial_step_length=2., terminal_speed=0.1,
+                           n_points=25, source_radius=2.0,
+                           source_center=(133.1, 116.3, 5.0) )
+    assert stream.n_points > 0
+    assert src.n_points == 25
