+++ vtki/common.py
@@ -653,6 +653,11 @@ class Common(DataSetFilters, object):
         return list(self.GetBounds())
 
     @property
+    def length(self):
+        """the length of the diagonal of the bounding box"""
+        return self.GetLength()
+
+    @property
     def center(self):
         """ Center of the bounding box """
         return list(self.GetCenter())
