### 2019-04-28 d0baf2c
Add interpolate filter
