+++ examples/01-filter/resample.py
@@ -1,6 +1,6 @@
 """
-Resampling Data Sets
-~~~~~~~~~~~~~~~~~~~~
+Resampling & Interpolating
+~~~~~~~~~~~~~~~~~~~~~~~~~~
 
 Resample one mesh's point/cell arrays onto another meshes nodes.
 """
@@ -8,7 +8,7 @@ Resample one mesh's point/cell arrays onto another meshes nodes.
 # This example will resample a volumetric mesh's  scalar data onto the surface
 # of a sphere contained in that volume.
 
-# sphinx_gallery_thumbnail_number = 1
+# sphinx_gallery_thumbnail_number = 4
 import vtki
 from vtki import examples
 import numpy as np
@@ -32,3 +32,32 @@ result = mesh.sample(data_to_probe)
 # Plot result
 name = 'Spatial Point Data'
 result.plot(scalars=name, clim=data_to_probe.get_data_range(name))
+
+
+
+###############################################################################
+# Interpolate
+# +++++++++++
+#
+# Run a resampling from an interpolation built using a Gaussian kernal
+
+# Download sample data
+surface = examples.download_saddle_surface()
+points = examples.download_sparse_points()
+
+
+p = vtki.Plotter()
+p.add_mesh(points, point_size=30.0, render_points_as_spheres=True)
+p.add_mesh(surface)
+p.show()
+
+###############################################################################
+# Run the interpolation
+
+interpolated = surface.interpolate(points, radius=12.0)
+
+
+p = vtki.Plotter()
+p.add_mesh(points, point_size=30.0, render_points_as_spheres=True)
+p.add_mesh(interpolated, scalars='val')
+p.show()
