### 2019-04-28 31dc882
Update glyphing example to use non-trivial data
