### 2019-04-28 6cc7c25
Update tri surface example
