### 2019-04-28 1dc01e7
Remove plot theme code in examples
