+++ examples/00-load/create-tri-surface.py
@@ -30,7 +30,6 @@ print(points[0:5, :])
 
 # simply pass the numpy points to the PolyData constructor
 cloud = vtki.PolyData(points)
-vtki.set_plot_theme('doc')
 cloud.plot(point_size=15)
 
 ################################################################################
