### 2019-04-29 9301abc
Update streamlines example
