### 2019-04-29 b3958a6
Remove notebook argument from docs examples
