### 2019-04-29 f7e2f20
Fix kitchen downloads test
