### 2019-04-29 ea3f93a
Fix streamlines example icon
