### 2019-04-29 8f52ddb
Fix SegY reader import issue
