+++ vtki/grid.py
@@ -174,7 +174,7 @@ class RectilinearGrid(vtkRectilinearGrid, Grid):
         filename = os.path.abspath(os.path.expanduser(filename))
         # check file exists
         if not os.path.isfile(filename):
-            raise Exception('%s does not exist')
+            raise Exception('{} does not exist'.format(filename))
 
         # Check file extention
         if '.vtr' in filename:
@@ -446,7 +446,7 @@ class UniformGrid(vtkImageData, Grid):
         filename = os.path.abspath(os.path.expanduser(filename))
         # check file exists
         if not os.path.isfile(filename):
-            raise Exception('%s does not exist')
+            raise Exception('{} does not exist'.format(filename))
 
         # Check file extention
         if '.vti' in filename:
