### 2019-04-30 367e5bb
Refactor scale_factor -> factor
