### 2019-04-30 03992d4
Attempt fix binder examples deploy
