### 2019-05-01 6097da9
Add support for string scalar arrays and beter point labelling
