### 2019-05-01 #1382 @leouieda
:wave: Hi @rgmyr @Chilipp would you be interested in reviewing this submission for JOSS?
