### 2019-05-01 acd2990
Add helper function to label scalar arrays in plotter
