+++ examples/02-plot/plot-labels.py
@@ -4,23 +4,27 @@ Label Points
 
 Use string arrays in a point set to label points
 """
+# sphinx_gallery_thumbnail_number = 3
 from vtki import examples
 import vtki
 import numpy as np
 
 ###############################################################################
-# Label Array
-# +++++++++++
+# Label String Array
+# ++++++++++++++++++
+#
+# This example will label the nodes of a mesh with a given array of string
+# labels for each of the nodes.
 
+# Make some random points
 poly = vtki.PolyData(np.random.rand(10, 3))
-
-print(poly)
-
 
 ###############################################################################
 # Add string labels to the point data - this associates a label with every node:
 
 poly['My Labels'] = ['Label {}'.format(i) for i in range(poly.n_points)]
+
+print(poly)
 
 ###############################################################################
 # Now plot the points with labels:
@@ -33,10 +37,14 @@ plotter.show()
 ###############################################################################
 # Label Node Locations
 # ++++++++++++++++++++
+#
+# This example will label the nodes of a mesh with their coordinate locations
 
 # Load example beam file
 grid = vtki.UnstructuredGrid(examples.hexbeamfile)
 
+
+###############################################################################
 # Create plotting class and add the unstructured grid
 plotter = vtki.Plotter()
 plotter.add_mesh(grid, show_edges=True, color='tan')
@@ -44,11 +52,38 @@ plotter.add_mesh(grid, show_edges=True, color='tan')
 # Add labels to points on the yz plane (where x == 0)
 points = grid.points
 mask = points[:, 0] == 0
-plotter.add_point_labels(points[mask], points[mask].tolist())
+plotter.add_point_labels(points[mask], points[mask].tolist(),
+                         point_size=20, font_size=36)
 
 plotter.camera_position = [
-                (-1.4643015810492384, 1.5603923627830638, 3.16318236536270),
-                (0.05268120500967251, 0.639442034364944, 1.204095304165153),
-                (0.2364061044392675, 0.9369426029156169, -0.25739213784721)]
+    (-1.5, 1.5, 3.),
+    (0.05, 0.6, 1.2),
+    (0.2, 0.9, -0.25)]
 
 plotter.show()
+
+
+
+###############################################################################
+# Label Scalar Values
+# +++++++++++++++++++
+#
+# This example will label pooints with their scalar values
+
+mesh = examples.load_uniform().slice()
+
+###############################################################################
+p = vtki.Plotter()
+
+# Add the mesh:
+p.add_mesh(mesh, scalars='Spatial Point Data', show_edges=True)
+# Add the points with scalar labels:
+p.add_point_scalar_labels(mesh, 'Spatial Point Data',
+                          point_size=20, font_size=36)
+
+# Use a nice camera position:
+p.camera_position = [(7, 4, 5),
+                     (4.4, 7., 7.2),
+                     (0.8, 0.5, 0.25)]
+
+p.show()
