### 2019-05-01 7fd7b33
Add support for string scalar arrays and beter point labelling (#200)
* Add support for string scalar arrays and beter point labelling

* Add helper function to label scalar arrays in plotter

* Fix bad error message
