+++ docs/conf.py
@@ -51,7 +51,7 @@ master_doc = 'index'
 
 # General information about the project.
 project = u'vtki'
-copyright = u'2017-2019, vtkiorg Developers'
+copyright = u'2017-2019, The vtkiorg Developers'
 author = u'Alex Kaszynski and Bane Sullivan'
 
 # The version info for the project you're documenting, acts as replacement for
