+++ docs/tools/utilities.rst
@@ -1,10 +1,29 @@
 Utilities
 =========
 
-
 .. automodule:: vtki.utilities
 
+.. autofunction:: vtki.set_error_output_file
+
+.. autofunction:: vtki.trans_from_matrix
+
+.. autofunction:: vtki.is_inside_bounds
+
+
+Object Conversions
+------------------
+
 .. autofunction:: vtki.wrap
+
+.. autofunction:: vtki.is_vtki_obj
+
+.. autofunction:: vtki.image_to_texture
+
+.. autofunction:: vtki.numpy_to_texture
+
+
+File IO
+-------
 
 .. autofunction:: vtki.read
 
@@ -14,6 +33,10 @@ Utilities
 
 .. autofunction:: vtki.read_legacy
 
+
+Mesh Creation
+-------------
+
 .. autofunction:: vtki.lines_from_points
 
 .. autofunction:: vtki.vtk_points
@@ -22,6 +45,20 @@ Utilities
 
 .. autofunction:: vtki.fit_plane_to_points
 
-.. autofunction:: vtki.trans_from_matrix
 
-.. autofunction:: vtki.set_error_output_file
+Array Access
+------------
+
+.. autofunction:: vtki.get_scalar
+
+.. autofunction:: vtki.convert_array
+
+.. autofunction:: vtki.point_scalar
+
+.. autofunction:: vtki.cell_scalar
+
+.. autofunction:: vtki.get_vtk_type
+
+.. autofunction:: vtki.vtk_bit_array_to_char
+
+.. autofunction:: vtki.convert_string_array
