+++ AUTHORS.rst
@@ -1,6 +1,7 @@
-The following are a list of authors who have made substantial contributions to
+The following is a list of authors who have made substantial contributions to
 the conception or design of this software; or the creation of new code used in
-this software; or have drafted the work or substantively revised it:
+this software; or have drafted the work or substantively revised it and are
+considered "The vtkiorg Developers":
 
 - Alex Kaszynski, (`@akaszynski <https://github.com/akaszynski/>`_)
 - Bane Sullivan, (`@banesullivan <https://github.com/banesullivan/>`_)
