+++ docs/external_examples.rst
@@ -9,18 +9,18 @@ Here are a list of longer, more technical examples of what ``vtki`` can do!
     If any of these links are broken, please raise an issue on the repository.
 
 
+Do you have a technical processing workflow or visualization routine you
+would like to share?
+If so, please consider sharing your work here by clicking the `Edit on GitHub`
+button at the top of this page and adding a link to your project!
+
+
 .. toctree::
    :maxdepth: 2
 
    3D visualization for the Open Mining Format (omf) <https://github.com/OpenGeoVis/omfvtk/blob/master/Example.ipynb>
+   Live example notebook showcasing omfvtk <https://mybinder.org/v2/gh/OpenGeoVis/omfvtk-demo/master?filepath=notebooks%2FExample.ipynb>
    3D Rendering with Discretize <http://discretize.simpeg.xyz/en/master/auto_examples/vtki_demo.html#sphx-glr-auto-examples-vtki-demo-py>
    OpenFOAM Rendering <https://pswpswpsw.github.io/posts/2018/09/blog-post-modify-vtk-openfoam/>
    FORGE Geothermal Project <http://forge.pvgeo.org>
-
-
-.. attention::
-
-   Do you have a technical processing workflow or visualization routine you
-   would like to share?
-   If so, please consider sharing your code here in the ``vtki`` documentation
-   by linking your example to this page!
+   PVGeo's example gallery <https://pvgeo.org/examples/index.html>
