### 2019-05-03 1481696
Add multiblock copy method and negative indexing
