### 2019-05-03 3aeb812
Enable default plotting theme environment variable
