### 2019-05-03 ab8a1db
Add edge_color to rcParams
