### 2019-05-03 f4919e2
added extract cells from point indices
