### 2019-05-03 31612dd
removed unneeded structured grid docs
