### 2019-05-03 8d1995b
Add logo per #201
