### 2019-05-03 3ac11db
added field arrays, in addition to the current point and cell ones
