### 2019-05-03 973fd46
made sure field data cannot be made active
