### 2019-05-03 4286352
Update dev testing guide (note #206)
