### 2019-05-03 15cb426
added fields feature referenced in #206
