+++ vtki/common.py
@@ -11,7 +11,7 @@ from vtk.util.numpy_support import numpy_to_vtk, vtk_to_numpy
 
 import vtki
 from vtki import DataSetFilters
-from vtki.utilities import (CELL_DATA_FIELD, POINT_DATA_FIELD, get_scalar,
+from vtki.utilities import (CELL_DATA_FIELD, POINT_DATA_FIELD, FIELD_DATA_FIELD, get_scalar,
                             vtk_bit_array_to_char, is_vtki_obj,
                             _raise_not_matching, convert_array)
 
@@ -37,6 +37,7 @@ class Common(DataSetFilters, object):
         self.references = []
         self._point_bool_array_names = []
         self._cell_bool_array_names = []
+        self._field_bool_array_names = []
 
     @property
     def active_scalar_info(self):
@@ -48,7 +49,11 @@ class Common(DataSetFilters, object):
         # rare error where scalar name isn't a valid scalar
         if name not in self.point_arrays:
             if name not in self.cell_arrays:
-                name = None
+                if name in self.field_arrays:
+                    raise RuntimeError('Field arrays cannot be made active. ' +
+                                       'Convert to point/cell arrays if possible.')
+                else:
+                    name = None
 
         if name is None:
             if self.n_scalars < 1:
@@ -74,7 +79,11 @@ class Common(DataSetFilters, object):
         # rare error where scalar name isn't a valid scalar
         if name not in self.point_arrays:
             if name not in self.cell_arrays:
-                name = None
+                if name in self.field_arrays:
+                    raise RuntimeError('Field arrays cannot be made active. ' +
+                                       'Convert to point/cell array if possible.')
+                else:
+                    name = None
 
         return self._active_vectors_info
 
@@ -277,6 +286,8 @@ class Common(DataSetFilters, object):
             self.point_arrays[new_name] = self.point_arrays.pop(old_name)
         elif field == CELL_DATA_FIELD:
             self.cell_arrays[new_name] = self.cell_arrays.pop(old_name)
+        elif field == FIELD_DATA_FIELD:
+            self.field_arrays[new_name] = self.field_arrays.pop(old_name)
         else:
             raise RuntimeError('Array not found.')
         if self.active_scalar_info[1] == old_name:
@@ -379,6 +390,79 @@ class Common(DataSetFilters, object):
         """ Makes points double precision """
         if self.points.dtype != np.double:
             self.points = self.points.astype(np.double)
+
+    def _field_scalar(self, name=None):
+        """
+        Returns field scalars of a vtk object
+
+        Parameters
+        ----------
+        name : str
+            Name of field scalars to retrive.
+
+        Returns
+        -------
+        scalars : np.ndarray
+            Numpy array of scalars
+
+        """
+        if name is None:
+            raise RuntimeError('Must specify an array to fetch.')
+        vtkarr = self.GetFieldData().GetAbstractArray(name)
+        if vtkarr is None:
+            raise AssertionError('({}) is not a valid field scalar array'.format(name))
+
+        # numpy does not support bit array data types
+        if isinstance(vtkarr, vtk.vtkBitArray):
+            vtkarr = vtk_bit_array_to_char(vtkarr)
+            if name not in self._point_bool_array_names:
+                self._field_bool_array_names.append(name)
+
+        array = convert_array(vtkarr)
+        if array.dtype == np.uint8 and name in self._field_bool_array_names:
+            array = array.view(np.bool)
+        return array
+
+    def _add_field_scalar(self, scalars, name, deep=True):
+        """
+        Adds field scalars to the mesh
+
+        Parameters
+        ----------
+        scalars : numpy.ndarray
[…]
