### 2019-05-06 pyvista/pyvista#208 @banesullivan
Renaming vtki → PyVista
@akaszynski and I will be renaming `vtki` to avoid confusion between this library and other similarly named projects as well as to reflect that `vtki` is not a direct product of VTK or Kitware and to ensure that were are **absolutely not using VTK's name to endorse or promote a product derived from VTK.**

Renaming the software will require quite a lot of refactoring both internally and across dependent projects - even so, we feel renaming `vtki` is necessary.

## Call for Help

**Users of `vtki`:** we would like to solicit your help in making a new name for this software and we welcome any and all suggestions that are compliant with most open-source software licensing terms.

Please use this issue as a place to post your thoughts on what this software shall be named and be sure to give a +1 (👍) to any suggestions that you like!
