### 2019-05-06 vtkiorg/vtki#208 @aashish24
how about ipyvis?
