### 2019-05-06 vtkiorg/vtki#208 @banesullivan
Thanks for the suggestion @aashish24!

I'm concerned it might be confused with https://github.com/WestHealth/pyvis
