### 2019-05-06 vtkiorg/vtki#208 @banesullivan
Thanks, @mmagnuski! That's a really cool visualization - did you make that leveraging `vtki`?!
