### 2019-05-06 vtkiorg/vtki#208 @akaszynski
This is of course right after we made the logo!

It used to be ``vtk_interface``, which is where ``vtki`` came from.  Ideally the module would be short and sweet, but still an informative pep8 name.  I'm thinking:
- ``visplot``
- ``easyvis``
- ``meshplot`` or ``gridplot`` or ``simplot`` (as in simple plotting)
- ``not_vtk`` (just kidding, it's not pep8)

``vtkplotter`` might have to change their name as well...
