### 2019-05-06 vtkiorg/vtki#208 @thast
how about simply `visualize`? (that is probably my `simpeg/discretize` bias ^^ ) and the repo name seems to not have been taken yet.
