### 2019-05-06 vtkiorg/vtki#208 @tomtomatron
How about `Envisage` or `nviz` for short. Maybe `nvz`.
