### 2019-05-06 vtkiorg/vtki#208 @akaszynski
``visualize`` sounds good.  I'm amazed the module name isn't taken.

I'll probably just end up importing as:
```python
import visualize as vis
```

It's a straightforward module name.  Thanks @thast

It's also great to see the support.  Didn't know how many people were active here on the repo.
