### 2019-05-06 vtkiorg/vtki#208 @banesullivan
I'm shocked `visualize` and `visualizer` aren't taken on PyPI yet... we may want to jump on those
