### 2019-05-06 vtkiorg/vtki#208 @craigmillernz
Something that ties into OpenGeoVis, which is the parent project might be an idea?
