### 2019-05-06 vtkiorg/vtki#208 @craigmillernz
I like visualize, it's simple, generic easy to remember and shorten when importing
