### 2019-05-06 vtkiorg/vtki#208 @banesullivan
My current favorites (in no particular order) are:

- `visualize`
- `simplot`
- `Envisage`
