### 2019-05-06 vtkiorg/vtki#208 @banesullivan
It's worth noting that we'll also have to rename the organizational account @vtkiorg - likely to the same name as the main software library or something different if we land on `visualize` since that GitHub account is taken

Perhaps something in line with @OpenGeoVis like @craigmillernz mentioned?

- I love "vispy" and "pyviz" but these are taken 😞
- "openvis"/"openviz" - again these are taken on GitHub but something similar would be nice
