### 2019-05-06 vtkiorg/vtki#208 @leouieda
I like `visualize` as well. Is there anything else that could be more unique? There are already so many `*vis*` packages out there. I feel like vtki has a very direct mission (a better VTK interface) that calling it `visualize` might be too generic.
