### 2019-05-06 vtkiorg/vtki#208 @banesullivan
Thanks, @leouieda! This is a great point - `vtki`'s mission has been to make VTK's data types, filtering algorithms, and rendering more accessible to new users and researchers looking for an easy-to-use mesh analysis and visualization toolset.

Considering most of the recent development has been to make analysis and filtering algorithms for all the datatypes - a name like `visualize` might not capture scope of what this software implements.

For example, we have all types of thresholding, slicing, mesh subsetting, etc. algorithms available
