### 2019-05-06 vtkiorg/vtki#208 @banesullivan
Another good point, @aashish24 - "simplot" does imply simulation in my mind so that likely isn't an appropriate name

`easyviz`/`easyvis` is growing on me - both GitHub org accounts (I just snagged them) and PyPI package names are available
