### 2019-05-06 vtkiorg/vtki#208 @leouieda
How about something that plays on the concept of "improving a toolkit": knife sharpening, lubricating (WD-40), upgrading (hammer -> nail gun). I tend to like quirky names but there is benefit to being easily googled.
