### 2019-05-06 vtkiorg/vtki#208 @banesullivan
### Org Account

I suppose we avoided putting `vtki` under @OpenGeoVis from the start because `vtki`'s applications are more general than geo-visualization and we wanted to attract users from all disciplines.

Considering that almost every project under @OpenGeoVis is built on top of `vtki` and we seem to be gaining traction mostly in the geoscience realm at this time, this might be appropriate. However, I want to make sure this software gains visibility from users across disciplines as it is used as a supporting module in more non-geo software packages at this time - see dependents graph for details.

@akaszynski, thoughts?

### Software Name

@leouieda: I very much like the "improving a toolkit" concept - I'll have to brainstorm this in regards to how `vtki` "sharpens"/improves VTK - do you have any projects in mind that do this kind of software improvement/wrapping?
