### 2019-05-06 vtkiorg/vtki#208 @leouieda
@banesullivan I imagined that was the case for opengeovis. Off the top of my head:

* [requests](https://2.python-requests.org/en/master/) wraps `urllib3` (somewhat) and is used by pretty much everyone doing we stuff. Tag line: HTTP for humans.
* [gitless](https://gitless.com/) is a redesigned version of git built on top of git
