### 2019-05-06 vtkiorg/vtki#208 @banesullivan
what about `meshy`?

```py
import meshy

mesh = meshy.PolyData()

mesh.plot()
```
