### 2019-05-06 vtkiorg/vtki#208 @mmagnuski
A few ideas that aim to emphasise 3d:
* `thirdeye`
* `depth` (or `depthviz`)
* `trivis` (threevis is taken) or `freevi`

And another to retain the `vtk` + `i` identity:
`vistki` (`import vistki as vtki` ;P) which would translate to
"**vis**ualize **t**hreedimensional (**k**)onstructs **i**nteractively" to avoid any copyright issues :D
