### 2019-05-06 vtkiorg/vtki#208 @leouieda
On the 3D theme, how about "stereoscope" or "parallax"?
