### 2019-05-06 vtkiorg/vtki#208 @craigmillernz
>
>
> what about `meshy`?
>
> ```python
> import meshy
> from meshy import examples
>
> mesh = meshy.PolyData()
> # ... or ....
> mesh = examples.download_st_helens().warp_by_scalar()
>
> p = meshy.Plotter()
> p.add_mesh(mesh)
> p.show()
> ```
>
> inspired by @AntonBiryukovUofC - "Boaty McBoatface" → "Meshy McMeshface"

meshy implies its about creating meshes not visualizing. plus too much overlap with discretize which has a Mesh
