### 2019-05-07 vtkiorg/vtki#208 @aashish24
parallax would be misleading - it is used for stereo, meshy will sounds more like a mesh library. Also, I would think you want to make it general 2D and 3D library as opposed to just 3D.
