### 2019-05-07 vtkiorg/vtki#208 @tomtomatron
If 'Envisage' doesn't sit well, maybe `nviz` for `N-dimensional visualization`, or `Envison`.
