### 2019-05-07 vtkiorg/vtki#208 @jmp75
`nviz` is nice and short, but may also be misleading as `vtki` is not for any N-dimensional data. People searching for this may seek more something like [ggobi](http://www.ggobi.org).
