### 2019-05-07 cd28fae
Fix get_scalar bug
