### 2019-05-07 4794552
Update data repr methods
