+++ vtki/common.py
@@ -959,17 +959,24 @@ class Common(DataSetFilters, object):
             fmt += "</td><td>"
             fmt += "\n"
             fmt += "<table>\n"
-            row = "<tr><th>{}</th><th>{}</th><th>{}</th><th>{}</th><th>{}</th></tr>\n"
-            fmt += row.format("Name", "Field", "Type", "Min", "Max")
-            row = "<tr><td>{}</td><td>{}</td><td>{}</td><td>{:.3e}</td><td>{:.3e}</td></tr>\n"
+            titles = ["Name", "Field", "Type", "N Comp", "Min", "Max"]
+            fmt += "<tr>" + "".join(["<th>{}</th>".format(t) for t in titles]) + "</tr>\n"
+            row = "<tr><td>{}</td><td>{}</td><td>{}</td><td>{}</td><td>{}</td><td>{}</td></tr>\n"
+            row = "<tr>" + "".join(["<td>{}</td>" for i in range(len(titles))]) + "</tr>\n"
 
             def format_array(key, field):
                 """internal helper to foramt array information for printing"""
-                arr = get_scalar(self, key)
+                arr = get_scalar(self, key, preference=field)
                 dl, dh = self.get_data_range(key)
+                dl = '{:.3e}'.format(dl)
+                dh = '{:.3e}'.format(dh)
                 if key == self.active_scalar_info[1]:
                     key = '<b>{}</b>'.format(key)
-                return row.format(key, field, arr.dtype, dl, dh)
+                if arr.ndim > 1:
+                    ncomp = arr.shape[1]
+                else:
+                    ncomp = 1
+                return row.format(key, field, arr.dtype, ncomp, dl, dh)
 
             for i in range(self.GetPointData().GetNumberOfArrays()):
                 key = self.GetPointData().GetArrayName(i)
