### 2019-05-07 9117a59
Install panel from master branch on Travis
