### 2019-05-07 vtkiorg/vtki#208 @banesullivan
On the comments about `vtki`'s focus - `vtki` is built to be both an underlying mesh creation and analysis utility as well as a 3D visualization tool for those mesh types. We want to emphasize that `vtki` is **not just 3D visualization** and it can be used as a supporting module for the underlying types in software that use discretized finite volume and element meshes with simple geometries using voxels/hexahedrons (think `discretize`) or complex geometries using tetrahedrons, etc. (think applications like [PyGIMLi](https://www.pygimli.org)).

## Future of `vtki`

**My (somewhat overzealous) goal for `vtki`:** to become the underlying library of choice for Python-based finite volume/element analysis, processing, and simulation software. I envision`vtki`'s mesh types being used as the underlying data structures in finite mesh problems across a wide range of software libraries. This would bring interoperability amongst software to a whole new level - think of the possibilities if people from all disciplines collaborated on a single toolset for finite mesh problems. The caveat is that we have quite a bit of development to go before we could get to that point.

### Use Case

If you are developing a finite volume simulation software and dealing with 3D data, then the underlying mesh types need to be able to be visualized in 3D - rather than create your own mesh types which are likely equivalent to underlying structures in VTK, simply define your operators to work with `vtki` mesh types and you'll gain the all the additional features of `vtki`'s mesh analysis and filtering... The 3D plotting is a required feature after that which is why we have so much code dedicated to making the process of visualizing complex `vtki` datasets intuitive/easy - having powerful 3D viz features is how I hope to get folks using `vtki` and onboard with trying to use `vtki`'s types for their project.

### 2D vs. 3D

This is likely better suited for its own issue, but I think our focus with `vtki`'s viz efforts should be for 3D visualization. `matplotlib` is an excellent tool for 2D viz and if our users want 2D representations of their 3D `vtki` data, then I think it would be best for us to create an interface for visualizing `vtki` datasets with `matplotlib`... especially since `matplotlib` is in a far better state in regards to creating publication-quality images (reference #147 and #165)
