### 2019-05-07 6235f57
Fix scalar bar slot issue in multi plots
