### 2019-05-07 vtkiorg/vtki#208 @banesullivan
how about **`meshviz`**

```py
import meshviz

mesh = meshviz.UniformGrid()
...
p = meshviz.Plotter()
...
```
