### 2019-05-07 vtkiorg/vtki#208 @banesullivan
Another (lazy) option is **`ivtk`** as in the "**I**nterface to the **V**isualization **T**ool**k**it" which avoids copyright issues (to the best of my knowledge) and adheres to our mission of building "A Streamlined Python Interface to the Visualization Toolkit"
