### 2019-05-07 vtkiorg/vtki#208 @leouieda
So Kitware would be fine with `ivtk` but not `vtki`?
