### 2019-05-07 vtkiorg/vtki#208 @banesullivan
We should do something different... if using `vtk` in our current name is an issue then we should avoid it altogether to make sure everyone is happy and adhering to copyright laws
