### 2019-05-07 vtkiorg/vtki#208 @banesullivan
Also, `mvis` which is a short, 4-letter verison of `meshvis`

```py
import mvis

mesh = mvis.UniformGrid()
...
p = mvis.Plotter()
...
```
