### 2019-05-07 vtkiorg/vtki#208 @lheagy
One that popped into my head: `vista` (def: a pleasing view)
