### 2019-05-07 vtkiorg/vtki#208 @banesullivan
I like `vista`... plus it's available on PyPI
