### 2019-05-07 vtkiorg/vtki#208 @banesullivan
Let's see how other's react but I'm liking `vista` (even though it implies visualization only which I have some hesitation about).

This would also be easy for other tools that expand on `vtki`/`vista` to easily tag our name to the end and new users know to associate that new software with this software: e.g. `***vista`.

With the name change to `vtki`, I will also be changing the names of offshoot projects for consistency like [`omfvtk`](https://github.com/OpenGeoVis/omfvtk) → `omfvista`.
