### 2019-05-07 vtkiorg/vtki#208 @leouieda
`vista` is my favorite so far. It sounds unique and pleasant. Might be tough to get a good domain name, though: https://www.fatiando.org/verde/
