### 2019-05-07 vtkiorg/vtki#208 @dannyseim
anthro-py-scene
