### 2019-05-07 vtkiorg/vtki#208 @banesullivan
I'm leaning towards `vistas` which has a bit more availability than `vista` when it comes to GitHub organization accounts, domains, and non-similarity with things like https://github.com/geovista, and `VistaPro`: https://en.wikipedia.org/wiki/VistaPro
