### 2019-05-07 vtkiorg/vtki#208 @banesullivan
I propose:

- rename `vtki` to `vistas`
- rename @vtkiorg to @pyvistas
- forward https://vtki.org to https://pyvistas.org

I already acquired those accounts/domains
