### 2019-05-07 vtkiorg/vtki#208 @banesullivan
Actually `vista` has the same availability:

- rename `vtki` to `vista`
- rename @vtkiorg to @pyvista
- forward https://vtki.org to https://pyvista.org

I already acquired those accounts/domains
