### 2019-05-07 vtkiorg/vtki#208 @craigmillernz
i'm confused between the distinction of vista vs pyvista ?
