### 2019-05-07 vtkiorg/vtki#208 @banesullivan
domain names for `vista` are through the roof expensive... `pyvista` is way cheaper and more in line with what the organizational account is doing: "Creating Python visualization software leveraging the Visualization Toolkit (VTK)"

We could also call the software itself `pyvista` I just shortened because:

```py
import vista
```

is easier/ cleaner than

```py
import pyvista
```
but I'm totally open to either
