### 2019-05-07 vtkiorg/vtki#208 @lheagy
Pytorch makes the same distinction - the website is: https://pytorch.org/, but the import is
```
import torch
```

:+1: from me on the shorter import `vista` plus the longer / more searchable website domain `pyvista`
