### 2019-05-07 vtkiorg/vtki#208 @banesullivan
I'm sold. @akaszynski, any reservations?
