### 2019-05-07 vtkiorg/vtki#208 @craigmillernz
but torch/pytorch have
conda install pytorch torchvision cudatoolkit=9.0 -c pytorch

so is how is it labelled on pip/conda?
