### 2019-05-07 vtkiorg/vtki#208 @banesullivan
I'm thinking the organizational aspect of this project should be referred to as "pyvista" and the software itself be "vista" so all pip/conda installs would be `pip install vista`.. `pyvitsa`/`pytorch` seems redundant to me when working with a Python package manager...
