### 2019-05-07 vtkiorg/vtki#208 @craigmillernz
what about all the external facing names (website, pip, conda etc) should refer to pyvista, but within the code it gets shortened to vista?  Just thinking it is less confusing for new people to find and to make it clear that pyvista isnt just a wrapper on some other thing called vista, but is the actually code in its own right.  I agree the py is a bit redundant but i'm just aiming for clarity.
