### 2019-05-07 vtkiorg/vtki#208 @banesullivan
> Just thinking it is less confusing for new people to find and to make it clear that pyvista isnt just a wrapper on some other thing called vista, but is the actually code in its own right

That's an excellent point... "pyvista" might imply its some sort of python wrapping for another "vista" library. However, I think what I am proposing here is in line with pyviz: http://pyviz.org
