### 2019-05-07 vtkiorg/vtki#208 @leouieda
I am very much against naming the pip/conda packages one thing and importing another. This is terrible for users and causes a lot of confusion. It's fine if the website and github organization have different names because people usually don't have to know/remember those. The repository, package, and library names should be consistent, though.
