### 2019-05-07 vtkiorg/vtki#208 @akaszynski
Agreed with @leouied.  I like ``vista``, and a domain name of ``pyvista`` works, but the repo, package, and library should all be ``vista``.
