### 2019-05-07 vtkiorg/vtki#208 @banesullivan
It appears that we are all converging on:

```bash
pip install vista

conda install -c conda-forge vista
```

```py
import vista
```

Docs: https://docs.pyvista.org

Then related projects under the @pyvista organization can have subdomains under http://**.pyvista.org

And the main website http://pyvista.org will have a high level overview of everything we are building in this space
