### 2019-05-07 vtkiorg/vtki#208 @banesullivan
Possibly... I might end up migrating @OpenGeoVis under @pyvistas as everything in that org is pretty much built on top of `vista`.

I also snagged the @geopyvista account just in case we want to separate the geoscience applications
