### 2019-05-08 vtkiorg/vtki#208 @banesullivan
We’ve landed!! 🚀

I’ll begin the internal refactoring process in a PR then start switching ownership over to @pyvistas

Should we stage a release of `vtki` that will throw a runtime error on import telling users to install vistas instead?

Is there an easy way to link `pip install vtki` to `pip install vistas`? And same for conda?
