### 2019-05-08 vtkiorg/vtki#208 @akaszynski
Agreed.  Last release can include the runtime error.  We can remove it off of pypi after a few months have passed.

Goodbye old friend `vtki`, long live `vista`!
