### 2019-05-08 vtkiorg/vtki#208 @leouieda
@banesullivan, you can make `vista` a dependency of `vtki`.
