+++ examples/README.txt
@@ -8,5 +8,5 @@ Here is a gallery of all the quick examples demonstrating what ``vtki`` can do!
 All of these examples are live and available on MyBinder!
 
 .. image:: https://mybinder.org/badge_logo.svg
-   :target: https://mybinder.org/v2/gh/vtkiorg/vtki-examples/master
+   :target: https://mybinder.org/v2/gh/pyvista/vtki-examples/master
    :alt: Launch on Binder
