+++ appveyor.yml
@@ -26,7 +26,7 @@ init:
 
 install:
   # Use ci-helpers
-  - "git clone --depth 1 git://github.com/vtkiorg/gl-ci-helpers.git"
+  - "git clone --depth 1 git://github.com/pyvista/gl-ci-helpers.git"
   - "powershell gl-ci-helpers/appveyor/install_opengl.ps1"
   # Install requirements
   - "%PYTHON% -m ensurepip"
