+++ README.rst
@@ -53,16 +53,16 @@ papers as well as a supporting module for other mesh dependent Python modules.
 Documentation
 =============
 
-Refer to the `documentation <http://docs.vtki.org/>`_ for detailed
+Refer to the `documentation <http://docs.pyvista.org/>`_ for detailed
 installation and usage details.
 
 For general questions about the project, its applications, or about software
 usage, please do not create an issue but join us on Slack_ or send one
 of the developers an email. The project support team can be reached at
-`info@vtki.org`_.
+`info@pyvista.org`_.
 
 .. _Slack: http://slack.opengeovis.org
-.. _info@vtki.org: mailto:info@vtki.org
+.. _info@pyvista.org: mailto:info@pyvista.org
 
 
 Installation
@@ -77,7 +77,7 @@ You can also visit `PyPi <http://pypi.python.org/pypi/vtki>`_,
 `Anaconda <https://anaconda.org/conda-forge/vtki>`_, or
 `GitHub <https://github.com/pyvista/vtki>`_ to download the source.
 
-See the `Installation <http://docs.vtki.org/getting-started/installation.html#install-ref.>`_
+See the `Installation <http://docs.pyvista.org/getting-started/installation.html#install-ref.>`_
 for more details if the installation through pip doesn't work out.
 
 
@@ -87,7 +87,7 @@ Highlights
 Head over to the `Quick Examples`_ page in the docs to learn more about using
 ``vtki``.
 
-.. _Quick Examples: http://docs.vtki.org/examples/index.html
+.. _Quick Examples: http://docs.pyvista.org/examples/index.html
 
 Want to test-drive ``vtki``? Check out our live examples on MyBinder:
 
@@ -102,9 +102,9 @@ Want to test-drive ``vtki``? Check out our live examples on MyBinder:
 * Intuitive plotting routines with ``matplotlib`` similar syntax (see Plotting_)
 
 
-.. _IPython Tools: http://docs.vtki.org/tools/ipy_tools.html
-.. _Filters: http://docs.vtki.org/tools/filters.html
-.. _Plotting: http://docs.vtki.org/tools/plotting.html
+.. _IPython Tools: http://docs.pyvista.org/tools/ipy_tools.html
+.. _Filters: http://docs.pyvista.org/tools/filters.html
+.. _Plotting: http://docs.pyvista.org/tools/plotting.html
 
 
 Connections
@@ -140,7 +140,7 @@ Please take a look at the `contributors page`_ and the active `list of authors`_
 to learn more about the developers of ``vtki``.
 
 .. _contributors page: https://GitHub.com/akaszynski/vtki/graphs/contributors/
-.. _list of authors: http://docs.vtki.org/authors
+.. _list of authors: http://docs.pyvista.org/authors
 
 
 Contributing
@@ -153,4 +153,4 @@ need to foster a community that can support user questions and develop new
 features to make this software a useful tool for all users. To learn more about
 contributing to ``vtki``, please see the `Contributing Guide`_.
 
-.. _Contributing Guide: http://docs.vtki.org/dev/contributing.html
+.. _Contributing Guide: http://docs.pyvista.org/dev/contributing.html
