+++ joss/paper.md
@@ -1,5 +1,5 @@
 ---
-title: 'vtki: A Streamlined Python Interface for the Visualization Toolkit'
+title: 'vista: A Streamlined Python Interface for the Visualization Toolkit'
 tags:
   - Python
   - visualization
@@ -31,13 +31,13 @@ a powerful scientific visualization software library, and with Python bindings,
 it combines the speed of C++ with the rapid prototyping of Python.
 Despite this, VTK code programmed in Python using the base VTK python package
 is unnecessarily complicated as its API merely binds existing C++ calls.
-The `vtki` Python package provides a concise, well-documented interface
+The `vista` Python package provides a concise, well-documented interface
 exposing VTK’s powerful visualization backend; enabling researchers to
 rapidly explore large datasets, communicate their spatial findings, and
-facilitate reproducibility. `vtki` further seeks to simplify standard mesh
+facilitate reproducibility. `vista` further seeks to simplify standard mesh
 creation and plotting routines without compromising on the speed of the C++
 VTK backend. Figure 1 demonstrates an integrated scene of geospatial data
-generated by `vtki`; to learn more about how this visualization was created,
+generated by `vista`; to learn more about how this visualization was created,
 please visit [PVGeo's FORGE project website](http://forge.pvgeo.org).
 
 
@@ -54,7 +54,7 @@ Reading a VTK supported file and plotting it requires a user to write a
 complicated sequence of routines to render the data object while
 having to remember which VTK classes to use for file reading and dataset mapping.
 An example can be found in [this creative commons VTK example](https://vtk.org/Wiki/VTK/Examples/Python/STLReader).
-`vtki` includes plotting routines that are intended to be intuitive and
+`vista` includes plotting routines that are intended to be intuitive and
 highly controllable with `matplotlib` [@matplotlib] similar syntax and keyword
 arguments. These plotting routines are defined to make the process of
 visualizing spatially referenced data straightforward and easily implemented
@@ -63,14 +63,14 @@ like that shown in Figure 1.
 
 
 VTK implements an object-oriented approach to 3D visualization [@vtkbook],
-and `vtki` adhere's to that underlying structure to provide an API that
+and `vista` adhere's to that underlying structure to provide an API that
 expands on VTK's data types. These expanded, wrapped types hold methods and
 attributes for quickly accessing scalar arrays, inspecting properties of
 the dataset, or using filtering algorithms to transform datasets.
-`vtki` wrapped objects have a suite of common filters ready for immediate
+`vista` wrapped objects have a suite of common filters ready for immediate
 use directly on the objects. These filters are commonly used algorithms in the
 VTK library that have been made more accessible by binding a method to control
-that algorithm directly onto all `vtki` datasets, providing a shared set of
+that algorithm directly onto all `vista` datasets, providing a shared set of
 functionality. Through the use of these bound filtering methods, powerful VTK
 algorithms can be leveraged and controlled via keyword arguments designed to
 be intuitive for novice users.
@@ -83,26 +83,26 @@ slicing through volume, and D) geometric glyphing at mesh nodes and scaled by
 a scalar array.](./images/filters.png)
 
 
-At its core, `vtki` is a pure Python helper module for VTK
+At its core, `vista` is a pure Python helper module for VTK
 that interfaces back to VTK data objects through NumPy [@numpy]
 and direct array access adhering to VTK's object-oriented approach to
 3D visualization [@vtkbook].
-The `vtki` Python package provides an accessible and intuitive interface back
+The `vista` Python package provides an accessible and intuitive interface back
 to the VTK library to facilitate rapid prototyping, analysis, and visual
[…]
