+++ joss/images/code.py
@@ -1,9 +1,9 @@
 # Obligatory set up code
-import vtki
-from vtki import examples
+import vista
+from vista import examples
 import numpy as np
 # Set a document-friendly plotting theme
-vtki.set_plot_theme('document')
+vista.set_plot_theme('document')
 
 # Load an example uniform grid
 dataset = examples.load_uniform()
@@ -13,10 +13,10 @@ threshed = dataset.threshold([100, 500]) # Figure 4 A
 outline = dataset.outline()
 contours = dataset.contour() # Figure 4 B
 slices = dataset.slice_orthogonal() # Figure 4 C
-glyphs = dataset.glyph(factor=1e-3, geom=vtki.Sphere()) # Figure 4 D
+glyphs = dataset.glyph(factor=1e-3, geom=vista.Sphere()) # Figure 4 D
 
 # Two by two comparison
-vtki.plot_compare_four(threshed, contours, slices, glyphs,
+vista.plot_compare_four(threshed, contours, slices, glyphs,
                         {'show_scalar_bar':False},
                         {'border':False},
                         camera_position=[-2,5,3], outline=outline,
