+++ vtki/sphinx_gallery.py
@@ -1,15 +1,15 @@
 """
-Utilities for using vtki with sphinx-gallery.
+Utilities for using vista with sphinx-gallery.
 """
 
 import shutil
 
-import vtki
+import vista
 
 
 class Scraper(object):
     """
-    Save ``vtki.Plotter`` objects.
+    Save ``vista.Plotter`` objects.
 
     Used by sphinx-gallery to generate the plots from the code in the examples.
 
@@ -28,7 +28,7 @@ class Scraper(object):
             raise ImportError('You must install `sphinx_gallery`')
         image_names = list()
         image_path_iterator = block_vars["image_path_iterator"]
-        figures = vtki.plotting._ALL_PLOTTERS
+        figures = vista.plotting._ALL_PLOTTERS
         for address, plotter in figures.items():
             fname = next(image_path_iterator)
             if hasattr(plotter, '_gif_filename'):
@@ -37,5 +37,5 @@ class Scraper(object):
             else:
                 plotter.screenshot(fname)
             image_names.append(fname)
-        vtki.close_all() # close and clear all plotters
+        vista.close_all() # close and clear all plotters
         return figure_rst(image_names, gallery_conf["src_dir"])
