+++ tests/test_container.py
@@ -5,54 +5,54 @@ import numpy as np
 import pytest
 import vtk
 
-import vtki
-from vtki import examples as ex
-from vtki.plotting import system_supports_plotting
+import vista
+from vista import examples as ex
+from vista.plotting import system_supports_plotting
 
 
 def test_multi_block_init_vtk():
     multi = vtk.vtkMultiBlockDataSet()
     multi.SetBlock(0, vtk.vtkRectilinearGrid())
     multi.SetBlock(1, vtk.vtkTable())
-    multi = vtki.MultiBlock(multi)
-    assert isinstance(multi, vtki.MultiBlock)
+    multi = vista.MultiBlock(multi)
+    assert isinstance(multi, vista.MultiBlock)
     assert multi.n_blocks == 2
-    assert isinstance(multi[0], vtki.RectilinearGrid)
+    assert isinstance(multi[0], vista.RectilinearGrid)
     assert isinstance(multi[1], vtk.vtkTable)
     multi = vtk.vtkMultiBlockDataSet()
     multi.SetBlock(0, vtk.vtkRectilinearGrid())
     multi.SetBlock(1, vtk.vtkTable())
-    multi = vtki.MultiBlock(multi, deep=True)
-    assert isinstance(multi, vtki.MultiBlock)
+    multi = vista.MultiBlock(multi, deep=True)
+    assert isinstance(multi, vista.MultiBlock)
     assert multi.n_blocks == 2
-    assert isinstance(multi[0], vtki.RectilinearGrid)
+    assert isinstance(multi[0], vista.RectilinearGrid)
     assert isinstance(multi[1], vtk.vtkTable)
 
 def test_multi_block_init_dict():
     data = dict()
     data['grid'] = ex.load_rectilinear()
     data['poly'] = ex.load_airplane()
-    multi = vtki.MultiBlock(data)
-    assert isinstance(multi, vtki.MultiBlock)
+    multi = vista.MultiBlock(data)
+    assert isinstance(multi, vista.MultiBlock)
     assert multi.n_blocks == 2
     # Note that disctionaries do not maintain order
-    assert isinstance(multi[0], (vtki.RectilinearGrid, vtki.PolyData))
+    assert isinstance(multi[0], (vista.RectilinearGrid, vista.PolyData))
     assert multi.get_block_name(0) in ['grid','poly']
-    assert isinstance(multi[1], (vtki.RectilinearGrid, vtki.PolyData))
+    assert isinstance(multi[1], (vista.RectilinearGrid, vista.PolyData))
     assert multi.get_block_name(1) in ['grid','poly']
 
 def test_multi_block_init_list():
     data = [ex.load_rectilinear(), ex.load_airplane()]
-    multi = vtki.MultiBlock(data)
-    assert isinstance(multi, vtki.MultiBlock)
+    multi = vista.MultiBlock(data)
+    assert isinstance(multi, vista.MultiBlock)
     assert multi.n_blocks == 2
-    assert isinstance(multi[0], vtki.RectilinearGrid)
-    assert isinstance(multi[1], vtki.PolyData)
+    assert isinstance(multi[0], vista.RectilinearGrid)
+    assert isinstance(multi[1], vista.PolyData)
 
 
 def test_multi_block_append():
     """This puts all of the example data objects into a a MultiBlock container"""
-    multi = vtki.MultiBlock()
+    multi = vista.MultiBlock()
     # Add examples
     multi.append(ex.load_ant())
     multi.append(ex.load_sphere())
@@ -62,32 +62,32 @@ def test_multi_block_append():
     # Now check everything
     assert multi.n_blocks == 5
     assert multi.bounds is not None
-    assert isinstance(multi[0], vtki.PolyData)
[…]
