+++ examples/02-plot/plot-edges.py
@@ -12,8 +12,8 @@ Show the edges of all geometries within a mesh
 # the ``show_edges`` keyword argument to ``True`` when plotting a dataset.
 
 # sphinx_gallery_thumbnail_number = 1
-import vtki
-from vtki import examples
+import vista
+from vista import examples
 
 nefertiti = examples.download_nefertiti()
 

[…]
