+++ vtki/_version.py
@@ -1,6 +1,6 @@
 """ version info for vtki """
 # major, minor, patch
-version_info = 0, 18, 4
+version_info = 0, 18, 5
 
 # Nice string for the version
 __version__ = '.'.join(map(str, version_info))
