+++ .travis.yml
@@ -34,147 +34,150 @@ before_script:
   - echo $TRAVIS_BRANCH
   - echo $TRAVIS_BUILD_NUMBER
   - echo $TRAVIS_REPO_SLUG
+#
+# matrix:
+#   include:
+#     # Test the core API on Python 3.5, 3.6, and 3.7
+#     - name: Core API 3.5
+#       python: 3.5
+#       # Run the test suite and generate coverage report
+#       script: pytest -v --cov vtki
+#     - name: Core API 3.6
+#       python: 3.6
+#       # Run the test suite and generate coverage report
+#       script: pytest -v --cov vtki
+#     - name: Core API 3.7
+#       python: 3.7
+#       env:
+#         - REPORT_COVERAGE: true
+#         # Deploy to PyPI when core testing passes on 3.7 on tags
+#         - VTKI_DEPLOY: true
+#       # Run the test suite and generate coverage report
+#       script: pytest -v --cov vtki
+#
+#     # Test the package docstrings on Python 3.5, 3.6, and 3.7
+#     - name: Package Docstrings 3.5
+#       python: 3.5
+#       # Run all code examples in the docstrings
+#       script: pytest -v --doctest-modules vtki
+#     - name: Package Docstrings 3.6
+#       python: 3.6
+#       # Run all code examples in the docstrings
+#       script: pytest -v --cov --doctest-modules vtki
+#     - name: Package Docstrings 3.7
+#       python: 3.7
+#       env:
+#         - REPORT_COVERAGE: true
+#       # Run all code examples in the docstrings
+#       script: pytest -v --cov --doctest-modules vtki
+#
+#     # Test notebooks on Python 3.6 and 3.7
+#     - name: Notebooks 3.6
+#       python: 3.6
+#       # Now make sure notebooks are running
+#       # TODO: add below: pytest -v --nbval-lax --current-env --disable-warnings tests/*.ipynb;
+#       script: |
+#         pytest -v --nbval-lax --current-env --disable-warnings tests/*.ipynb;
+#         pytest -v --nbval-lax --current-env --disable-warnings notebooks/*.ipynb;
+#     - name: Notebooks 3.6 with panel
+#       python: 3.6
+#       # Now make sure notebooks are running
+#       # Install panel from master branch to keep up with bug fixes
+#       script: |
+#         git clone https://github.com/pyviz/panel.git;
+#         mv panel .panel
+#         cd .panel
+#         python setup.py install
+#         cd ..
+#         pytest -v --nbval-lax --current-env --disable-warnings tests/*.ipynb;
+#         pytest -v --nbval-lax --current-env --disable-warnings notebooks/*.ipynb;
+#     - name: Notebooks 3.7 with panel
+#       python: 3.7
+#       # Now make sure notebooks are running
+#       # Install panel from master branch to keep up with bug fixes
+#       script: |
+#         git clone https://github.com/pyviz/panel.git;
+#         mv panel .panel
+#         cd .panel
+#         python setup.py install
+#         cd ..
+#         pytest -v --nbval-lax --current-env --disable-warnings tests/*.ipynb;
+#         pytest -v --nbval-lax --current-env --disable-warnings notebooks/*.ipynb;
+#     - name: Notebooks 3.7
+#       python: 3.7
+#       env:
+#         - REPORT_COVERAGE: false
+#       # Now make sure notebooks are running
+#       script: |
+#         pytest -v --nbval-lax --current-env --disable-warnings tests/*.ipynb;
+#         pytest -v --nbval-lax --current-env --disable-warnings notebooks/*.ipynb;
+#
+#     # Build/test/deploy documentation on 3.7 only
[…]
