+++ docs/getting-started/installation.rst
@@ -125,7 +125,7 @@ time your Docker container is launched:
     #!/bin/bash
     set -x
     export DISPLAY=:99.0
-    export VTKI_OFF_SCREEN=True
+    export VISTA_OFF_SCREEN=True
     which Xvfb
     Xvfb :99 -screen 0 1024x768x24 > /dev/null 2>&1 &
     sleep 3
@@ -172,7 +172,7 @@ Then, configure the headless display:
 
     sudo apt-get install xvfb
     export DISPLAY=:99.0
-    export VTKI_OFF_SCREEN=True
+    export VISTA_OFF_SCREEN=True
     Xvfb :99 -screen 0 1024x768x24 > /dev/null 2>&1 &
     sleep 3
 
