### 2019-05-08 3472626
Fix docs issues from refactoring
