+++ docs/why.rst
@@ -1,5 +1,5 @@
 Why ``vista``?
-=============
+==============
 
 VTK is an excellent visualization toolkit, and with Python bindings it should be
 able to combine the speed of C++ with the rapid prototyping of Python.
@@ -57,7 +57,7 @@ loading and plotting an STL file requires a lot of code when using only the
 
 
 Plot a Mesh using vista
-~~~~~~~~~~~~~~~~~~~~~~
+~~~~~~~~~~~~~~~~~~~~~~~
 
 The same stl can be loaded and plotted using vista with:
 
