### 2019-05-08 9542481
Rename vtki → vista (#211)
* Refactor vtkiorg → pyvista

* vtki.org → pyvista.org

* vtki → vista

* rename *vtki* files → *vista*

* Update contact info

* Fix README banner

* VTKI -> VISTA

* Stop autogenerating logo

* Fix docs issues from refactoring

* Update README title
