+++ vista/examples/__init__.py
@@ -1,0 +1,2 @@
+from vista.examples.examples import *
+from vista.examples.downloads import *
