+++ vista/_version.py
@@ -1,6 +1,6 @@
 """ version info for vista """
 # major, minor, patch
-version_info = 0, 18, 4
+version_info = 0, 19, 0
 
 # Nice string for the version
 __version__ = '.'.join(map(str, version_info))
