### 2019-05-08 a932119
Fix docs deployment
