+++ .travis.yml
@@ -10,7 +10,7 @@ language: python
 env:
   global:
     # Doctr deploy key for pyvista/vista
-    - secure: "NHUV04B//rkOUxfBd5Bk3QN3g/iQ91wdRZowNTwCIPAbznFez2mbwQcsULDeZk7blPsv1xEAnqy/T/nPyYZuHO9qU2/OElvxKspaKgcGZ9Ig/czmjrO7Y5VnU2KOZJNwBLgb6mEpDvkKoI2j9WhhcTuPcDbayk6ntXz03snYRUsn2nwHfvqDmIqMQnqSwgVejPJQxBsQSikzYzkTYT9P/z94VoSTp4UrvDVj2Tygrw4PnSGe/9x6qTolo3Ofd+REt4LBth6/lBX2OrBhHwnp9wWpEX1o98HOdf5ZUHI9v8gZacnncVXNlNu1biBlQR58bXZX/4RqRdM2No6NcszC8CbhAS+UhNdsPlbimt9L60s4MWwq+IaxCOA2/lzLVdFl8p5wDjz+DDZsdVMNyjhO3MBP5Am/4qJ0BsQEQ2vtZMF3Jvil9oiSSWcp7IGYsj3aPqEcKFm5uc2SwU1CXg4uBfgDsowqBi0t5/1rWqsJQ6p+6ZlU2NoN4ppNjuenAYyDla49J6I7wLp6Z3GwHuupE75WncVhL5BY5aYRcTDrkKuRx8nFIMe+Xuzp6NKmWaDxIwoztA0VPZ9FtMYiM5ahvHwSnZ8wIxUngjMHPuE7+e6KkCBcLt/e7DlzBQhi4kPJesbzw3NX/VmYkAoZsK5tkwKBZakKIKudpjzW73DxGE8="
+    - secure: "WAFIXdY3+gZaSYp5AKkVcEqs1tUlt27t1AsisyuZMSxGCrWJ6QzvYUsNH1ayb/0g0llnrY02d3FsLMswNpdXhD7A/JbfMZly06duplwvrUcZHw+YaVAZ3XECs7E5x58Pn9T6vqWq7TUWN2HFdjVMZFpKsbKnTt2EBfqabX2bRHioJKzfN4ilYHNP/8J1GRel7HMKKYkJ5ImVxpBwkevaH4OCzJIIbndKw3qcDnHnpA2dHENMCPWSShHtE6LoStqoEPj44YHN5MZDgv1HutRa2M9tC9Jfq2DeDUICEjq1LJK109ksq+vwmL8AAUm+RPLNw9AQ6agPX7hCeUE0/6FJqWrGKwMrneCXBSWh7P96Fyvvly6yNieK7UJoSLFUqT47i98W9fTWLdDvsyCO8O+xtHTPqnw1oGnseOB4+gXmT2ljHR2RWntXwWrUy3SlKhzjKfdNLzZQJIWnYaBOszJJAquDvYTdsbrkDugnKZJkuZBvHbIGGWfIdkLhu66z0nBS7rh1Xt/8KTdVYRUcUG2cZfmEzoksbqB7QO6N3zNf1kZnsI8WWMMbHj0iKLycWBQOZvwR0NyLiNlyL88uEh1nuoUORhSF8v8GA2jmjPgiksOJGKQGvfh5bnfqMsOVC3oAhfqeoXaLlrORpatJrdJ0PsfbfLh/EKxTrtH2ctsq66k="
 
 before_install:
   # configure a headless display
