### 2019-05-08 pyvista/vista#208 @florian-wagner
Just joking. I like the name and the software, independent of its name, a lot. Next pyGIMLi release will ship with a vista-based 3D viewer. Thanks for the great work.
