### 2019-05-08 #1382 @leouieda
@Chilipp sorry this is taking so long. @banesullivan is in the process of renaming the vtki package to `vista` (see pyvista/vista#208). We'll need to update this issue before we can start the review.
