### 2019-05-08 #1382 @leouieda
:wave: Hi @RcSepp would be interested in reviewing this submission for JOSS?
