### 2019-05-09 pyvista/vista#208 @banesullivan
The transition is almost complete! Shoutout to @lheagy for recommending the name `vista` 🎆  and major thanks to everyone here that posted suggestions and their thoughts!

It really means a lot to us that all of you are so active and interested in seeing `vista` succeed - I never could have imagined this software would be embraced by you all so rapidly - thanks for the support everyone!

- `vista` version `0.19.0` is live on PyPI 🎉
- `omfvista` version `0.1.0` is live on PyPI 🎉
- `PVGeo` is showing green lights on after refactoring ✅
- I've raised issues with all external projects I know of leveraging `vtki`/`vista` to let them know the name has changed. Unfortunately, I didn't take a look at the dependents graph before changing the name so I may be forgetting a few

I'm in the process of fixing all the deployment tests but all is looking good at this point.

I need some help fixing the Conda-forge deployment... @hadim originally implemented that for us and to be honest I have no idea what to do to fix it: https://github.com/conda-forge/vtki-feedstock/issues/14
