### 2019-05-09 23ea93a
Fix offline use of examples in subdirectories
