### 2019-05-09 26d4fa8
Add grid creation feature utitlity
