### 2019-05-09 a86c3ff
Enhance repr methods
