### 2019-05-09 4475ea9
Improve dimesnions docs
