+++ docs/external_examples.rst
@@ -18,8 +18,8 @@ button at the top of this page and adding a link to your project!
 .. toctree::
    :maxdepth: 2
 
-   3D visualization for the Open Mining Format (omf) <https://github.com/OpenGeoVis/omfvtk/blob/master/Example.ipynb>
-   Live example notebook showcasing omfvtk <https://mybinder.org/v2/gh/OpenGeoVis/omfvtk-demo/master?filepath=notebooks%2FExample.ipynb>
+   3D visualization for the Open Mining Format (omf) <https://github.com/OpenGeoVis/omfvista/blob/master/Example.ipynb>
+   Live example notebook showcasing omfvista <https://mybinder.org/v2/gh/OpenGeoVis/omfvista-demo/master?filepath=notebooks%2FExample.ipynb>
    3D Rendering with Discretize <http://discretize.simpeg.xyz/en/master/auto_examples/vista_demo.html#sphx-glr-auto-examples-vista-demo-py>
    OpenFOAM Rendering <https://pswpswpsw.github.io/posts/2018/09/blog-post-modify-vtk-openfoam/>
    FORGE Geothermal Project <http://forge.pvgeo.org>
