### 2019-05-10 pyvista/vista#208 @banesullivan
I think I'm going to rename the repository `pyvista` but still keep the package on PyPI/conda as `vista`. I think if we officially call the software "PyVista", we can avoid any potential legal issues with the name being too similar to all the other "vista" things out there.

This is exactly what PyTorch does: https://github.com/pytorch/pytorch

Any objections?
