### 2019-05-10 pyvista/vista#208 @banesullivan
Note the discussion in https://github.com/softwareunderground/awesome-open-geoscience/pull/78
