### 2019-05-10 pyvista/vista#208 @leouieda
I would suggest going all the way and renaming everything pyvista (package included). From my own experience, having my repository/pypi packages called `gmt-python` while the package itself was `gmt` generated a whole lot of trouble for new users and a bunch of confused bug reports.
