### 2019-05-10 pyvista/vista#208 @banesullivan
With a name that long, should we push an alias on import? Perhaps:

```py
import pyvista as pv

mesh = pv.StructuredGrid()
...

p = pv.Plotter()
```
