### 2019-05-10 pyvista/vista#208 @leouieda
I like `pyvista as pv`. That leads nicely to `pvgeo` which could now be `pv -> paraview` or `pv -> pyvista`.
