### 2019-05-10 pyvista/vista#208 @banesullivan
That’s an excellent point!

As for other offshoot projects like `omfvtk` which is `omfvista` at this moment, would `**vista` be appropriate (copyright/trademark friendly) or should we try to build a brand of `pv**` products that revolve around PyVista?
