+++ tests/test_grid.py
@@ -4,24 +4,24 @@ import numpy as np
 import pytest
 import vtk
 
-import vista
-from vista import examples
-from vista.plotting import system_supports_plotting
+import pyvista
+from pyvista import examples
+from pyvista.plotting import system_supports_plotting
 
-beam = vista.UnstructuredGrid(examples.hexbeamfile)
+beam = pyvista.UnstructuredGrid(examples.hexbeamfile)
 
 # create structured grid
 x = np.arange(-10, 10, 2)
 y = np.arange(-10, 10, 2)
 z = np.arange(-10, 10, 2)
 x, y, z = np.meshgrid(x, y, z)
-sgrid = vista.StructuredGrid(x, y, z)
+sgrid = pyvista.StructuredGrid(x, y, z)
 
 try:
     test_path = os.path.dirname(os.path.abspath(__file__))
     test_data_path = os.path.join(test_path, 'test_data')
 except:
-    test_path = '/home/alex/afrl/python/source/vista/tests'
+    test_path = '/home/alex/afrl/python/source/pyvista/tests'
 
 
 def test_volume():
@@ -41,22 +41,22 @@ def test_struct_example():
 
 
 def test_init_from_structured():
-    unstruct_grid = vista.UnstructuredGrid(sgrid)
+    unstruct_grid = pyvista.UnstructuredGrid(sgrid)
     assert unstruct_grid.points.shape[0] == x.size
     assert np.all(unstruct_grid.celltypes == 12)
 
 
 def test_init_from_unstructured():
-    grid = vista.UnstructuredGrid(beam, deep=True)
+    grid = pyvista.UnstructuredGrid(beam, deep=True)
     grid.points += 1
     assert not np.any(grid.points == beam.points)
 
 def test_init_bad_input():
     with pytest.raises(Exception):
-        unstruct_grid = vista.UnstructuredGrid(np.array(1))
+        unstruct_grid = pyvista.UnstructuredGrid(np.array(1))
 
     with pytest.raises(Exception):
-        unstruct_grid = vista.UnstructuredGrid(np.array(1),
+        unstruct_grid = pyvista.UnstructuredGrid(np.array(1),
                                               np.array(1),
                                               np.array(1),
                                               'woa')
@@ -86,7 +86,7 @@ def test_init_from_arrays():
                       [0, 1, 3]])
 
     points = np.vstack((cell1, cell2)).astype(np.int32)
-    grid = vista.UnstructuredGrid(offset, cells, cell_type, points)
+    grid = pyvista.UnstructuredGrid(offset, cells, cell_type, points)
 
     assert grid.n_cells == 2
     assert np.allclose(grid.offset, offset)
@@ -111,28 +111,28 @@ def test_save(extension, binary, tmpdir):
     filename = str(tmpdir.mkdir("tmpdir").join('tmp.%s' % extension))
     beam.save(filename, binary)
 
-    grid = vista.UnstructuredGrid(filename)
+    grid = pyvista.UnstructuredGrid(filename)
     assert grid.cells.shape == beam.cells.shape
     assert grid.points.shape == beam.points.shape
 
-    grid = vista.read(filename)
+    grid = pyvista.read(filename)
     assert grid.cells.shape == beam.cells.shape
     assert grid.points.shape == beam.points.shape
-    assert isinstance(grid, vista.UnstructuredGrid)
+    assert isinstance(grid, pyvista.UnstructuredGrid)
 
 
 def test_init_bad_filename():
     filename = os.path.join(test_path, 'test_grid.py')
[…]
