+++ examples/01-filter/compute-volume.py
@@ -12,9 +12,9 @@ import pyvista
 from pyvista import examples
 
 ################################################################################
-# Computing mass properties such as the volume or area of datasetsin ``pyvista``
+# Computing mass properties such as the volume or area of datasetsin PyVista
 # is quite easy using the :func:`pyvista.DataSetFilters.compute_cell_sizes` filter
-# and the :attr:`pyvista.Common.volume` property on all ``pyvista`` meshes.
+# and the :attr:`pyvista.Common.volume` property on all PyVista meshes.
 #
 # Let's get started with a simple gridded mesh:
 
