### 2019-05-11 5d20e5f
Fix docs issues
