+++ AUTHORS.rst
@@ -1,7 +1,7 @@
 The following is a list of authors who have made substantial contributions to
 the conception or design of this software; or the creation of new code used in
 this software; or have drafted the work or substantively revised it and are
-considered "The pyvista Developers":
+considered "The PyVista Developers":
 
 - Alex Kaszynski, (`@akaszynski <https://github.com/akaszynski/>`_)
 - Bane Sullivan, (`@banesullivan <https://github.com/banesullivan/>`_)
