+++ docs/pyvista-binder-config.yml
@@ -2,5 +2,5 @@ default_context:
     repo_name: "pyvista-examples"
     github_account: "pyvista"
     description: "Quick examples demonstrating what PyVista can do! Simply launch this Binder project and test drive PyVista"
-    author_name: "pyvista Developers"
+    author_name: "PyVista Developers"
     year: "2019"
