### 2019-05-11 e226507
Fix doctr deployment
