+++ .travis.yml
@@ -10,7 +10,7 @@ language: python
 env:
   global:
     # Doctr deploy key for pyvista/pyvista
-    - secure: "WAFIXdY3+gZaSYp5AKkVcEqs1tUlt27t1AsisyuZMSxGCrWJ6QzvYUsNH1ayb/0g0llnrY02d3FsLMswNpdXhD7A/JbfMZly06duplwvrUcZHw+YaVAZ3XECs7E5x58Pn9T6vqWq7TUWN2HFdjVMZFpKsbKnTt2EBfqabX2bRHioJKzfN4ilYHNP/8J1GRel7HMKKYkJ5ImVxpBwkevaH4OCzJIIbndKw3qcDnHnpA2dHENMCPWSShHtE6LoStqoEPj44YHN5MZDgv1HutRa2M9tC9Jfq2DeDUICEjq1LJK109ksq+vwmL8AAUm+RPLNw9AQ6agPX7hCeUE0/6FJqWrGKwMrneCXBSWh7P96Fyvvly6yNieK7UJoSLFUqT47i98W9fTWLdDvsyCO8O+xtHTPqnw1oGnseOB4+gXmT2ljHR2RWntXwWrUy3SlKhzjKfdNLzZQJIWnYaBOszJJAquDvYTdsbrkDugnKZJkuZBvHbIGGWfIdkLhu66z0nBS7rh1Xt/8KTdVYRUcUG2cZfmEzoksbqB7QO6N3zNf1kZnsI8WWMMbHj0iKLycWBQOZvwR0NyLiNlyL88uEh1nuoUORhSF8v8GA2jmjPgiksOJGKQGvfh5bnfqMsOVC3oAhfqeoXaLlrORpatJrdJ0PsfbfLh/EKxTrtH2ctsq66k="
+    - secure: "LffT6SJeAEvH2qD6nJXbw8E9VW5ji5vm6bLywUdEjequgKDuV8SvkKuZQq6pBrz2l1YEWlJ70c9oShz7lJ7H2mpteIpjvf5qDx6eXiRyOSwWdsvr8sbQe8C7J0nGxnfEWSANvZ9IA7gXwURbs3wzbnLoUpIiubW8UPnbfk8KAymh3S3wDwNrUXNszc4pzVE/0tkEjIRft1EYfdUpBQfZ1pABltkQcMFqGjDR6LFmkU7EONrRw4cKzg5IFSWRfHtS6S04i+c5PXuMoReM7EkFlUKO4j+fKHJhSxlSuMGYugVlroXWZUjVNZuEWjF1mKLuQnq13FzUHr1skbUXl2LgWHHmHGJ76+8ZtvSPhuOSll4kgWjTVplnW4iqbPcGWnvs3GnSeuAtz5gOjpKbY879/QpFanTXSA4a/WTy0D87aRKjn2p2zQ46/d8oeIfdo1ww59b983ubjJ4jccqlANVyw6AUuND1iXtktCN485SnUfX5bHEj5kEbpRmWBTIfDeMLgyYq7KM0ZHTfd5Zd+V1JD9idJxw3o3L7DytojPgjE190L+DhhyoZdWWZXGFF3DiivQdR/huKvaGG5o2Jg+k9/lBYmq1/AhnOJflakVHbcxQf2FR0DkylMclgLo8X5+MySkMsiSU9kHCCQgzzOOUPClRdxN+4i/tuXn1/H97SAPo="
 
 before_install:
   # configure a headless display
