### 2019-05-12 ced600b
Bump version: 0.19.0 -> 0.20.0
