### 2019-05-12 792b5da
alias import pyvista in examples
