+++ examples/00-load/create-unstructured-surface.py
@@ -7,7 +7,7 @@ Creating an Unstructured Surface
 Create an irregular, unstructured grid from NumPy arrays
 """
 
-import pyvista
+import pyvista as pv
 import vtk
 import numpy as np
 
@@ -50,7 +50,7 @@ cell2 = np.array([[0, 0, 2],
 points = np.vstack((cell1, cell2))
 
 # create the unstructured grid directly from the numpy arrays
-grid = pyvista.UnstructuredGrid(offset, cells, cell_type, points)
+grid = pv.UnstructuredGrid(offset, cells, cell_type, points)
 
 # plot the grid
 grid.plot(show_edges=True)
