+++ examples/01-filter/volume-smoothing.py
@@ -6,7 +6,7 @@ Smoothing rough edges of a volumetric surface
 """
 
 # sphinx_gallery_thumbnail_number = 4
-import pyvista
+import pyvista as pv
 from pyvista import examples
 
 ################################################################################
