+++ examples/00-load/read-parallel.py
@@ -6,7 +6,7 @@ The VTK library supports parallel file foramts. Reading meshes broken up into
 several files is natively supported by VTK and PyVista
 """
 # sphinx_gallery_thumbnail_number = 1
-import pyvista
+import pyvista as pv
 from pyvista import examples
 import os
 
@@ -24,7 +24,7 @@ examples.download_blood_vessels()
 # files.
 #
 # Let's inspect where this downloaded our dataset:
-path = os.path.join(pyvista.EXAMPLES_PATH, 'blood_vessels')
+path = os.path.join(pv.EXAMPLES_PATH, 'blood_vessels')
 print(os.listdir(path))
 
 ################################################################################
@@ -35,7 +35,7 @@ print(os.listdir(os.path.join(path, 'T0000000500')))
 # contains all the parallel files or pieces that make the whole mesh. We can
 # simply read the ``.pvtu`` file and VTK will handle putting the mesh together.
 filename = os.path.join(path, 'T0000000500.pvtu')
-mesh = pyvista.read(filename)
+mesh = pv.read(filename)
 print(mesh)
 
 ################################################################################
