+++ examples/01-filter/resample.py
@@ -9,18 +9,18 @@ Resample one mesh's point/cell arrays onto another meshes nodes.
 # of a sphere contained in that volume.
 
 # sphinx_gallery_thumbnail_number = 4
-import pyvista
+import pyvista as pv
 from pyvista import examples
 import numpy as np
 
 ###############################################################################
 # Querry a grids points onto a sphere
-mesh = pyvista.Sphere(center=(4.5,4.5,4.5), radius=4.5)
+mesh = pv.Sphere(center=(4.5,4.5,4.5), radius=4.5)
 data_to_probe = examples.load_uniform()
 
 ###############################################################################
 # Plot the two datasets
-p = pyvista.Plotter()
+p = pv.Plotter()
 p.add_mesh(mesh, color=True)
 p.add_mesh(data_to_probe, opacity=0.5)
 p.show()
@@ -46,7 +46,7 @@ surface = examples.download_saddle_surface()
 points = examples.download_sparse_points()
 
 
-p = pyvista.Plotter()
+p = pv.Plotter()
 p.add_mesh(points, point_size=30.0, render_points_as_spheres=True)
 p.add_mesh(surface)
 p.show()
@@ -57,7 +57,7 @@ p.show()
 interpolated = surface.interpolate(points, radius=12.0)
 
 
-p = pyvista.Plotter()
+p = pv.Plotter()
 p.add_mesh(points, point_size=30.0, render_points_as_spheres=True)
 p.add_mesh(interpolated, scalars='val')
 p.show()
