+++ pyvista/_version.py
@@ -1,6 +1,6 @@
 """ version info for pyvista """
 # major, minor, patch
-version_info = 0, 20, 0
+version_info = 0, 20, 1
 
 # Nice string for the version
 __version__ = '.'.join(map(str, version_info))
