### 2019-05-12 #1382 @banesullivan
We have settled on PyVista as the new name of this software - sorry for the delay!
