### 2019-05-13 32227f1
Fix default panel issue
