### 2019-05-13 9c7417c
Update Travis script
