### 2019-05-13 #1382 @banesullivan
Please note, @akaszynski and I have changed the paper title to "PyVista: 3D plotting and mesh analysis through a streamlined interface for the Visualization Toolkit (VTK)"
