+++ docs/index.rst
@@ -1,5 +1,8 @@
 .. title:: pyvista
 
+**🚨🚨 PyVista was formerly vtki 🚨🚨 We recently had to change the name of this
+software and we apologize for any confusion this may be causing**
+
 .. raw:: html
 
     <div class="banner">
