### 2019-05-15 #1382 @leouieda
OK, I just renamed the issue and updated the link in the description to the new package. Everything seems to work and the link inserted in the PDF is the correct one now. I'm just waiting on some reviewer invitations to start the review. Hopefully that can be resolved within a week or so.
