### 2019-05-15 0484114
Add support details and JOSS badge
