### 2019-05-15 #1450 @banesullivan
Probably not a big issue but I see it's on the checklist - `@whedon` marks version `0.18.1` in the checklist but PyVista is released officially on `0.20.1`
