### 2019-05-15 #1450 @nicoguaro
I did a first review of the package and I have the following comments:

- The installation instructions for ``pip`` work properly while they do not work for ``conda``.
- The examples work properly, although the colors differ from the documentation.
