### 2019-05-15 #1450 @banesullivan
Thanks for the comments @nicoguaro

> The installation instructions for pip work properly while they do not work for conda.

we are currently waiting on the Conda-forge team to merge our anaconda recipe in https://github.com/conda-forge/staged-recipes/pull/8394

> The examples work properly, although the colors differ from the documentation.

This has to do with how we allow users to set plotting themes via `pyvista.set_plot_theme` the default theme a user experiences is a bit different than the `'document'` theme we use when building the docs. To have the same colors, you could add:

```py
import pyvista as pv
pv.set_plot_theme('document')
```

which is what we add in out `conf.py` when building the documentation/examples
