### 2019-05-15 c0cfe88
Restructure documentation navigation bar
