+++ docs/dev/contributing.rst
@@ -1,67 +1,0 @@
-.. _contributing_ref:
-
-Overview
-========
-
-This page is dedicated to outline where you should start with your question,
-concern, feature request, or desire to contribute.
-
-Questions
-~~~~~~~~~
-
-For general questions about the project, its applications, or about software
-usage, please do not create an issue but join us on Slack_ or send one
-of the developers an email. The project support team can be reached at
-`info@pyvista.org`_.
-
-.. _Slack: http://slack.opengeovis.org
-.. _info@pyvista.org: mailto:info@pyvista.org
-
-For more technical questions, you are welcome to create an issue on the
-`issues page`_ which we will address promptly.
-Through posting on the issues page, your question can be addressed by community
-members with the needed expertise and the information gained will remain
-available on the issues page for other users.
-
-.. _issues page: https://github.com/pyvista/pyvista/issues
-
-Reporting Bugs
-~~~~~~~~~~~~~~
-
-If you stumble across any bugs, crashes, or concerning quirks while using code
-distributed here, please report it on the `issues page`_ with an
-appropriate label so we can promptly address it.
-When reporting an issue, please be overly descriptive so that we may reproduce
-it. Whenever possible, please provide tracebacks, screenshots, and sample files
-to help us address the issue.
-
-Feature Requests
-~~~~~~~~~~~~~~~~
-
-We encourage users to submit ideas for improvements to PyVista code base!
-Please create an issue on the `issues page`_ with a *Feature Request*
-label to suggest an improvement.
-Please use a descriptive title and provide ample background information to help
-the community implement that functionality. For example, if you would like a
-reader for a specific file format, please provide a link to documentation of
-that file format and possibly provide some sample files with screenshots to work
-with. We will use the issue thread as a place to discuss and provide feedback.
-
-Contributing New Code
-~~~~~~~~~~~~~~~~~~~~~
-
-If you have an idea for how to improve PyVista, please first create an issue as
-a feature request which we can use as a discussion thread to work through how to
-implement the contribution.
-
-Once you are ready to start coding and develop for PyVista, please take a look
-at the remainder of the pages in this *Development Guide*.
-
-Licensing
-~~~~~~~~~
-
-All contributed code will be licensed under The MIT License found in the
-repository. If you did not write the code yourself, it is your responsibility
-to ensure that the existing license is compatible and included in the
-contributed files or you can obtain permission from the original author to
-relicense the code.
