+++ docs/tools/filters.rst
@@ -1,14 +1,14 @@
 .. _filters_ref:
 
 Filters
-=======
+-------
 
 
 .. automodule:: pyvista.filters
 
 
 Dataset Filters
----------------
+~~~~~~~~~~~~~~~
 
 The :class:`pyvista.DataSetFilters` is inherited by :class:`pyvista.Common` making
 all the the following filters available as callable methods directly from any
