### 2019-05-15 5f7c4df
Update docs deploy condtion
