+++ docs/deploy_docs_from_travis.sh
@@ -8,7 +8,7 @@ find ./examples -type f -name '*.ipynb' | cpio -p -d -v ../../pyvista-examples/;
 cd ../../pyvista-examples/;
 git init;
 git add .;
-git commit -m "$TRAVIS_BUILD_NUMBER : Autogenerated notebooks from Travis";
+git commit -m "$TRAVIS_JOB_NUMBER : Autogenerated notebooks from Travis";
 REMOTE="https://${GH_TOKEN}@github.com/pyvista/pyvista-examples";
 git config --global user.name "${GH_NAME}";
 git config --global user.email "${GH_EMAIL}";
