### 2019-05-15 ad5fe4c
Attempt fix examples deployment
