+++ .travis.yml
@@ -33,6 +33,7 @@ before_script:
   - echo $TRAVIS_TAG
   - echo $TRAVIS_BRANCH
   - echo $TRAVIS_BUILD_NUMBER
+  - echo $TRAVIS_JOB_NUMBER
   - echo $TRAVIS_REPO_SLUG
 
 matrix:
