### 2019-05-15 pyvista/pyvista#208 @banesullivan
The name change appears to be complete! `pyvista` is live on both PyPI and Conda-forge and it looks to me like all of the dependents have caught up:

I'm going to close this issue but see if I can keep it pinned
