### 2019-05-15 a62c935
Add better warning about name change in docs
