### 2019-05-15 80a29f1
Create CODE_OF_CONDUCT.md
