### 2019-05-15 #1450 @Chilipp
Dear @banesullivan,

I am done with the review and I have to say that this is a very nice package that provides valuable new functionalities for visualizations with python. I worked with Paraview before and it's Python interface, and I have to say that your methodologies are both, useful and well-documented.

There are only a few minor issues:

- [ ] **version:** The version is not set correctly (as already mentioned in https://github.com/openjournals/joss-reviews/issues/1450#issuecomment-492497996 by @banesullivan)
- [ ] **dependendies:** You list the optional dependencies in the installation instructions, but not the necessary dependencies. You should be more transparent here
- [ ] **tests:** Although you state in your [CONTRIBUTING.md](https://github.com/pyvista/pyvista/blob/master/CONTRIBUTING.md#testing) how to run the tests locally, you do not state this in the installation instructions. I recommend to change this. You should at least include a link to the CONTRIBUTING.md file, such that someone who wants to verify the installation before using it, knows how to do this.

Otherwise, if the three issues are resolved, I recommend to accept this publication.
