### 2019-05-15 6e0762b
Update installation guide per https://github.com/openjournals/joss-reviews/issues/1450#issuecomment-492813464
