### 2019-05-15 9563643
Update testing guide per https://github.com/openjournals/joss-reviews/issues/1450#issuecomment-492813464
