### 2019-05-15 #1450 @banesullivan
@Chilipp, Please see changes in https://github.com/pyvista/pyvista/commit/6e0762b7ae946b6141746ba134d0b3526ee356f5 and https://github.com/pyvista/pyvista/commit/9563643928a2f907e3001c481194adccacf62c48 addressing your comment.

## Dependencies

> You list the optional dependencies in the installation instructions, but not the necessary dependencies. You should be more transparent here

I have updated the installation instructions to include an itemized list of PyVista's required dependencies and a brief description for why - please note the list here:  https://docs.pyvista.org/getting-started/installation.html#dependencies

## Testing

> Although you state in your CONTRIBUTING.md how to run the tests locally, you do not state this in the installation instructions. I recommend to change this. You should at least include a link to the CONTRIBUTING.md file, such that someone who wants to verify the installation before using it, knows how to do this.

I have added  a link in the installation guide to point directly to the "Testing" section of [CONTRIBUTING.md](https://github.com/pyvista/pyvista/blob/master/CONTRIBUTING.md#testing).

PyVista's tests are not deployed with the software but rather are available after cloning the repository so I added some text on that page to let users know that they can run PyVista's comprehensive testing suite after cloning the source and point them to [CONTRIBUTING.md](https://github.com/pyvista/pyvista/blob/master/CONTRIBUTING.md) for further details:

Considering that documentation page is intended to be viewed by a general audience (non-developers), I believe this approach to be welcoming to new users

@Chilipp - do you think our approach here is appropriate?
