### 2019-05-15 #1450 @leouieda
> Probably not a big issue but I see it's on the checklist - @whedon marks version 0.18.1 in the checklist but PyVista is released officially on 0.20.1

@banesullivan thanks for pointing it out :+1: We'll update this when the review is done. The reviews might lead to changes that should be released before publication.
