### 2019-05-15 #1450 @leouieda
@nicoguaro @Chilipp thank you so much for the quick reviews! :rocket: As @banesullivan addresses your comments/questions, please make sure you tick the corresponding items from the checklist when you're satisfied.

> The examples work properly, although the colors differ from the documentation.

@banesullivan this might be confusing to other users as well since they'll run the code and get something different from what is "advertised". It might be more transparent to include `pv.set_plot_theme('document')` in the examples. This way, the user also knows that this is possible.
