### 2019-05-15 49444c9
Add plot theme example per https://github.com/openjournals/joss-reviews/issues/1450
