+++ examples/02-plot/plot-edl.py
@@ -12,7 +12,7 @@ To learn more, please see `this blog post`_.
 ################################################################################
 
 # sphinx_gallery_thumbnail_number = 1
-import pyvista
+import pyvista as pv
 from pyvista import examples
 
 ################################################################################
@@ -29,7 +29,7 @@ nefertiti.plot(eye_dome_lighting=True, cpos=[-1,-1,0.2])
 ################################################################################
 # Here we will compare a EDL shading side by side with normal shading
 
-p = pyvista.Plotter(shape=(1, 2), border=False)
+p = pv.Plotter(shape=(1, 2), border=False)
 
 # With eye-dome lighting
 p.subplot(0,0)
@@ -60,7 +60,7 @@ point_cloud = examples.download_lidar()
 # And now plot this point cloud as-is:
 
 # Plot a typical point cloud with no EDL
-p = pyvista.Plotter()
+p = pv.Plotter()
 p.add_mesh(point_cloud, color='tan', point_size=5)
 p.show()
 
@@ -70,7 +70,7 @@ p.show()
 # Reference :func:`pyvista.Renderer.enable_eye_dome_lighting`.
 
 # Plot with EDL
-p = pyvista.Plotter()
+p = pv.Plotter()
 p.add_mesh(point_cloud, color='tan', point_size=5)
 p.enable_eye_dome_lighting()
 p.show()
@@ -80,7 +80,7 @@ p.show()
 # The eye dome lighting mode can also handle plotting scalar arrays:
 
 # Plot with EDL and scalar data
-p = pyvista.Plotter()
+p = pv.Plotter()
 p.add_mesh(point_cloud, scalars='Elevation', point_size=5)
 p.enable_eye_dome_lighting()
 p.show()
