+++ examples/02-plot/plot-labels.py
@@ -6,7 +6,7 @@ Use string arrays in a point set to label points
 """
 # sphinx_gallery_thumbnail_number = 3
 from pyvista import examples
-import pyvista
+import pyvista as pv
 import numpy as np
 
 ###############################################################################
@@ -17,7 +17,7 @@ import numpy as np
 # labels for each of the nodes.
 
 # Make some random points
-poly = pyvista.PolyData(np.random.rand(10, 3))
+poly = pv.PolyData(np.random.rand(10, 3))
 
 ###############################################################################
 # Add string labels to the point data - this associates a label with every node:
@@ -29,7 +29,7 @@ print(poly)
 ###############################################################################
 # Now plot the points with labels:
 
-plotter = pyvista.Plotter()
+plotter = pv.Plotter()
 plotter.add_point_labels(poly, 'My Labels', point_size=20, font_size=36)
 plotter.show()
 
@@ -41,12 +41,12 @@ plotter.show()
 # This example will label the nodes of a mesh with their coordinate locations
 
 # Load example beam file
-grid = pyvista.UnstructuredGrid(examples.hexbeamfile)
+grid = pv.UnstructuredGrid(examples.hexbeamfile)
 
 
 ###############################################################################
 # Create plotting class and add the unstructured grid
-plotter = pyvista.Plotter()
+plotter = pv.Plotter()
 plotter.add_mesh(grid, show_edges=True, color='tan')
 
 # Add labels to points on the yz plane (where x == 0)
@@ -73,7 +73,7 @@ plotter.show()
 mesh = examples.load_uniform().slice()
 
 ###############################################################################
-p = pyvista.Plotter()
+p = pv.Plotter()
 
 # Add the mesh:
 p.add_mesh(mesh, scalars='Spatial Point Data', show_edges=True)
