+++ examples/02-plot/plot-movie.py
@@ -6,15 +6,15 @@ Create an animated MP4 movie of a rendering scene
 """
 
 
-import pyvista
+import pyvista as pv
 import numpy as np
 
 filename = 'sphere-shrinking.mp4'
 
-mesh = pyvista.Sphere()
+mesh = pv.Sphere()
 mesh.cell_arrays['data'] = np.random.random(mesh.n_cells)
 
-plotter = pyvista.Plotter()
+plotter = pv.Plotter()
 # Open a movie file
 plotter.open_movie(filename)
 
