+++ examples/02-plot/plot-multi-window.py
@@ -6,7 +6,7 @@ Multi-Window Plot
 Subplotting: having multiple scenes in a signle window
 """
 
-import pyvista
+import pyvista as pv
 from pyvista import examples
 
 ################################################################################
@@ -15,7 +15,7 @@ from pyvista import examples
 # ``shape=(2, 2)``. Use the :func:`pyvista.BasePlotter.subplot` function to select
 # the subplot you wish to be the active subplot.
 
-plotter = pyvista.Plotter(shape=(2, 2))
+plotter = pv.Plotter(shape=(2, 2))
 
 plotter.subplot(0,0)
 plotter.add_text('Render Window 0', position=None, font_size=30)
@@ -23,11 +23,11 @@ plotter.add_mesh(examples.load_globe())
 
 plotter.subplot(0, 1)
 plotter.add_text('Render Window 1', font_size=30)
-plotter.add_mesh(pyvista.Cube(), show_edges=True, color='tan')
+plotter.add_mesh(pv.Cube(), show_edges=True, color='tan')
 
 plotter.subplot(1, 0)
 plotter.add_text('Render Window 2', font_size=30)
-sphere = pyvista.Sphere()
+sphere = pv.Sphere()
 plotter.add_mesh(sphere, scalars=sphere.points[:, 2])
 plotter.add_scalar_bar('Z')
 # plotter.add_axes()
@@ -35,7 +35,7 @@ plotter.add_axes(interactive=True)
 
 plotter.subplot(1, 1)
 plotter.add_text('Render Window 3', font_size=30)
-plotter.add_mesh(pyvista.Cone(), color='g', show_edges=True)
+plotter.add_mesh(pv.Cone(), color='g', show_edges=True)
 plotter.show_bounds(all_edges=True)
 
 # Display the window
@@ -44,7 +44,7 @@ plotter.show()
 
 ################################################################################
 
-plotter = pyvista.Plotter(shape=(1, 2))
+plotter = pv.Plotter(shape=(1, 2))
 
 # Note that the (0, 0) location is active by default
 # load and plot an airplane on the left half of the screen
