+++ examples/02-plot/plot-screenshot.py
@@ -3,21 +3,21 @@ Saving Screenshots
 ~~~~~~~~~~~~~~~~~~
 """
 # sphinx_gallery_thumbnail_number = 2
-import pyvista
+import pyvista as pv
 from pyvista import examples
 import numpy as np
 import matplotlib.pyplot as plt
 
 # Get a sample file
 filename = examples.planefile
-mesh = pyvista.read(filename)
+mesh = pv.read(filename)
 
 
 ################################################################################
 # You can also take a screenshot without creating an interactive plot window using
 # the ``Plotter``:
 
-plotter = pyvista.Plotter(off_screen=True)
+plotter = pv.Plotter(off_screen=True)
 plotter.add_mesh(mesh, color='orange')
 plotter.show(auto_close=False)
 # plotter.screenshot('airplane.png')
