### 2019-05-15 ceec922
Make sure first axis matches length when setting arrays
