### 2019-05-15 #1450 @banesullivan
> this might be confusing to other users as well since they'll run the code and get something different from what is "advertised". It might be more transparent to include pv.set_plot_theme('document') in the examples. This way, the user also knows that this is possible.

That's definitely a fair point - We should make sure users will have the same results "as advertised" when running the examples.

As a starting point, I have added a new example in the *Plotting* section showing off all the different themes in https://github.com/pyvista/pyvista/commit/49444c99d507338fed303341324a2a758f2c4e20
