### 2019-05-15 #1450 @Chilipp
Thanks for your quick response @banesullivan! Yes, the changes are sufficient for me.

@leouieda: To me this paper is ready for acceptance
