### 2019-05-16 f7f0190
Fix plot texture example
