### 2019-05-16 0f7fda6
Put the range checker for texture coordinate back
