### 2019-05-16 9b77bba
Update docstring
