### 2019-05-16 cf4e6cd
BUGFIX: RGB plotting bug-fix
