+++ docs/index.rst
@@ -9,7 +9,7 @@
 
     <div class="banner">
         <img src="_static/pyvista_logo.png" alt="pyvista" width="500px">
-        <h2>A Streamlined Python Interface for the Visualization Toolkit</h2>
+        <h2>3D plotting and mesh analysis through a streamlined interface for the Visualization Toolkit (VTK)</h2>
     </div>
 
 
