### 2019-05-16 #1450 @nicoguaro
First, I would like to congratulate the authors for the good work PyVista. I have used VTK directly from Python and this really helps in the process. I have also, used it since it was named vtkInterface and it would say that it is in a mature state now.

Regarding the review:

- I have installed using both ``pip`` and ``conda`` without problems.
- Have checked the examples.
- Have run the tests without problems.
- Built the documentation, and had some (minor) problems when using ``conda``.

The last point is due to some conflicting channels.

Regarding the paper, I have a comment related to the first paragraph and the visualization landscape in Python. The authors do not mention two libraries that I consider that are mature and provide (somewhat) easy to use API, namely:

- [Mayavi](https://arxiv.org/abs/1010.4891v1)
- [yt](https://yt-project.org/)

I would suggest mentioning these in this paragraph.
