### 2019-05-16 #1450 @banesullivan
Thank you @nicoguaro! That is a fair point about including mentions of other Python visualization software by name. I will make updates to address your concerns and have the paper regenerated.

Would you please open an issue on the main repository about the conflicting channels or perhaps on the pyvista-feedstock repo (which ever is more relevant): https://github.com/conda-forge/pyvista-feedstock
