### 2019-05-16 #1450 @nicoguaro
@banesullivan, I did not document what I did. I will repeat it using only ``conda`` and open an issue with the info that I get from there.
