### 2019-05-17 064c85c
Response to https://github.com/openjournals/joss-reviews/issues/1450#issuecomment-493243818
