### 2019-05-17 9e97587
Response to https://github.com/openjournals/joss-reviews/issues/1450#issuecomment-493243818 (#220)
