+++ joss/paper.bib
@@ -145,3 +145,26 @@ issn="0742-4795",
 doi="10.1115/1.4042079",
 url="http://dx.doi.org/10.1115/1.4042079"
 }
+
+
+@article{ytproject,
+    title={yt: A multi-code analysis toolkit for astrophysical simulation data},
+    author={Turk, Matthew J and Smith, Britton D and Oishi, Jeffrey S and Skory, Stephen and Skillman, Samuel W and Abel, Tom and Norman, Michael L},
+    journal={The Astrophysical Journal Supplement Series},
+    volume={192},
+    number={1},
+    pages={9},
+    year={2010},
+    publisher={IOP Publishing}
+}
+
+@article{mayavi,
+    title={Mayavi: 3D visualization of scientific data},
+    author={Ramachandran, Prabhu and Varoquaux, Ga{\"e}l},
+    journal={Computing in Science \& Engineering},
+    volume={13},
+    number={2},
+    pages={40--51},
+    year={2011},
+    publisher={IEEE}
+}
