### 2019-05-17 0688476
Add repeat_texture as keyword argument and RuntimeWarning
