### 2019-05-17 7bbb871
Remove repeat_texture parameter
