### 2019-05-17 2ddde61
Edit example notebooks for use on MyBinder
