+++ docs/conf.py
@@ -107,8 +107,8 @@ sphinx_gallery_conf = {
     "image_scrapers": (pyvista.Scraper(), 'matplotlib'),
     "thumbnail_size": (350, 350),
     'first_notebook_cell': ("%matplotlib inline\n"
-                            "from pyvista import set_plot_theme",
-                            "set_plot_theme('document')",)
+                            "from pyvista import set_plot_theme\n"
+                            "set_plot_theme('document')")
 }
 
 
