### 2019-05-17 #1450 @leouieda
@nicoguaro thanks for the quick review! :+1:  I see you have a couple of items left in your checklist. Are these intentional or forgotten?

@banesullivan I noticed a few of missing DOIs from the paper:

* Matplotlib: https://doi.org/10.1109/MCSE.2007.55
* Mayavi: https://doi.org/10.1109/MCSE.2011.35
* yt: https://doi.org/10.1088/0067-0049/192/1/9

Could you please add these to the bibtex?
