### 2019-05-17 34bcc49
Update paper.bib with DOIs and ISBNs per https://github.com/openjournals/joss-reviews/issues/1450#issuecomment-493608701
