+++ joss/paper.bib
@@ -1,15 +1,17 @@
 @book{vtkbook,
-    title={The visualization toolkit: an object-oriented approach to 3D graphics},
+    title={The Visualization Toolkit (4th ed.)},
     author={Schroeder, Will J and Lorensen, Bill and Martin, Ken},
-    year={2004},
-    publisher={Kitware}
+    year={2006},
+    publisher={Kitware},
+    isbn={978-1-930934-19-1}
 }
 
 @misc{numpy,
     title={Numerical Python},
     author={Ascher, David and Dubois, Paul F and Hinsen, Konrad and Hugunin, Jim and Oliphant, Travis and others},
     year={2001},
-    publisher={Citeseer}
+    publisher={Citeseer},
+    doi={10.1109/MCSE.2011.37}
 }
 
 @article{matplotlib,
@@ -20,7 +22,8 @@
     number={3},
     pages={90},
     year={2007},
-    publisher={IEEE Computer Society}
+    publisher={IEEE Computer Society},
+    doi={10.1109/MCSE.2007.55}
 }
 
 @article{edl,
@@ -33,117 +36,117 @@
 
 
 @INPROCEEDINGS{Gillaugh2017,
-author="Gillaugh, Daniel
-and Kaszynski, Alexander
-and Brown, Jeffrey M.
-and Johnston, David A.
-and Slater, Joseph C.",
-title="Accurate Strain Gage Limits Through Geometry Mistuning Modeling",
-series="AIAA SciTech Forum",
-year="2017",
-month="Jan",
-day="05",
-publisher="American Institute of Aeronautics and Astronautics",
-doi="10.2514/6.2017-0865",
-url="https://doi.org/10.2514/6.2017-0865"
+    author="Gillaugh, Daniel
+    and Kaszynski, Alexander
+    and Brown, Jeffrey M.
+    and Johnston, David A.
+    and Slater, Joseph C.",
+    title="Accurate Strain Gage Limits Through Geometry Mistuning Modeling",
+    series="AIAA SciTech Forum",
+    year="2017",
+    month="Jan",
+    day="05",
+    publisher="American Institute of Aeronautics and Astronautics",
+    doi="10.2514/6.2017-0865",
+    url="https://doi.org/10.2514/6.2017-0865"
 }
 
 @Article{Brown2018,
-author="Brown, Jeffrey M.
-and Beck, Joseph
-and Kaszynski, Alexander
-and Clark, John",
-title="Surrogate Modeling of Manufacturing Variation Effects on Unsteady Interactions in a Transonic Turbine",
-journal="Journal of Engineering for Gas Turbines and Power",
-year="2018",
-month="Oct",
-day="11",
-publisher="ASME",
-volume="141",
-number="3",
-pages="032506--032506-12",
-issn="0742-4795",
-doi="10.1115/1.4041314",
-url="http://dx.doi.org/10.1115/1.4041314"
+    author="Brown, Jeffrey M.
+    and Beck, Joseph
+    and Kaszynski, Alexander
+    and Clark, John",
+    title="Surrogate Modeling of Manufacturing Variation Effects on Unsteady Interactions in a Transonic Turbine",
+    journal="Journal of Engineering for Gas Turbines and Power",
+    year="2018",
+    month="Oct",
+    day="11",
+    publisher="ASME",
+    volume="141",
+    number="3",
+    pages="032506--032506-12",
+    issn="0742-4795",
+    doi="10.1115/1.4041314",
+    url="http://dx.doi.org/10.1115/1.4041314"
 }
 
 @INPROCEEDINGS{Brown2018_pro,
-author="Brown, Jeffrey M.
-and Beck, Joseph
[…]
