### 2019-05-17 #1450 @banesullivan
@leouieda, thank you for pointing out the missing DOIs. I have updated the `paper.bib` to have DOIs for all references.

One exception being the VTK book - I cannot find a DOI for this but I added an ISBN which `@whedon` doesn't seem to render in the references section of the paper. I filled out our citation for the VTK book per how VTK's website requests (https://vtk.org/about/#citation):

> To cite VTK, please reference the VTK textbook.
>
> Schroeder, Will; Martin, Ken; Lorensen, Bill (2006), The Visualization Toolkit (4th ed.), Kitware, ISBN 978-1-930934-19-1
