### 2019-05-17 pyvista/pyvista#221 @leouieda
Opening sentence in the JOSS paper
Hi @banesullivan, could you please rephrase the opening sentence in the JOSS paper?

> The availability of software for 3D visualization in Python is limited to
libraries that are either not capable of handling large, spatially referenced
datasets, e.g. `matplotlib` [@matplotlib], or libraries including Mayavi
[@mayavi], the yt Project [@ytproject], and the Visualization Toolkit (VTK)
[@vtkbook] which are powerful yet have inherently complex application
programming interfaces (APIs) that create a barrier to entry for new users.

The reason is that this might come across as too harsh on these packages. For example, I have never used `yt` but I don't their API is as complex as VTK. And there are other options out there as well, like the [pythreejs](https://github.com/jupyter-widgets/pythreejs) and [ipyvolume](https://github.com/maartenbreddels/ipyvolume). So maybe you can phrase without having to list all current options, like "There are several options for 3D visualization in Python, like matplotlib (...), Mayavi (...), .... However, few are capable of handling large spatially referenced datasets and some are powerful yet have inherently complex APIs, which might create a barrier to entry for new users."

openjournals/joss-reviews#1450
