### 2019-05-17 2d9a61e
Fix typos and links in the JOSS paper
The links are made to be visible in the PDF so that they are printer friendly and can be copy-pasted.
Removed the link to examples in the VTK wiki. It's not crucial for the paper and not very future proof.

openjournals/joss-reviews#1450
