+++ joss/paper.md
@@ -32,8 +32,8 @@ programming interfaces (APIs) that create a barrier to entry for new users.
 Notably, VTK is a powerful scientific visualization software library, and with
 Python bindings, it combines the speed of C++ with the rapid prototyping of
 Python.
-Despite this, VTK code programmed in Python using the base VTK python package
-is unnecessarily complicated as its API merely binds existing C++ calls.
+Despite this, VTK code programmed in Python using the base VTK Python package
+is unnecessarily complicated as its API binds existing C++ calls.
 The PyVista Python package provides a concise, well-documented interface
 exposing VTK’s powerful visualization backend; enabling researchers to
 rapidly explore large datasets, communicate their spatial findings, and
@@ -41,7 +41,8 @@ facilitate reproducibility. PyVista further seeks to simplify standard mesh
 creation and plotting routines without compromising on the speed of the C++
 VTK backend. Figure 1 demonstrates an integrated scene of geospatial data
 generated by PyVista; to learn more about how this visualization was created,
-please visit [PVGeo's FORGE project website](http://forge.pvgeo.org).
+please visit PVGeo's FORGE project website 
+([forge.pvgeo.org](http://forge.pvgeo.org)).
 
 
 ![A visually integrated scene of geospatial data (FORGE Geothermal Site).
@@ -56,10 +57,9 @@ visualization library is often an ambitious programming endeavor.
 Reading a VTK supported file and plotting it requires a user to write a
 complicated sequence of routines to render the data object while
 having to remember which VTK classes to use for file reading and dataset mapping.
-An example can be found in [this creative commons VTK example](https://vtk.org/Wiki/VTK/Examples/Python/STLReader).
 PyVista includes plotting routines that are intended to be intuitive and
-highly controllable with `matplotlib` [@matplotlib] similar syntax and keyword
-arguments. These plotting routines are defined to make the process of
+highly controllable with syntax and keyword arguments similar to `matplotlib` 
+[@matplotlib]. These plotting routines are defined to make the process of
 visualizing spatially referenced data straightforward and easily implemented
 by novice programmers; streamlining the process for creating integrated scenes
 like that shown in Figure 1.
@@ -104,10 +104,10 @@ AFRL publications leveraging PyVista for 3D visualization include:
 [@Gillaugh2017], [@Brown2018], [@Brown2018_pro], [@Beck2018_subspace],
 [@Kaszynski2018], [@Gillaugh2018_art], [@Gillaugh2019]
 
-[PVGeo](https://github.com/OpenGeoVis/PVGeo) is Python package of VTK-based
-algorithms to analyze geoscientific data and models. PyVista is used to make
-the inputs and outputs of PVGeo's algorithms more accessible and to streamline
-the process of visualizing geoscientific data.
+PVGeo ([github.com/OpenGeoVis/PVGeo](https://github.com/OpenGeoVis/PVGeo)) is
+a Python package of VTK-based algorithms to analyze geoscientific data and models. 
+PyVista is used to make the inputs and outputs of PVGeo's algorithms more 
+accessible and to streamline the process of visualizing geoscientific data.
 
 
 ## References
