+++ joss/paper.md
@@ -42,7 +42,7 @@ facilitate reproducibility. PyVista further seeks to simplify standard mesh
 creation and plotting routines without compromising on the speed of the C++
 VTK backend. Figure 1 demonstrates an integrated scene of geospatial data
 generated by PyVista; to learn more about how this visualization was created,
-please visit PVGeo's FORGE project website 
+please visit PVGeo's FORGE project website
 ([forge.pvgeo.org](http://forge.pvgeo.org)).
 
 
@@ -59,7 +59,7 @@ Reading a VTK supported file and plotting it requires a user to write a
 complicated sequence of routines to render the data object while
 having to remember which VTK classes to use for file reading and dataset mapping.
 PyVista includes plotting routines that are intended to be intuitive and
-highly controllable with syntax and keyword arguments similar to `matplotlib` 
+highly controllable with syntax and keyword arguments similar to Matplotlib
 [@matplotlib]. These plotting routines are defined to make the process of
 visualizing spatially referenced data straightforward and easily implemented
 by novice programmers; streamlining the process for creating integrated scenes
@@ -106,8 +106,8 @@ AFRL publications leveraging PyVista for 3D visualization include:
 [@Kaszynski2018], [@Gillaugh2018_art], [@Gillaugh2019]
 
 PVGeo ([github.com/OpenGeoVis/PVGeo](https://github.com/OpenGeoVis/PVGeo)) is
-a Python package of VTK-based algorithms to analyze geoscientific data and models. 
-PyVista is used to make the inputs and outputs of PVGeo's algorithms more 
+a Python package of VTK-based algorithms to analyze geoscientific data and models.
+PyVista is used to make the inputs and outputs of PVGeo's algorithms more
 accessible and to streamline the process of visualizing geoscientific data.
 
 
