### 2019-05-17 e782d7c
JOSS Paper updates (#223)
* Address #221

* `matplotlib` -> Matplotlib
