+++ joss/paper.md
@@ -23,12 +23,13 @@ bibliography: paper.bib
 
 # Summary
 
-The availability of software for 3D visualization in Python is limited to
-libraries that are either not capable of handling large, spatially referenced
-datasets, e.g. `matplotlib` [@matplotlib], or libraries including Mayavi
-[@mayavi], the yt Project [@ytproject], and the Visualization Toolkit (VTK)
-[@vtkbook] which are powerful yet have inherently complex application
-programming interfaces (APIs) that create a barrier to entry for new users.
+There are several options for 3D visualization in Python, a few notable projects
+include Matplotlib [@matplotlib], Mayavi [@mayavi], the yt Project [@ytproject],
+and the Visualization Toolkit (VTK) [@vtkbook].
+However, few open-source packages are capable of handling large,
+spatially-referenced datasets and some are powerful yet have inherently complex
+application programming interfaces (APIs), which might create a barrier to entry
+for new users.
 Notably, VTK is a powerful scientific visualization software library, and with
 Python bindings, it combines the speed of C++ with the rapid prototyping of
 Python.
@@ -41,7 +42,7 @@ facilitate reproducibility. PyVista further seeks to simplify standard mesh
 creation and plotting routines without compromising on the speed of the C++
 VTK backend. Figure 1 demonstrates an integrated scene of geospatial data
 generated by PyVista; to learn more about how this visualization was created,
-please visit PVGeo's FORGE project website 
+please visit PVGeo's FORGE project website
 ([forge.pvgeo.org](http://forge.pvgeo.org)).
 
 
@@ -58,7 +59,7 @@ Reading a VTK supported file and plotting it requires a user to write a
 complicated sequence of routines to render the data object while
 having to remember which VTK classes to use for file reading and dataset mapping.
 PyVista includes plotting routines that are intended to be intuitive and
-highly controllable with syntax and keyword arguments similar to `matplotlib` 
+highly controllable with syntax and keyword arguments similar to Matplotlib
 [@matplotlib]. These plotting routines are defined to make the process of
 visualizing spatially referenced data straightforward and easily implemented
 by novice programmers; streamlining the process for creating integrated scenes
@@ -105,8 +106,8 @@ AFRL publications leveraging PyVista for 3D visualization include:
 [@Kaszynski2018], [@Gillaugh2018_art], [@Gillaugh2019]
 
 PVGeo ([github.com/OpenGeoVis/PVGeo](https://github.com/OpenGeoVis/PVGeo)) is
-a Python package of VTK-based algorithms to analyze geoscientific data and models. 
-PyVista is used to make the inputs and outputs of PVGeo's algorithms more 
+a Python package of VTK-based algorithms to analyze geoscientific data and models.
+PyVista is used to make the inputs and outputs of PVGeo's algorithms more
 accessible and to streamline the process of visualizing geoscientific data.
 
 
