### 2019-05-17 pyvista/pyvista#221 @banesullivan
@leouieda, I have made changes to address this in #223 - could you let me know if the changes are okay?
