+++ joss/paper.md
@@ -40,17 +40,7 @@ exposing VTK’s powerful visualization backend; enabling researchers to
 rapidly explore large datasets, communicate their spatial findings, and
 facilitate reproducibility. PyVista further seeks to simplify standard mesh
 creation and plotting routines without compromising on the speed of the C++
-VTK backend. Figure 1 demonstrates an integrated scene of geospatial data
-generated by PyVista; to learn more about how this visualization was created,
-please visit PVGeo's FORGE project website
-([forge.pvgeo.org](http://forge.pvgeo.org)).
-
-
-![A visually integrated scene of geospatial data (FORGE Geothermal Site).
-This rendering includes a digital land surface with overlain satellite
-imagery and geologic map, a subsurface temperature model, scattered points
-of the sampled temperature values, geophysical well logging data, GIS site
-boundary, and interpreted faulting surfaces.](./images/forge-iso.png)
+VTK backend.
 
 
 Plotting VTK datasets using only the VTK Python package or a similar
@@ -78,13 +68,6 @@ that algorithm directly onto all PyVista datasets, providing a shared set of
 functionality. Through the use of these bound filtering methods, powerful VTK
 algorithms can be leveraged and controlled via keyword arguments designed to
 be intuitive for novice users.
-Figure 2 show examples of standard filtering algorithms.
-
-
-![Examples of common filtering algorithms: A) threshold volume
-extraction by scalar array, B) iso-contouring by scalar array, C) orthogonal
-slicing through volume, and D) geometric glyphing at mesh nodes and scaled by
-a scalar array.](./images/filters.png)
 
 
 At its core, PyVista is a pure Python helper module for VTK
@@ -94,6 +77,19 @@ and direct array access adhering to VTK's object-oriented approach to
 The PyVista Python package provides an accessible and intuitive interface back
 to the VTK library to facilitate rapid prototyping, analysis, and visual
 integration of spatially referenced datasets.
+
+Figure 1 demonstrates an integrated scene of geospatial data
+generated by PyVista; to learn more about how this visualization was created,
+please visit PVGeo's FORGE project website
+([forge.pvgeo.org](http://forge.pvgeo.org)).
+
+
+![A visually integrated scene of geospatial data (FORGE Geothermal Site).
+This rendering includes a digital land surface with overlain satellite
+imagery and geologic map, a subsurface temperature model, scattered points
+of the sampled temperature values, geophysical well logging data, GIS site
+boundary, and interpreted faulting surfaces.](./images/forge-iso.png)
+
 
 ## Mentions
 
