### 2019-05-17 #1450 @banesullivan
Looks like we got it!

@leouieda, can you inspect the last preview of the PDF? The figure looks much better than previously

I'll get a PR going to bring these changes into the master branch
