### 2019-05-17 0d0c68f
Reduce size of FORGE image
