### 2019-05-17 #1450 @leouieda
@banesullivan looks good now :+1: After you merge these changes, please do the following:

* Make a new release of Pyvista with the modifications made during review
* Archive this new version in Zenodo or figshare and post the DOI here
