### 2019-05-18 f9a37bb
Refactor .plot() to .show() on Plotter examples
