+++ tests/test_plotting.py
@@ -87,7 +87,7 @@ def test_set_camera_position():
     plotter.camera_position = 'zx'
     plotter.camera_position = 'zy'
     plotter.camera_position = cpos
-    cpos_out = plotter.plot()
+    cpos_out = plotter.show()
     assert cpos_out == cpos
 
 
@@ -120,14 +120,14 @@ def test_plot_show_bounds():
                             show_ylabels=False,
                             show_zlabels=False,
                             use_2d=True)
-    plotter.plot()
+    plotter.show()
 
 @pytest.mark.skipif(NO_PLOTTING, reason="Requires system to support plotting")
 def test_plot_label_fmt():
     plotter = pyvista.Plotter(off_screen=OFF_SCREEN)
     plotter.add_mesh(sphere)
     plotter.show_bounds(xlabel='My X', fmt=r'%.3f')
-    plotter.plot()
+    plotter.show()
 
 
 @pytest.mark.skipif(NO_PLOTTING, reason="Requires system to support plotting")
@@ -205,7 +205,7 @@ def test_make_movie():
     plotter.remove_actor(actor)
     plotter.add_mesh(movie_sphere,
                      scalars=np.random.random(movie_sphere.n_faces))
-    plotter.plot(auto_close=False, window_size=[304, 304])
+    plotter.show(auto_close=False, window_size=[304, 304])
     plotter.set_focus([0, 0, 0])
     for i in range(10):
         plotter.write_frame()
@@ -237,7 +237,7 @@ def test_add_legend():
     legend_labels = [['sphere', 'r']]
     plotter.add_legend(labels=legend_labels, border=True, bcolor=None,
                        size=[0.1, 0.1])
-    plotter.plot()
+    plotter.show()
 
 
 @pytest.mark.skipif(NO_PLOTTING, reason="Requires system to support plotting")
@@ -259,7 +259,7 @@ def test_add_point_labels():
     plotter.set_background('k')
     plotter.add_point_labels(points, range(n), show_points=True, point_color='r')
     plotter.add_point_labels(points - 1, range(n), show_points=False, point_color='r')
-    plotter.plot()
+    plotter.show()
 
 
 @pytest.mark.skipif(NO_PLOTTING, reason="Requires system to support plotting")
@@ -268,7 +268,7 @@ def test_add_points():
     plotter = pyvista.Plotter(off_screen=OFF_SCREEN)
     points = np.random.random((n, 3))
     plotter.add_points(points, scalars=np.arange(10), cmap=None, flip_scalars=True)
-    plotter.plot()
+    plotter.show()
 
 
 @pytest.mark.skipif(NO_PLOTTING, reason="Requires system to support plotting")
@@ -296,7 +296,7 @@ def test_plot_cell_scalars():
     scalars = np.arange(sphere.n_faces)
     plotter.add_mesh(sphere, interpolate_before_map=True, scalars=scalars,
                      n_colors=5, rng=10)
-    plotter.plot()
+    plotter.show()
 
 @pytest.mark.skipif(NO_PLOTTING, reason="Requires system to support plotting")
 def test_plot_clim():
@@ -304,7 +304,7 @@ def test_plot_clim():
     scalars = np.arange(sphere.n_faces)
     plotter.add_mesh(sphere, interpolate_before_map=True, scalars=scalars,
                      n_colors=5, clim=10)
-    plotter.plot()
+    plotter.show()
     assert plotter.mapper.GetScalarRange() == (-10, 10)
 
 
@@ -313,7 +313,7 @@ def test_invalid_n_scalars():
[…]
