+++ docs/getting-started/simple.rst
@@ -155,7 +155,7 @@ manually interact with the plotting window:
     plotter = pv.Plotter(off_screen=True)
     plotter.add_mesh(data, color='tan')
     plotter.camera_position = cpos
-    plotter.plot(auto_close=False)
+    plotter.show(auto_close=False)
     # plotter.screenshot('airplane.png')
     plotter.close()
 
