### 2019-05-18 ecb59e3
Fix plot texture example (#216)
* Fix plot texture example

* Put the range checker for texture coordinate back

* Update docstring

* Add repeat_texture as keyword argument and RuntimeWarning

* Remove repeat_texture parameter
