### 2019-05-18 1405583
Fix texture tests
