+++ tests/test_common.py
@@ -362,11 +362,6 @@ def test_set_t_coords():
     with pytest.raises(AssertionError):
         grid.t_coords = np.empty((grid.n_points, 1))
 
-    with pytest.warns(RuntimeWarning):
-        arr = np.empty((grid.n_points, 2))
-        arr[:] = -1
-        grid.t_coords = arr
-
 
 def test_activate_texture_none():
     assert grid._activate_texture('not a key') is None
