### 2019-05-18 #1450 @banesullivan
@leouieda, I have tagged a new release and the CIs are working to publish everything!

Version `0.20.2` is released, being pushed to PyPI, conda-forge, and live on Zenodo.

The Zenodo DOI that always resolves to the latest version is: `10.5281/zenodo.2647611`  (This DOI represents all versions, and will always resolve to the latest one)  Zenodo is still pushing this to DOI.org:

- https://doi.org/10.5281/zenodo.2647611
