### 2019-05-18 #1450 @leouieda
OK @banesullivan, given that both reviewers have signed off on this submission, I think we are ready to accept! :confetti_ball:  A minor note: you might want to include ORCIDs of authors (specially yours) in the Zenodo archive.

@nicoguaro and @Chilipp many thanks for the speedy and constructive review! We really appreciate that you donate some of your time to JOSS.
