### 2019-05-18 #1450 @leouieda
@openjournals/joss-eics this submission is ready for publication. I checked the generated PDF and XML and they look good to me.
