### 2019-05-18 #1450 @banesullivan
> A minor note: you might want to include ORCIDs of authors (specially yours) in the Zenodo archive.

Thanks for pointing this out @leouieda! I didn't realize our ORCIDs were missing (or that we could do this...) I'm struggling to find how/where this goes, do you have any pointers?
