### 2019-05-18 #1450 @leouieda
> I'm struggling to find how/where this goes, do you have any pointers?

When editing the metadata for authors, you can specify name, affiliation, and ORCID for each one.
