### 2019-05-18 #1450 @banesullivan
> When editing the metadata for authors, you can specify name, affiliation, and ORCID for each one.

Ah, thanks! I didn't realize I'd have to edit the archive

Relevant to https://github.com/zenodo/zenodo/issues/596
