### 2019-05-18 47a50fa
Add Slack link to Getting Connected section in docs
