+++ CONTRIBUTING.md
@@ -27,7 +27,7 @@ python -m pip install -e .
 For general questions about the project, its applications, or about software
 usage, please create an issue in the [pyvista/pyvista-support](https://github.com/pyvista/pyvista-support)
 repository where the community can collectively address your questions.
-You are also welcome to join us on join us on [Slack](http://slack.opengeovis.org)
+You are also welcome to join us on join us on [Slack](http://slack.pyvista.org)
 or send one of the developers an email.
 The project support team can be reached at [info@pyvista.org](mailto:info@pyvista.org)
 
