### 2019-05-19 cb91b5e
Fix typos in docs
