### 2019-05-19 9b66434
Add PYVISTA_USE_PANEL to server set details
