+++ docs/getting-started/installation.rst
@@ -143,7 +143,8 @@ time your Docker container is launched:
     #!/bin/bash
     set -x
     export DISPLAY=:99.0
-    export PYVISTA_OFF_SCREEN=True
+    export PYVISTA_OFF_SCREEN=true
+    export PYVISTA_USE_PANEL=true
     which Xvfb
     Xvfb :99 -screen 0 1024x768x24 > /dev/null 2>&1 &
     sleep 3
@@ -190,7 +191,8 @@ Then, configure the headless display:
 
     sudo apt-get install xvfb
     export DISPLAY=:99.0
-    export PYVISTA_OFF_SCREEN=True
+    export PYVISTA_OFF_SCREEN=true
+    export PYVISTA_USE_PANEL=true
     Xvfb :99 -screen 0 1024x768x24 > /dev/null 2>&1 &
     sleep 3
 
