+++ docs/tools/geometric.rst
@@ -16,3 +16,4 @@ functions create, check out the example: :ref:`ref_geometric_example`.
 .. autofunction:: pyvista.Polygon
 .. autofunction:: pyvista.Disc
 .. autofunction:: pyvista.Text3D
+.. autofunction:: pyvista.SuperToroid
