### 2019-05-20 fa89777
Add SG contractual method per https://github.com/sphinx-gallery/sphinx-gallery/pull/494
