### 2019-05-20 0a814e8
Make _get_sg_image_scraper accesible
