### 2019-05-20 d474c94
Provide test for _get_sg_image_scraper per https://github.com/sphinx-gallery/sphinx-gallery/pull/494#issuecomment-494097672
