### 2019-05-20 b9953ee
Test for https://github.com/sphinx-gallery/sphinx-gallery/pull/494
