+++ requirements_docs.txt
@@ -6,4 +6,4 @@ sphinxcontrib-websupport
 pytest-sphinx
 git+git://github.com/rtfd/sphinx-notfound-page@master
 sphinx-copybutton
-sphinx-gallery
+git+git://github.com/larsoner/sphinx-gallery@scrape
