### 2019-05-20 f468607
Test string get for https://github.com/sphinx-gallery/sphinx-gallery/pull/494
