### 2019-05-20 #1450 @banesullivan
Thanks @leouieda!!! This is very exciting!

@leouieda, would you please confirm that the following would be a correct BibTex citation for this paper:

```txt
@article{sullivan2019pyvista,
  author = {Sullivan, C. Bane and Kaszynski, Alexander A.},
  journal = {Journal of Open Source Software},
  pages = {4(37), 1450},
  title = {{PyVista: 3D plotting and mesh analysis through a streamlined interface for the Visualization Toolkit (VTK)}},
  year = {2019},
  doi = {10.21105/joss.01450},
}
```

Also, is there a way to autogenerate BibTex citations on the JOSS website? I see the current autogenerated citation of:

> Sullivan et al., (2019). PyVista: 3D plotting and mesh analysis through a streamlined interface for the Visualization Toolkit (VTK). Journal of Open Source Software, 4(37), 1450, https://doi.org/10.21105/joss.01450
