### 2019-05-20 ed3b124
Add citation after https://github.com/openjournals/joss-reviews/issues/1450
