+++ docs/index.rst
@@ -125,7 +125,14 @@ Here are a few open-source projects that leverage PyVista:
 .. _tetgen: https://github.com/akaszynski/tetgen
 
 
+Citing PyVista
+==============
 
+There is a `paper about PyVista <https://doi.org/10.21105/joss.01450>`_!
+
+If you are using PyVista in your scientific research, please help our scientific
+visibility by citing our work! Head over to :ref:`citation_ref` to learn more
+about citing PyVista.
 
 
 Getting Started
