+++ CITATION.rst
@@ -1,0 +1,26 @@
+.. _citation_ref:
+
+Citing PyVista
+--------------
+
+There is a `paper about PyVista <https://doi.org/10.21105/joss.01450>`_!
+
+If you are using PyVista in your scientific research, please help our scientific
+visibility by citing our work!
+
+
+    Sullivan et al., (2019). PyVista: 3D plotting and mesh analysis through a streamlined interface for the Visualization Toolkit (VTK). Journal of Open Source Software, 4(37), 1450, https://doi.org/10.21105/joss.01450
+
+
+BibTex:
+
+.. code::
+
+    @article{sullivan2019pyvista,
+      author = {Sullivan, C. Bane and Kaszynski, Alexander A.},
+      journal = {Journal of Open Source Software},
+      pages = {4(37), 1450},
+      title = {{PyVista: 3D plotting and mesh analysis through a streamlined interface for the Visualization Toolkit (VTK)}},
+      year = {2019},
+      doi = {10.21105/joss.01450},
+    }
