+++ AUTHORS.rst
@@ -1,3 +1,12 @@
+.. _authors_ref:
+
+Authors
+-------
+
+.. image:: https://img.shields.io/github/contributors/pyvista/pyvista.svg?logo=github&logoColor=white
+   :target: https://github.com/pyvista/pyvista/graphs/contributors/
+
+
 The following is a list of authors who have made substantial contributions to
 the conception or design of this software; or the creation of new code used in
 this software; or have drafted the work or substantively revised it and are
