+++ docs/authors.rst
@@ -1,10 +1,15 @@
-Authors
-=======
+Authors & Citation
+==================
 
-.. image:: https://img.shields.io/github/contributors/pyvista/pyvista.svg?logo=github&logoColor=white
-   :target: https://github.com/pyvista/pyvista/graphs/contributors/
+PyVista is an open-source project with a community of contributors.
+While most of PyVista's contributors are listed in the :ref:`authors_ref`
+section, please be sure to head over to the `Contributors Graph`_.
 
-.. _authors_ref:
+.. _Contributors Graph: https://github.com/pyvista/pyvista/graphs/contributors/
+
+
+.. include:: ../CITATION.rst
+
 
 .. include:: ../AUTHORS.rst
 
