### 2019-05-20 70ae792
Add paper link to Getting Connected section
