### 2019-05-22 #1450 @leouieda
Hi @banesullivan I think the bibtex should be:

```
@article{Sullivan2019,
  doi = {10.21105/joss.01450},
  url = {https://doi.org/10.21105/joss.01450},
  year = {2019},
  month = may,
  publisher = {The Open Journal},
  volume = {4},
  number = {37},
  pages = {1450},
  author = {C. Sullivan and Alexander Kaszynski},
  title = {{PyVista}: 3D plotting and mesh analysis through a streamlined interface for the Visualization Toolkit ({VTK})},
  journal = {Journal of Open Source Software}
}
```

At least according to https://www.doi2bib.org/bib/10.21105/joss.01450
