### 2019-05-22 14bbec7
First try at smooth shading
