### 2019-05-22 c2be9b3
Control smooth shading with a parameter
