### 2019-05-22 4d11545
Revert bad changes for ffmpeg. see #215
