+++ appveyor.yml
@@ -31,6 +31,7 @@ install:
   # Install requirements
   - "%PYTHON% -m ensurepip"
   - "%PYTHON% -m pip install -r requirements.txt"
+  - "%PYTHON% -c \"import imageio; imageio.plugins.ffmpeg.download()\""
 
 build_script:
   - "%PYTHON% setup.py sdist"
