### 2019-05-22 #1450 @banesullivan
@leouieda - we have a slight issue. The publication is showing up under `vtki` still. On the JOSS homepage:

But after clicking on it, all is good:
