### 2019-05-23 be7b956
Precompute point normals for smoothing
