+++ pyvista/plotting.py
@@ -221,12 +221,21 @@ def plot(var_item, off_screen=None, full_screen=False, screenshot=None,
                 plotter.add_arrows(var_item[0], var_item[1])
             else:
                 for item in var_item:
-                    plotter.add_mesh(item, **kwargs)
+                    if isinstance(item, np.ndarray) and item.ndim == 3:
+                        plotter.add_volume(item, **kwargs)
+                    else:
+                        plotter.add_mesh(item, **kwargs)
         else:
             for item in var_item:
-                plotter.add_mesh(item, **kwargs)
+                if isinstance(item, np.ndarray) and item.ndim == 3:
+                    plotter.add_volume(item, **kwargs)
+                else:
+                    plotter.add_mesh(item, **kwargs)
     else:
-        plotter.add_mesh(var_item, **kwargs)
+        if isinstance(var_item, np.ndarray) and var_item.ndim == 3:
+            plotter.add_volume(var_item, **kwargs)
+        else:
+            plotter.add_mesh(var_item, **kwargs)
 
     if text:
         plotter.add_text(text)
@@ -1095,6 +1104,153 @@ class BasePlotter(object):
         # Add scalar bar if available
         if stitle is not None and show_scalar_bar and not rgb:
             self.add_scalar_bar(stitle, **scalar_bar_args)
+
+        return actor
+
+    def add_volume(self, data, resolution=[1,1,1], opacity='linear', n_colors=256,
+                   cmap=None, flip_scalars=False, reset_camera=None, name=None,
+                   ambient=0.0, categories=False, loc=None, backface_culling=False,
+                   **kwargs):
+        """
+        Adds a volume, rendered using a GPU ray cast mapper.
+
+        Requires a 3D numpy.ndarray.
+
+        Parameters
+        ----------
+        data: 3D numpy.ndarray
+            The input array to visualize, assuming the array values denote
+            scalar intensities.
+
+        opacity : float or string, optional
+            Opacity of input array. Options are: linear, linear_r, geom, geom_r.
+            Defaults to 'linear'. Can also be given as a scalar value between 0
+            and 1.
+
+        flip_scalars : bool, optional
+            Flip direction of cmap.
+
+        n_colors : int, optional
+            Number of colors to use when displaying scalars.  Default
+            256.
+
+        cmap : str, optional
+           cmap string.  See available matplotlib cmaps. Only applicable for when
+           displaying scalars.  Defaults to None (jet).  Requires matplotlib.
+
+        name : str, optional
+            The name for the added actor so that it can be easily
+            updated.  If an actor of this name already exists in the
+            rendering window, it will be replaced by the new actor.
+
+        ambient : float, optional
+            The amount of light from 0 to 1 that reaches the actor when not
+            directed at the light source emitted from the viewer.  Default 0.0.
+
+        loc : int, tuple, or list
+            Index of the renderer to add the actor to.  For example,
+            ``loc=2`` or ``loc=(1, 1)``.  If None, selects the last
+            active Renderer.
+
+        backface_culling : bool optional
+            Does not render faces that should not be visible to the
+            plotter.  This can be helpful for dense surface meshes,
+            especially when edges are visible, but can cause flat
+            meshes to be partially displayed.  Default False.
+
+        categories : bool, optional
+            If fetching a colormap from matplotlib, this is the number of
+            categories to use in that colormap. If set to ``True``, then
+            the number of unique values in the scalar array will be used.
+
+        Returns
+        -------
+        actor: vtk.vtkVolume
[…]
