### 2019-05-23 c0093db
added convenience function for clearing scalar arrays
