+++ tests/test_plotting.py
@@ -13,6 +13,16 @@ from pyvista.plotting import system_supports_plotting
 
 NO_PLOTTING = not system_supports_plotting()
 
+ffmpeg_failed = False
+try:
+    try:
+        import imageio_ffmpeg
+    except ImportError:
+        import imageio
+        imageio.plugins.ffmpeg.download()
+except:
+    ffmpeg_failed = True
+
 
 if __name__ != '__main__':
     OFF_SCREEN = 'pytest' in sys.modules
@@ -193,6 +203,7 @@ def test_open_gif_invalid():
         plotter.open_gif('file.abs')
 
 
+@pytest.mark.skipif(ffmpeg_failed, reason="Requires imageio-ffmpeg")
 @pytest.mark.skipif(NO_PLOTTING, reason="Requires system to support plotting")
 def test_make_movie():
     # Make temporary file
