+++ tests/test_plotting.py
@@ -17,6 +17,7 @@ ffmpeg_failed = False
 try:
     try:
         import imageio_ffmpeg
+        imageio_ffmpeg.get_ffmpeg_exe()
     except ImportError:
         import imageio
         imageio.plugins.ffmpeg.download()
