### 2019-05-25 85b7a14
clear_scalars -> clear_arrays
