### 2019-05-25 e87350d
Make clear methods for all scalar array types and test
