+++ tests/test_common.py
@@ -7,13 +7,14 @@ import vtk
 import pyvista
 from pyvista import examples
 
-grid = pyvista.UnstructuredGrid(examples.hexbeamfile)
+GRID = pyvista.UnstructuredGrid(examples.hexbeamfile)
 
 py2 = sys.version_info.major == 2
 
 
 def test_point_arrays():
     key = 'test_array_points'
+    grid = GRID.copy()
     grid[key] = np.arange(grid.n_points)
     assert key in grid.point_arrays
 
@@ -29,8 +30,12 @@ def test_point_arrays():
 
     assert np.allclose(grid[key], np.arange(grid.n_points))
 
+    grid.clear_point_arrays()
+    assert len(grid.point_arrays.keys()) == 0
+
 
 def test_point_arrays_bad_value():
+    grid = GRID.copy()
     with pytest.raises(TypeError):
         grid.point_arrays['new_array'] = None
 
@@ -40,6 +45,7 @@ def test_point_arrays_bad_value():
 
 def test_cell_arrays():
     key = 'test_array_cells'
+    grid = GRID.copy()
     grid[key] = np.arange(grid.n_cells)
     assert key in grid.cell_arrays
 
@@ -57,6 +63,7 @@ def test_cell_arrays():
 
 
 def test_cell_arrays_bad_value():
+    grid = GRID.copy()
     with pytest.raises(TypeError):
         grid.cell_arrays['new_array'] = None
 
@@ -66,6 +73,7 @@ def test_cell_arrays_bad_value():
 
 def test_cell_arrays():
     key = 'test_array_field'
+    grid = GRID.copy()
     # Add array of lenght not equal to n_cells or n_points
     n = grid.n_cells // 3
     grid.field_arrays[key] = np.arange(n)
@@ -81,11 +89,13 @@ def test_cell_arrays():
 
 
 def test_cell_arrays_bad_value():
+    grid = GRID.copy()
     with pytest.raises(TypeError):
         grid.field_arrays['new_array'] = None
 
 
 def test_copy():
+    grid = GRID.copy()
     grid_copy = grid.copy(deep=True)
     grid_copy.points[0] = np.nan
     assert not np.any(np.isnan(grid.points[0]))
@@ -96,6 +106,7 @@ def test_copy():
 
 
 def test_transform():
+    grid = GRID.copy()
     trans = vtk.vtkTransform()
     trans.RotateX(30)
     trans.RotateY(30)
@@ -114,6 +125,7 @@ def test_transform():
 
 
 def test_transform_errors():
+    grid = GRID.copy()
     with pytest.raises(TypeError):
         grid.transform(None)
 
@@ -122,6 +134,7 @@ def test_transform_errors():
 
 
 def test_translate():
+    grid = GRID.copy()
     grid_copy = grid.copy()
     xyz = [1, 1, 1]
     grid_copy.translate(xyz)
@@ -131,6 +144,7 @@ def test_translate():
 
 
 def test_rotate_x():
+    grid = GRID.copy()
     angle = 30
     trans = vtk.vtkTransform()
     trans.RotateX(angle)
@@ -148,6 +162,7 @@ def test_rotate_x():
 
 
 def test_rotate_y():
+    grid = GRID.copy()
     angle = 30
     trans = vtk.vtkTransform()
     trans.RotateY(angle)
@@ -165,6 +180,7 @@ def test_rotate_y():
 
 
 def test_rotate_z():
+    grid = GRID.copy()
     angle = 30
     trans = vtk.vtkTransform()
     trans.RotateZ(angle)
@@ -182,6 +198,7 @@ def test_rotate_z():
 
 
 def test_make_points_double():
+    grid = GRID.copy()
     grid_copy = grid.copy()
[…]
