### 2019-05-25 a6fd553
deprecate n_scalars -> n_arrays
