+++ tests/test_filters.py
@@ -272,13 +272,13 @@ def test_warp_by_scalar():
 def test_cell_data_to_point_data():
     data = examples.load_uniform()
     foo = data.cell_data_to_point_data()
-    assert foo.n_scalars == 2
+    assert foo.n_arrays == 2
     assert len(foo.cell_arrays.keys()) == 0
 
 def test_point_data_to_cell_data():
     data = examples.load_uniform()
     foo = data.point_data_to_cell_data()
-    assert foo.n_scalars == 2
+    assert foo.n_arrays == 2
     assert len(foo.point_arrays.keys()) == 0
 
 
