### 2019-05-26 e009801
Add cpos arg to .show()
