### 2019-05-27 82dcb16
add_volume can now take pv.UniformGrid. Also now supports different blending modes
