### 2019-05-27 100ad9c
Fix hook for normals computation
