### 2019-05-27 6927b51
added input data type error
