### 2019-05-27 e186e35
Update AppVeyor badge to show master branch only
