### 2019-05-27 a8e9c2c
Add ability to add single floors to scene
