### 2019-05-28 ebb6085
extract surface for normal computation when using smooth_shading
