### 2019-05-28 712d551
now splitting vertices when computing normals for smooth shading
