### 2019-05-30 e83cf30
improve error message for wrapping non VTK data types
