### 2019-05-31 ebfa547
Simplify internals for removing arrays
