### 2019-05-31 22fde83
Clearing scalar arrays (#232)
* added convenience function for clearing scalar arrays

* clear_scalars -> clear_arrays

* Make clear methods for all scalar array types and test

* deprecate n_scalars -> n_arrays

* Simplify internals for removing arrays
