### 2019-06-05 #1450 @leouieda
@banesullivan sorry, I missed this comment. Could you open an issue at https://github.com/openjournals/joss/issues for this? There are other issues related to the front page and it's currently being redesigned. At least the metadata and paper page itself are fine.
