### 2020-07-08 #1450 @banesullivan
I noticed recently that the "Software Repository" link on https://joss.theoj.org/papers/10.21105/joss.01450 is incorrect. Can we update this? cc'ing @leouieda and @arfon

This seems to be an artifact of changing the repository's URL during the pre-review process. Ref https://github.com/openjournals/joss-reviews/issues/1382#issuecomment-491614935
